% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{addInterval}
\alias{addInterval}
\title{Adds an interval to the output schema}
\usage{
addInterval(simulation, startTime, endTime, resolution = 4/60,
  intervalName = NULL)
}
\arguments{
\item{startTime}{Start time of the interval in min}

\item{endTime}{End time of the interval in min}

\item{resolution}{Optional resolution in points/min (Default is 4 pts/hr)}

\item{intervalName}{Optional Name of interval. If not specified, a unique name wil be assigned.}

\item{outputSchema}{Schema where the newly created interval will be added}
}
\description{
Adds an interval to the output schema
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

# clears the previous output schema
clearIntervals(sim)

# Adds a new interval starting at 1h and ending at 10h with a resolution of 10 points per hour
addInterval(sim, 1 * 60, 10 * 60, 1 / 6)

# Adds another interval starting at 10h and ending at 17h with a default resolution and a specified name
addInterval(sim, 10 * 60, 17 * 60, intervalName = "Second Interval")
}
