% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{addLocalMoleculeParameters}
\alias{addLocalMoleculeParameters}
\title{Extend a Parameter Values Building Block (BB) with local molecule parameters
for molecules from a molecules BB in all physical containers of a spatial structure BB.}
\usage{
addLocalMoleculeParameters(
  parameterValuesBuildingBlock,
  spatialStructureBB,
  moleculesBB,
  moleculeNames = NULL
)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}

\item{spatialStructureBB}{A \code{BuildingBlock} object of type \verb{Spatial Structure}.
Entries will be created for local parameters of the selected molecules in all physical containers of this
spatial structure.}

\item{moleculesBB}{A \code{BuildingBlock} object of type \code{Molecules}. The entries will be
created for all molecules from this building block, or for a subset of molecules
defined in the \code{moleculeNames} argument.}

\item{moleculeNames}{Optional list of molecule names. If provided, only the molecules
with these names will be added to the \code{parameterValuesBuildingBlock}.}
}
\value{
Path of entries added to the building block.
}
\description{
Existing entries will not be overwritten.
}
