% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{addLocalMoleculeParametersToParameterValuesBB}
\alias{addLocalMoleculeParametersToParameterValuesBB}
\title{Extend a Parameter Values Building Block (BB) with local molecule parameters
for molecules from a molecules module in all physical containers of a spatial structure module.}
\usage{
addLocalMoleculeParametersToParameterValuesBB(
  parameterValuesBuildingBlock,
  spatialStructureModule,
  moleculesModule,
  moleculeNames = NULL
)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}

\item{spatialStructureModule}{A \code{MoBiModule} containing a \verb{Spatial Structure} building block.
Entries will be created for local parameters of the selected molecules in all physical containers of this
spatial structure.}

\item{moleculesModule}{A \code{MoBiModule} containing a \code{Molecules} building block. The entries will be
created for all molecules from this building block, or for a subset of molecules
defined in the \code{moleculeNames} argument.}

\item{moleculeNames}{Optional list of molecule names. If provided, only the molecules
with these names will be added to the \code{parameterValuesBuildingBlock}. If a specified molecule is not present in the provided molecules BB, it will be ignored.}
}
\value{
Path of entries added to the building block.
}
\description{
Existing entries will not be overwritten.
}
\examples{
\dontrun{
project <- loadMoBiProject("path/to/project.mbp3")
module <- project$getModules("TestModule")[[1]]
pvBB <- module$getParameterValuesBBs()[[1]]

newPaths <- addLocalMoleculeParametersToParameterValuesBB(
  parameterValuesBuildingBlock = pvBB,
  spatialStructureModule = module,
  moleculesModule = module,
  moleculeNames = c("CYP3A4")
)
}
}
