% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{addOutputInterval}
\alias{addOutputInterval}
\title{Adds an interval to the output schema of the simulation}
\usage{
addOutputInterval(simulation, startTime, endTime, resolution,
  intervalName = NULL)
}
\arguments{
\item{simulation}{Simulation for which a new interval should be created}

\item{startTime}{Start time of the interval in min}

\item{endTime}{End time of the interval in min}

\item{resolution}{resolution in points/min}

\item{intervalName}{Optional Name of interval. If not specified, a unique name wil be assigned.}
}
\value{
Returns the interval created.
}
\description{
Adds an interval to the output schema of the simulation
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath, addToCache = FALSE, loadFromCache = FALSE)

# clears the previous output schema
clearIntervals(sim)

# Adds a new interval starting at 1h and ending at 10h with a resolution of 10 points per hour
addOutputInterval(sim, 1 * 60, 10 * 60, 1 / 6)

# Adds another interval starting at 10h and ending at 17h with a resolution of 4 points per hour
# and a specified name
addOutputInterval(sim, 10 * 60, 17 * 60, 4/60,  intervalName = "Second Interval")
}
