% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-parameter.R
\name{addUserDefinedPKParameter}
\alias{addUserDefinedPKParameter}
\title{Adds a User-Defined PK-Parameter to the managed list of PK-Parameters}
\usage{
addUserDefinedPKParameter(userDefinedPKParameter)
}
\arguments{
\item{userDefinedPKParameter}{User defined PK-Parameter to add}
}
\description{
Adds a User-Defined PK-Parameter to the managed list of PK-Parameters
}
\examples{

userDefinedPKParameter <- UserDefinedPKParameter$new(name = "MyAUC", standardPKParameter = StandardPKParameter$AucTend)
addUserDefinedPKParameter(userDefinedPKParameter)
}
