% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{allAvailableDimensions}
\alias{allAvailableDimensions}
\title{Returns the name of all available dimensions defined in the OSPSuite platform}
\usage{
allAvailableDimensions()
}
\description{
Returns the name of all available dimensions defined in the OSPSuite platform
}
\examples{
dims <- allAvailableDimensions()

}
