% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\docType{data}
\name{AllPathsInMethod}
\alias{AllPathsInMethod}
\title{Names of the \code{.NET} container tasks of the type \code{"AllXXXPathsIn"}}
\format{
An object of class \code{list} of length 4.
}
\usage{
AllPathsInMethod
}
\description{
Names of the \code{.NET} container tasks of the type \code{"AllXXXPathsIn"}
}
\keyword{internal}
