% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-config.R
\docType{class}
\name{ApiConfig}
\alias{ApiConfig}
\title{ApiConfig}
\description{
Global configuration for the OSPSuite .NET API
}
\keyword{internal}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{DotNetWrapper}} -> \code{ApiConfig}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{pkParametersFilePath}}{Full path of the pkParameter file}

    \item{\code{dimensionFilePath}}{Full path of the dimension file}
  }
  \if{html}{\out{</div>}}
}
