% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-block.R
\docType{data}
\name{BuildingBlockTypes}
\alias{BuildingBlockTypes}
\title{Types of building blocks.}
\format{
An object of class \code{list} of length 10.
}
\usage{
BuildingBlockTypes
}
\description{
MUST WE DEFINE IT HERE, OR CAN WE GET IT FROM .NET SOMEHOW?
ARE INDIVIDUAL AND EXPRESSION PROFILE ALSO A BB TYPE?
}
\keyword{datasets}
