% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{class}
\name{Cache}
\alias{Cache}
\title{Cache}
\description{
A class that provides structure for caching objects.
}
\examples{
# Create a new instance of the class
myCache <- ospsuite:::Cache$new()

# Adding new key-value pairs
myCache$set("data1", iris)
myCache$set("data2", BOD)

# Print to see the current pairs
myCache

# Get value using a key
myCache$get("data2")

}
\keyword{internal}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{type}}{Type of the object that are cached. All objects must be of the same type.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{keys}}{List all key in the cache}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Cache-initialize}{\code{Cache$new()}}
    \item \href{#method-Cache-reset}{\code{Cache$reset()}}
    \item \href{#method-Cache-set}{\code{Cache$set()}}
    \item \href{#method-Cache-hasKey}{\code{Cache$hasKey()}}
    \item \href{#method-Cache-get}{\code{Cache$get()}}
    \item \href{#method-Cache-dropKey}{\code{Cache$dropKey()}}
    \item \href{#method-Cache-print}{\code{Cache$print()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-initialize}{}}}
\subsection{\code{Cache$new()}}{
  Initialize a new instance of the class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$new(type = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{Optional type constraints for object stored in the cache.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{Cache} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-reset}{}}}
\subsection{\code{Cache$reset()}}{
  Clear cache
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$reset()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-set"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-set}{}}}
\subsection{\code{Cache$set()}}{
  Store the given value in cache with the given key
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$set(key, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{Key to use to store the \code{value} in the case}
      \item{\code{value}}{Value to store in the cache}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-hasKey"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-hasKey}{}}}
\subsection{\code{Cache$hasKey()}}{
  Check if an entry with the given \code{key} exists in cache
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$hasKey(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{Key to check}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-get"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-get}{}}}
\subsection{\code{Cache$get()}}{
  Get cached object by key
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$get(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{Key to use to retrieve the value}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-dropKey"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-dropKey}{}}}
\subsection{\code{Cache$dropKey()}}{
  Remove the value associated with the key from cache.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$dropKey(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{Key used to find the value to remove. If the key is not found, nothing happens.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cache-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cache-print}{}}}
\subsection{\code{Cache$print()}}{
  Print the object to the console
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cache$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Rest arguments.}
    }
    \if{html}{\out{</div>}}
  }
}

}
