% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-analysis.R
\name{calculatePKAnalyses}
\alias{calculatePKAnalyses}
\title{Calculates the pkAnalyses for all output values available in \code{results}.}
\usage{
calculatePKAnalyses(results)
}
\arguments{
\item{results}{Results of simulation. Can be a single \code{SimulationResults} object
or a list of \code{SimulationResults} objects (typically the output from
\code{runSimulations}).}
}
\value{
If a single \code{SimulationResults} object is provided, returns a
\code{SimulationPKAnalyses} object. If a list of \code{SimulationResults} is provided,
returns a named list of \code{SimulationPKAnalyses} objects, with names corresponding to the names of the input list (usually the simulation IDs).
}
\description{
Calculates the pkAnalyses for all output values available in \code{results}.
}
\examples{

simPath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

addOutputs("Organism|VenousBlood|*|Aciclovir", sim)
results <- runSimulations(sim)[[1]]
pkAnalyses <- calculatePKAnalyses(results)
}
