% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{CellularPermeabilityMethods}
\alias{CellularPermeabilityMethods}
\title{Available methods for calculation of cellular permeabilities.}
\usage{
CellularPermeabilityMethods
}
\description{
Available methods for calculation of cellular permeabilities.
}
