% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{clearIntervals}
\alias{clearIntervals}
\title{Removes all intervals from the output schema defined in \code{simulation}}
\usage{
clearIntervals(simulation)
}
\arguments{
\item{simulation}{Instance of a simulation for which output intervals should be cleared}
}
\description{
Removes all intervals from the output schema defined in \code{simulation}
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

clearIntervals(sim)
}
