% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{clearOutputIntervals}
\alias{clearOutputIntervals}
\title{Removes all intervals from the output schema defined in \code{simulation}}
\usage{
clearOutputIntervals(simulation)
}
\arguments{
\item{simulation}{Instance of a simulation for which output intervals should be cleared}
}
\description{
Removes all intervals from the output schema defined in \code{simulation}
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

#Make sure we create a new simulation so that we do not impact other examples
sim <- loadSimulation(simPath, addToCache = FALSE, loadFromCache = FALSE)

clearOutputIntervals(sim)
}
