% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/enums.R
=======
% Please edit documentation in R/utilities-entity.R
>>>>>>> c4fbe917 (Fixes #1834 refactor DataCombined plot examples)
\name{CompareBy}
\alias{CompareBy}
\title{How should comparison of entities be performed}
\usage{
CompareBy
}
\description{
How should comparison of entities be performed
}
