% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-type.R
\docType{data}
\name{ContainerType}
\alias{ContainerType}
\title{All possible container types as defined in OSPSuite}
\format{An object of class \code{list} of length 14.}
\usage{
ContainerType
}
\description{
All possible container types as defined in OSPSuite
}
\keyword{datasets}
