% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-snapshots.R
\name{convertSnapshot}
\alias{convertSnapshot}
\title{Convert between snapshot and project formats}
\usage{
convertSnapshot(..., format, output = ".", runSimulations = FALSE)
}
\arguments{
\item{...}{character strings, path to files or a directory containing files to convert}

\item{format, }{character string, target format either "snapshot" or "project".}

\item{output}{character string, path to the output directory where to write the converted files}

\item{runSimulations}{logical, whether to run simulations during conversion (default = FALSE).
Only when converting from snapshot to project.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{convertSnapshot()} is deprecated and will be removed in a future release.
Use \code{\link[=snapshotToProject]{snapshotToProject()}} to convert a snapshot to a project, and
\code{\link[=projectToSnapshot]{projectToSnapshot()}} to convert a project to a snapshot.
}
