% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\docType{class}
\name{Covariates}
\alias{Covariates}
\title{Covariates}
\usage{
Covariates
}
\description{
All covariates for one individual
}
\section{Methods}{

\describe{
  \item{valueFor(covariateName)}{Returns the value of the covariate named \code{covariateName} or an empty string if not defined}
  }
}

\keyword{datasets}
