% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createConfigurationForFile}
\alias{createConfigurationForFile}
\title{Create a DataImporterConfiguration object for a XLS file}
\usage{
createConfigurationForFile(filePath, sheet = NULL)
}
\arguments{
\item{filePath}{Path to XLS file}

\item{sheet}{optional - name of the sheet}
}
\value{
DataImporterConfiguration object for XLS file
}
\description{
Create a DataImporterConfiguration object for a XLS file
}
