% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-individual.R
\name{createDistributions}
\alias{createDistributions}
\title{Creates the parameter distributions based on the given individual \code{individualCharacteristics}}
\usage{
createDistributions(individualCharacteristics)
}
\arguments{
\item{individualCharacteristics}{Characteristics of the individual to create as an instance of \code{OriginData}}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates the parameter distributions based on the given individual \code{individualCharacteristics}
}
