% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-expression-profile.R
\name{createExpressionProfileBuildingBlock}
\alias{createExpressionProfileBuildingBlock}
\title{Create an Expression Profile Building Block}
\usage{
createExpressionProfileBuildingBlock(type, moleculeName, speciesName)
}
\arguments{
\item{type}{String. Type of the protein the profile will be created for. Must be one of the values defined in the \code{ExpressionProfileCategories} enum.}

\item{moleculeName}{Name of the protein molecule.}

\item{speciesName}{Name of the species. Must be one of the values defined in the \code{Species} enum.}
}
\value{
A \code{BuildingBlock} object representing the created expression profile.
}
\description{
Creates an expression profile building block in MoBi for a given molecule,
species, and category. The expression profile is created with default parameter values, which can be modified using \code{setExpressionProfileParameters()}.
Important: The expression profile is not queried from the expression database.
}
\details{
Disease states are currently not supported.
}
\examples{
expressionProfile <- createExpressionProfileBuildingBlock(
  type = ExpressionProfileCategories$`Metabolizing Enzyme`,
  moleculeName = "CYP3A4",
  speciesName = "Human"
)
}
