% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-importer-configuration.R
\name{createImporterConfigurationForFile}
\alias{createImporterConfigurationForFile}
\title{Create a \code{DataImporterConfiguration} for an XLS sheet}
\usage{
createImporterConfigurationForFile(filePath, sheet = NULL)
}
\arguments{
\item{filePath}{Path to XLS file}

\item{sheet}{optional - name of the sheet. If no sheet is specified, the first
sheet of the XLS file is used.}
}
\value{
\code{DataImporterConfiguration} object for XLS file to be used in
\code{loadDataSetsFromExcel()}.
}
\description{
Create a \code{DataImporterConfiguration} for an XLS sheet
}
\details{
The function tries to parse the structure of the excel sheet and
creates a default configuration for this sheet. It is advised to check the
configuration and adjust if necessary before using with \code{loadDataSetsFromExcel()}.
}
\examples{
xlsFilePath <- system.file("extdata", "CompiledDataSet.xlsx", package = "ospsuite")
importerConfiguration <- createImporterConfigurationForFile(xlsFilePath)
importerConfiguration$sheets <- "TestSheet_1"

dataSets <- loadDataSetsFromExcel(
  xlsFilePath = xlsFilePath,
  importerConfigurationOrPath = importerConfiguration,
  importAllSheets = FALSE
)
}
