% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-individual.R
\name{createIndividual}
\alias{createIndividual}
\title{Creates an individual using the PKSim Database.}
\usage{
createIndividual(individualCharacteristics, useDistribution = FALSE)
}
\arguments{
\item{individualCharacteristics}{Characteristics of the individual to create as an instance of \code{OriginData}}

\item{useDistribution}{Boolean. Indicates wheather the function should returns all parameters defining an individual on only those parameters
that are actually distributed parameters}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates an individual using the PKSim Database.
}
