% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{createInitialConditionsBuildingBlock}
\alias{createInitialConditionsBuildingBlock}
\title{Create an empty Initial Conditions Building Block}
\usage{
createInitialConditionsBuildingBlock(name = "Initial Conditions")
}
\arguments{
\item{name}{Name of the building block. Must be a non-empty string. Defaults
to \code{"Initial Conditions"}.}
}
\value{
A \code{BuildingBlock} object of type \verb{Initial Conditions} with no entries.
}
\description{
Creates an empty Initial Conditions building block that can subsequently be
populated via \code{\link[=setInitialConditionsInBB]{setInitialConditionsInBB()}} or \code{\link[=extendInitialConditionsBB]{extendInitialConditionsBB()}},
and added to a module via \code{\link[=createMoBiModule]{createMoBiModule()}}.
}
\examples{
icBB <- createInitialConditionsBuildingBlock()
icBB <- createInitialConditionsBuildingBlock(name = "My ICs")
}
