% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-individual.R
\name{createMoBiIndividualBuildingBlock}
\alias{createMoBiIndividualBuildingBlock}
\title{Create a MoBi Individual Building Block}
\usage{
createMoBiIndividualBuildingBlock(
  species,
  population = NULL,
  gender = NULL,
  weight = NULL,
  weightUnit = "kg",
  height = NULL,
  heightUnit = "cm",
  age = NULL,
  ageUnit = "year(s)",
  gestationalAge = 40,
  gestationalAgeUnit = "week(s)",
  seed = NULL
)
}
\arguments{
\item{species}{Species of the individual as defined in PK-Sim (see \code{Species} enum).}

\item{population}{Population to use to create the individual. Required when
species is Human (see \code{HumanPopulation} enum).}

\item{gender}{Gender of the individual (see \code{Gender} enum).}

\item{weight}{Weight of the individual.}

\item{weightUnit}{Unit of the weight value. Default is \code{"kg"}.}

\item{height}{Height of the individual (human species only).}

\item{heightUnit}{Unit of the height value. Default is \code{"cm"}.}

\item{age}{Age of the individual (human species only).}

\item{ageUnit}{Unit of the age value. Default is \code{"year(s)"}.}

\item{gestationalAge}{Gestational age of the individual (for preterm human population).
Default is 40.}

\item{gestationalAgeUnit}{Unit of the gestational age value. Default is \code{"week(s)"}.}

\item{seed}{Optional seed for the individual creation algorithm.}
}
\value{
A \code{BuildingBlock} object representing the created individual.
}
\description{
Creates an individual building block in MoBi for a given species and
optional demographic characteristics by calling the MoBi \code{IndividualTask}.
}
\examples{
\dontrun{
individual <- createMoBiIndividualBuildingBlock(
  species = Species$Human,
  population = HumanPopulation$European_ICRP_2002,
  gender = Gender$Male,
  weight = 73,
  age = 30
)
}
}
