% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{createParameterValuesBuildingBlock}
\alias{createParameterValuesBuildingBlock}
\title{Create an empty Parameter Values Building Block}
\usage{
createParameterValuesBuildingBlock(name = "Parameter Values")
}
\arguments{
\item{name}{Name of the building block. Must be a non-empty string. Defaults
to \code{"Parameter Values"}.}
}
\value{
A \code{BuildingBlock} object of type \verb{Parameter Values} with no entries.
}
\description{
Creates an empty Parameter Values building block that can subsequently be
populated via \code{\link[=setParameterValuesInBB]{setParameterValuesInBB()}} or
\code{\link[=addLocalMoleculeParametersToParameterValuesBB]{addLocalMoleculeParametersToParameterValuesBB()}}, and added to a module via
\code{\link[=createMoBiModule]{createMoBiModule()}}.
}
\examples{
pvBB <- createParameterValuesBuildingBlock()
pvBB <- createParameterValuesBuildingBlock(name = "My PVs")
}
