% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plot-grid-configuration.R
\name{createPlotGridConfiguration}
\alias{createPlotGridConfiguration}
\title{Create \code{ospPlotGridConfiguration} class object}
\usage{
createPlotGridConfiguration(
  plotList,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tagLevels = NULL,
  tagPrefix = NULL,
  tagSuffix = NULL,
  tagSeparator = NULL,
  theme = NULL,
  nColumns = NULL,
  nRows = NULL,
  byRow = NULL,
  widths = NULL,
  heights = NULL,
  guides = NULL,
  design = NULL
)
}
\arguments{
\item{plotList}{A list containing \code{ggplot} objects.}

\item{title, subtitle, caption}{Text strings to use for the various plot
annotations, where plot refers to the grid of plots as a whole.}

\item{tagLevels}{A character vector defining the enumeration format to use
at each level. Possible values are \code{'a'} for lowercase letters, \code{'A'} for
uppercase letters, \code{'1'} for numbers, \code{'i'} for lowercase Roman numerals, and
\code{'I'} for uppercase Roman numerals. It can also be a list containing
character vectors defining arbitrary tag sequences. If any element in the
list is a scalar and one of \code{'a'}, \code{'A'}, \code{'1'}, \verb{'i}, or \code{'I'}, this level
will be expanded to the expected sequence.}

\item{tagPrefix, tagSuffix}{Strings that should appear before or after the
tag.}

\item{tagSeparator}{A separator between different tag levels}

\item{theme}{A ggplot theme specification to use for the plot. Only elements
related to the titles as well as plot margin and background will be used.}

\item{nColumns, nRows}{The dimensions of the grid to create - if both are \code{NULL} it
will use the same logic as \link[ggplot2:facet_wrap]{facet_wrap()} to set the
dimensions}

\item{byRow}{Analogous to \code{byrow} in \link[base:matrix]{matrix()}. If \code{FALSE} the
plots will be filled in in column-major order.}

\item{widths, heights}{The relative widths and heights of each column and row
in the grid. Will get repeated to match the dimensions of the grid.}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not. If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork.}

\item{design}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to \code{\link[=area]{area()}} together.
See the examples for further information on use.}
}
\description{
Create \code{ospPlotGridConfiguration} class object
}
