% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-population.R
\name{createPopulationCharacteristics}
\alias{createPopulationCharacteristics}
\title{Creates the population characteristics used to create a population}
\usage{
createPopulationCharacteristics(
  species,
  population = NULL,
  numberOfIndividuals,
  proportionOfFemales = 50,
  weightMin = NULL,
  weightMax = NULL,
  weightUnit = "kg",
  heightMin = NULL,
  heightMax = NULL,
  heightUnit = "cm",
  ageMin = NULL,
  ageMax = NULL,
  ageUnit = "year(s)",
  BMIMin = NULL,
  BMIMax = NULL,
  BMIUnit = "kg/m²",
  gestationalAgeMin = NULL,
  gestationalAgeMax = NULL,
  gestationalAgeUnit = "week(s)",
  moleculeOntogenies = NULL
)
}
\arguments{
\item{species}{Species of the individual as defined in PK-Sim (see Species enum)}

\item{population}{Population to use to create the individual. This is required only when the species is Human. (See HumanPopulation enum)}

\item{numberOfIndividuals}{Number of individuals in the population}

\item{proportionOfFemales}{Proportions of females. Default is 50 (50\%)}

\item{weightMin}{min weight for the population (optional)}

\item{weightMax}{max weight for the population (optional)}

\item{weightUnit}{Unit in which the weight value is defined. Default is kg}

\item{heightMin}{min height for the population (optional, for human species only)}

\item{heightMax}{max height for the population (optional, for human species only)}

\item{heightUnit}{Unit in which the height value is defined. Default is cm}

\item{ageMin}{min age for the population (optional, for human species only)}

\item{ageMax}{max age for the population (optional, for human species only)}

\item{ageUnit}{Unit in which the age value is defined. Default is year(s)}

\item{BMIMin}{min BMI for the population (optional, for human species only)}

\item{BMIMax}{max BMI for the population (optional, for human species only)}

\item{BMIUnit}{Unit in which the BMI value is defined. Default is year(s)}

\item{gestationalAgeMin}{min gestational age for the population (optional, for human species only)}

\item{gestationalAgeMax}{max gestational age for the population (optional, for human species only)}

\item{gestationalAgeUnit}{Unit in which the gestational age value is defined. Default is kg/m2}

\item{moleculeOntogenies}{Optional list of \code{MoleculeOntogeny} that will be used to retrieve ontogeny information for molecules.}
}
\value{
An instance of \code{PopulationCharacteristics} to be used in conjunction with \code{createPopulation}
}
\description{
Creates the population characteristics used to create a population
}
