% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{createSimulationBatch}
\alias{createSimulationBatch}
\title{Creates and returns an instance of a \code{SimulationBatch} that can be used to efficiently vary parameters and initial values in a simulation}
\usage{
createSimulationBatch(
  simulation,
  parametersOrPaths = NULL,
  moleculesOrPaths = NULL
)
}
\arguments{
\item{simulation}{Instance of a \code{Simulation} to simulate in a batch mode}

\item{parametersOrPaths}{Parameter instances (element or vector) typically retrieved using
\code{getAllParametersMatching} or parameter path (element or vector of strings) that will be varied in the simulation. (optional)
When providing the paths, only absolute full paths are supported (i.e., no matching with '*' possible).
If parametersOrPaths is \code{NULL}, you will not be able to set parameter values during batch run.}

\item{moleculesOrPaths}{Molecule instances (element or vector) typically retrieved using
\code{getAllMoleculesMatching} or molecule path (element or vector of strings) that will be varied in the simulation. (optional)
When providing the paths, only absolute full paths are supported (i.e., no matching with '*' possible).
If moleculesOrPaths is \code{NULL}, you will not be able to set molecule initial values during batch run.}
}
\value{
SimulationBatch that can be used to vary parameter values or molecule initial values and run simulation in an optimized manner
}
\description{
Creates and returns an instance of a \code{SimulationBatch} that can be used to efficiently vary parameters and initial values in a simulation
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Create a simulation batch that will allow batch run for one parameter value
simulationBatch <- createSimulationBatch(sim, "Organism|Liver|Volume")

# Create a simulation batch that will allow batch run for multiple parameter
# values and initial values
simulationBatch <- createSimulationBatch(
  sim,
  c("Organism|Liver|Volume", "R1|k1"),
  c("Organism|Liver|A")
)
}
