% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/enums.R
=======
% Please edit documentation in R/utilities-plotting.R
>>>>>>> c4fbe917 (Fixes #1834 refactor DataCombined plot examples)
\name{DataAggregationMethods}
\alias{DataAggregationMethods}
\title{Names of aggregation available for plotPopulationTimeProfile()}
\usage{
DataAggregationMethods
}
\description{
Names of aggregation available for plotPopulationTimeProfile()
}
