% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined-plotting.R
\name{dataCombinedToObsVsPredDataFrame}
\alias{dataCombinedToObsVsPredDataFrame}
\title{Extract "observed versus predicted" dataframe from \code{DataCombined} object}
\usage{
dataCombinedToObsVsPredDataFrame(dataCombined)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class.}
}
\description{
Extract "observed versus predicted" dataframe from \code{DataCombined} object
}
