% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/enums.R
=======
% Please edit documentation in R/data-set.R
>>>>>>> c4fbe917 (Fixes #1834 refactor DataCombined plot examples)
\name{DataErrorType}
\alias{DataErrorType}
\title{Supported types of the error}
\usage{
DataErrorType
}
\description{
Supported types of the error
}
