% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-internal-plot-configuration.R
\name{DefaultInternalPlotConfiguration}
\alias{DefaultInternalPlotConfiguration}
\title{Convenience class for plot configuration for OSP plots}
\description{
An intermediary R6 class that provides a single portal to prepare the
\code{PlotConfiguration} object needed by \code{{tlf}} plotting functions. The specific
type of \code{PlotConfiguration} objects can access all necessary objects formed
from user inputs using this internal object.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labels}}{\code{tlf::LabelConfiguration} object defining properties of labels.}

\item{\code{legend}}{\code{tlf::LegendConfiguration} object defining legend properties.}

\item{\code{xAxis}}{\code{tlf::XAxisConfiguration} object defining \code{x}-axis properties.}

\item{\code{yAxis}}{\code{tlf::YAxisConfiguration} object defining \code{y}-axis properties.}

\item{\code{background}}{\code{tlf::BackgroundConfiguration} object defining the
configuration of background.}

\item{\code{lines}}{\code{tlf::ThemeAestheticSelections} object or list defining how
lines are plotted.}

\item{\code{points}}{\code{tlf::ThemeAestheticSelections} object or list defining how
points are plotted.}

\item{\code{ribbons}}{\code{tlf::ThemeAestheticSelections} object or list defining
how ribbons are plotted.}

\item{\code{errorbars}}{\code{tlf::ThemeAestheticSelections} object or list defining
how errorbars are plotted.}

\item{\code{export}}{\code{tlf::SaveConfiguration} object defining saving properties.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DefaultInternalPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultInternalPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
