% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{deleteInitialConditionsFromBB}
\alias{deleteInitialConditionsFromBB}
\title{Delete entries from an Initial Conditions Building Block}
\usage{
deleteInitialConditionsFromBB(initialConditionsBuildingBlock, quantityPaths)
}
\arguments{
\item{initialConditionsBuildingBlock}{A \code{BuildingBlock} object of type \verb{Initial Conditions}.}

\item{quantityPaths}{A list of full paths of the quantities (usually molecules) that
will be deleted. Entries not present in the provided BB are ignored. Should contain
all path elements and the molecule name, separated by \code{|}.}
}
\value{
Invisibly returns the updated \code{initialConditionsBuildingBlock} object.
}
\description{
Delete entries from an Initial Conditions Building Block
}
\examples{
sim <- loadSimulation(system.file(
  "extdata",
  "simple.pkml",
  package = "ospsuite"
))
module <- sim$configuration$modules[[1]]
icBB <- module$getInitialConditionsBBs()[[1]]
deleteInitialConditionsFromBB(
  icBB,
  quantityPaths = c("Organism|Liver|A", "Organism|Liver|B")
)
}
