% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{deleteParameterValues}
\alias{deleteParameterValues}
\title{Delete entries from a Parameter Values Building Block}
\usage{
deleteParameterValues(parameterValuesBuildingBlock, quantityPaths)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}

\item{quantityPaths}{A list of full paths of the quantities (usually parameters)
that will be deleted. Should contain all path elements and the parameter name, separated by \code{|}.
Entries not present in the provided BB are ignored.}
}
\description{
Delete entries from a Parameter Values Building Block
}
