% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{deleteParameterValuesFromBB}
\alias{deleteParameterValuesFromBB}
\title{Delete entries from a Parameter Values Building Block}
\usage{
deleteParameterValuesFromBB(parameterValuesBuildingBlock, quantityPaths)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}

\item{quantityPaths}{A list of full paths of the quantities (usually parameters)
that will be deleted. Should contain all path elements and the parameter name, separated by \code{|}.
Entries not present in the provided BB are ignored.}
}
\value{
Invisibly returns the updated \code{parameterValuesBuildingBlock} object.
}
\description{
Delete entries from a Parameter Values Building Block
}
\examples{
sim <- loadSimulation(system.file(
  "extdata",
  "simple.pkml",
  package = "ospsuite"
))
module <- sim$configuration$modules[[1]]
pvBB <- module$getParameterValuesBBs()[[1]]
deleteParameterValuesFromBB(pvBB, quantityPaths = "Organism|Volume")
}
