% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\docType{data}
\name{Dimensions}
\alias{Dimensions}
\title{Supported dimensions.}
\format{
An object of class \code{list} of length 90.
}
\usage{
Dimensions
}
\description{
Supported dimensions.
}
\keyword{datasets}
