% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dimensions.R
\name{dimensionsConversionFactor}
\alias{dimensionsConversionFactor}
\title{Calculate factor for converting values of one dimension into another}
\usage{
dimensionsConversionFactor(dimension1, dimension2, mw = NULL)
}
\arguments{
\item{dimension1}{String name of the first dimension}

\item{dimension2}{String name fo the second dimension}

\item{mw}{Double value of the molecular weight of the molecule. Used
for conversion between molar and mass concentrations or amounts. Can be \code{NULL}
for other dimensions (default)}
}
\value{
A numerical factor the values of the first dimensions must be multiplied by to
be converted to second dimension. The conversion assumes values being in base units
of both dimensions.
If the values cannot be transformed between the dimensions, an error is thrown.
}
\description{
Calculate factor for converting values of one dimension into another
}
