% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\docType{data}
\name{DimensionsUnits}
\alias{DimensionsUnits}
\title{Supported units of dimensions
enum of all units for each dimension}
\format{
An object of class \code{list} of length 90.
}
\usage{
DimensionsUnits
}
\description{
Supported units of dimensions
enum of all units for each dimension
}
\keyword{datasets}
