% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.bbTypeToTaskName}
\alias{.bbTypeToTaskName}
\title{Mapping from a \code{BuildingBlockTypes} value to the name of the MoBi task that
implements \code{LoadFromPKML} for that type.}
\usage{
.bbTypeToTaskName(type)
}
\description{
Mapping from a \code{BuildingBlockTypes} value to the name of the MoBi task that
implements \code{LoadFromPKML} for that type.
}
\keyword{internal}
