% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.bbWithInitialConditionsToDataFrame}
\alias{.bbWithInitialConditionsToDataFrame}
\title{Convert a building block containing initial conditions to a data frame.}
\usage{
.bbWithInitialConditionsToDataFrame(buildingBlock)
}
\arguments{
\item{buildingBlock}{A \code{BuildingBlock} object of type \verb{Initial Conditions}
or \verb{Expression Profile}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \verb{Container Path}: Full path to the container where the molecule is located.
\item \verb{Molecule Name}: Name of the molecule.
\item \verb{Is Present}: Boolean indicating if the molecule is present.
\item \code{Value}: Initial value of the molecule. For values that are defined by a formula, the return value can be \code{NaN}.
\item \code{Unit}: Unit of the initial value.
\item \verb{Scale Divisor}: Scale divisor for the initial value.
\item \verb{Neg. Values Allowed}: Boolean indicating if negative values are allowed.
}
}
\description{
Works with any building block that contains initial conditions, including
\verb{Initial Conditions} and \verb{Expression Profile} building blocks.
}
\keyword{internal}
