% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.bbWithParameterValuesToDataFrame}
\alias{.bbWithParameterValuesToDataFrame}
\title{Convert a building block containing parameter values to a data frame.}
\usage{
.bbWithParameterValuesToDataFrame(buildingBlock)
}
\arguments{
\item{buildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}, \code{Individual}, or \verb{Expression Profile}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \verb{Container Path}: Full path to the container where the parameter is located.
\item \verb{Parameter Name}: Name of the parameter.
\item \code{Value}: Value of the parameter. For values that are defined by a formula, the return value can be \code{NaN}.
\item \code{Unit}: Unit of the parameter value.
\item \verb{Value Origin}: Origin of the parameter value.
}
}
\description{
Works with any building block that contains parameter values, including
\verb{Parameter Values}, \code{Individual}, and \verb{Expression Profile} building blocks.
}
\keyword{internal}
