% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{.cleanVectorArgs}
\alias{.cleanVectorArgs}
\title{Validate arguments provided as vectors}
\usage{
.cleanVectorArgs(arg = NULL, expectedLength = NULL, type)
}
\arguments{
\item{expectedLength}{An integer to denote the expected length of the vector.}

\item{type}{A single string or a vector of string representation or class of
the type that should be checked for.}

\item{x}{A vector of arguments.}
}
\value{
An atomic vector of desired data type.
}
\description{
Validate arguments provided as vectors
}
\details{
Cleaning an argument provided as (atomic or generic) vector involves:
\itemize{
\item Checking that it is of expected length.
\item Checking for \code{NULL} or other special constants (\code{NaN}, \code{Inf}, \code{NA} of the
wrong type) and standardizing them to \code{NA} of desired data type.
\item Checking that each element in the vector is of expected data type.
\item Making sure that an atomic vector is always returned, irrespective of if
the input was a list or an atomic vector.
}
}
\examples{
ospsuite:::.cleanVectorArgs(list(1, 2, NA, NULL), 4L, "numeric")
ospsuite:::.cleanVectorArgs(c(1, 2, NA, NA_complex), 4L, "numeric")
}
\seealso{
Other data-combined: 
\code{\link{.extractResidualsToTibble}()},
\code{\link{DataCombined}}
}
\concept{data-combined}
\keyword{internal}
