% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.convertGeneralToSpecificPlotConfiguration}
\alias{.convertGeneralToSpecificPlotConfiguration}
\title{Create plot-specific \code{tlf::PlotConfiguration} object}
\usage{
.convertGeneralToSpecificPlotConfiguration(
  specificPlotConfiguration,
  generalPlotConfiguration
)
}
\arguments{
\item{specificPlotConfiguration}{A specific subclass of
\code{tlf::PlotConfiguration} needed for the given plot.}

\item{generalPlotConfiguration}{A \code{DefaultPlotConfiguration} object.}
}
\description{
Create plot-specific \code{tlf::PlotConfiguration} object
}
\details{
The default plot configuration and the labels needs to vary from plot-to-plot
because each plot has its specific (default) aesthetic needs that need to be
met.

For example, although the axes labels for profile plots will be (e.g.) "Time
vs Fraction", they will be (e.g.) "Observed vs simulated values" for scatter
plots. Additionally, mapping group to line colors might be desirable for a
profile plot, it is not so for scatter plots.

This function generates object of specific subclass of
\code{tlf::PlotConfiguration} needed for the given plot but with suitable defaults
taken from the \code{DefaultPlotConfiguration} object.
}
\examples{

ospsuite:::.convertGeneralToSpecificPlotConfiguration(
  tlf::TimeProfilePlotConfiguration$new(),
  ospsuite::DefaultPlotConfiguration$new()
)

}
\seealso{
Other utilities-plotting:
\code{\link[=.addMissingGroupings]{.addMissingGroupings()}},
\code{\link[=.createAxesLabels]{.createAxesLabels()}},
\code{\link[=.extractAggregatedSimulatedData]{.extractAggregatedSimulatedData()}}
}
\concept{utilities-plotting}
\keyword{internal}
