% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.encodeUnit}
\alias{.encodeUnit}
\title{Encoding mu symbol}
\usage{
.encodeUnit(unit)
}
\arguments{
\item{unit}{Unit to encode.}
}
\description{
Encoding mu symbol
}
\details{
This is required to ensure that we have no issue using the mu symbol in different OS
See https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/476 for details.
}
\examples{
ospsuite:::.encodeUnit("µl")

}
\keyword{internal}
