% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.extractMatchingIndices}
\alias{.extractMatchingIndices}
\title{Custom function to extract matching indices}
\usage{
.extractMatchingIndices(x, y, tolerance = 0.001)
}
\arguments{
\item{x, y}{Numeric vectors to compare}

\item{tolerance}{Tolerance of comparison for observed and simulated time
points. Default is \code{NULL}, in which case the internal enumerated list
\code{.thresholdByTimeUnit} will be used to decide on what threshold to use
based on the unit of time measurement.}
}
\description{
None of the base equality/match operators (\code{\%in\%}, \code{==}, \code{all.equal}) allow
tolerance for comparing two numeric values. Therefore, \code{dplyr::near()} is
used.

But even \code{dplyr::near()} is not up to the task because it carries out vector
comparison element-wise, whereas what is needed is \code{match()}-like behavior,
where each element in the first vector is compared against all values in the
second vector for equality.

This custom function does exactly this.
}
\examples{

ospsuite:::.extractMatchingIndices(c(1, 2), c(1.001, 3, 4))
ospsuite:::.extractMatchingIndices(c(1, 2), c(1.001, 3, 4), tolerance = 0.00001)
ospsuite:::.extractMatchingIndices(c(1, 2), c(3, 4))

}
\seealso{
Other utilities-plotting: 
\code{\link{.addMissingGroupings}()},
\code{\link{.calculateResiduals}()},
\code{\link{.createAxesLabels}()},
\code{\link{.extractAggregatedSimulatedData}()},
\code{\link{.removeUnpairableDatasets}()}
}
\concept{utilities-plotting}
\keyword{internal}
