% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.extractPresentDatasetTypes}
\alias{.extractPresentDatasetTypes}
\title{Extracting dataset types present in \code{DataCombined}}
\usage{
.extractPresentDatasetTypes(dataCombined)
}
\description{
While creating plots, different code needs to be used depending on whether
only simulated, only observed, or both dataset types are present in the
\code{DataCombined}.
}
\keyword{internal}
