% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{.extractResidualsToTibble}
\alias{.extractResidualsToTibble}
\title{Created observed versus simulated paired data}
\usage{
.extractResidualsToTibble(data, scaling)
}
\arguments{
\item{data}{A data frame from \code{DataCombined$toDataFrame()}, which has been
further tidied using \code{.removeUnpairableDatasets()} and then
\code{.unitConverter()} functions.}
}
\description{
Created observed versus simulated paired data
}
\keyword{internal}
