% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-snapshots.R
\name{.gatherFiles}
\alias{.gatherFiles}
\title{Gather files and files from folders to one location}
\usage{
.gatherFiles(...)
}
\arguments{
\item{...}{character strings of file paths or folder paths}
}
\value{
A temporary directory with all files copied to it
}
\description{
Gather files and files from folders to one location
}
