% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{.getConcurrentSimulationRunnerResults}
\alias{.getConcurrentSimulationRunnerResults}
\title{Get SimulationResults from ConcurrentSimulationRunner}
\usage{
.getConcurrentSimulationRunnerResults(
  results,
  ids,
  simulationsIdMap,
  silentMode
)
}
\arguments{
\item{results}{.NET object created by \code{RunConcurrently()}}

\item{ids}{List of ids of simulations or simulation-batch values pairs in the order
as they were enqueued}

\item{simulationsIdMap}{A named list of simulation objects assigning the correct simulation
to the id of a result}

\item{silentMode}{If \code{TRUE}, no warnings are displayed if a simulation fails.}
}
\value{
A named list of \code{SimulationResults} objects with the names being the ids of simulations or
simulation-batch values pairs they were produced by
}
\description{
Get SimulationResults from ConcurrentSimulationRunner
}
\details{
Create a list of \code{SimulationResults}-objects from the results of a
\code{ConcurrentSimulationRunner}
}
