% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{.getContainerTask}
\alias{.getContainerTask}
\title{.getContainerTask}
\usage{
.getContainerTask()
}
\value{
an instance of the \code{.NET} Task \code{ContainerTask}
}
\description{
.getContainerTask
}
\keyword{internal}
