% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{.getCoreTask}
\alias{.getCoreTask}
\title{.getCoreTask}
\usage{
.getCoreTask(taskName)
}
\arguments{
\item{taskName}{The name of the task to retrieve (\strong{without} \code{Get} prefix).}
}
\value{
returns a new instance of of the specified \code{.NET} task.
}
\description{
Get an instance of the specified \code{.NET} Task in OSPSuite.R.Api
}
\keyword{internal}
