% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{.getCoreTaskFromCache}
\alias{.getCoreTaskFromCache}
\title{.getCoreTaskFromCache}
\usage{
.getCoreTaskFromCache(taskName)
}
\arguments{
\item{taskName}{The name of the task to retrieve (\strong{without} \code{Get} prefix).}
}
\value{
returns an instance of of the specified \code{.NET} task.
}
\description{
Get an instance of the specified \code{.NET} Task from OSPSuite.R.Api that is retrieved
from cache if already initiated. Otherwise a new task will be initiated and
cached in the \code{tasksEnv}.
}
\keyword{internal}
