% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{.getDimensionTask}
\alias{.getDimensionTask}
\title{.getDimensionTask}
\usage{
.getDimensionTask()
}
\value{
an instance of the \code{.NET} Task \code{DimensionTask}
}
\description{
.getDimensionTask
}
\keyword{internal}
