% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{.getOspDimensions}
\alias{.getOspDimensions}
\title{get OSP Dimensions from OSPSuite.Dimensions.xml data}
\usage{
.getOspDimensions(xmlData)
}
\arguments{
\item{xmlData}{XML data from \code{.parseDimensionsXML()}}
}
\value{
a list of supported dimensions
}
\description{
get OSP Dimensions from OSPSuite.Dimensions.xml data
}
