% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{.getOspUnits}
\alias{.getOspUnits}
\title{get OSP Units from OSPSuite.Dimensions.xml data}
\usage{
.getOspUnits(xmlData)
}
\arguments{
\item{xmlData}{XML data from \code{.parseDimensionsXML()}}
}
\value{
a list of supported units
}
\description{
get OSP Units from OSPSuite.Dimensions.xml data
}
