% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pksim.R
\name{.getPathToPKSimInstallDirFromRegistry}
\alias{.getPathToPKSimInstallDirFromRegistry}
\title{Tries to find the installation path for a specific version of PK-Sim from an Windows registry entry.}
\usage{
.getPathToPKSimInstallDirFromRegistry(pksim.version)
}
\arguments{
\item{pksim.version}{The version number of Pk-Sim as a string.}
}
\value{
The path to the PK-Sim installation for version pksim.version or NA if no path could be found.
The path is separated with slashes (unix-style) and in compliance with base-R without a trailing slash.
}
\description{
Tries to find the installation path for a specific version of PK-Sim from an Windows registry entry.
}
\examples{
\dontrun{
path <- .getPathToPKSimInstallDirFromRegistry("9")
}
}
\keyword{internal}
