% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.loadBBNetWithType}
\alias{.loadBBNetWithType}
\title{Load a building block of a known type via the matching MoBi task.}
\usage{
.loadBBNetWithType(expandedPath, type)
}
\description{
Load a building block of a known type via the matching MoBi task.
}
\keyword{internal}
