% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-repository.R
\name{.loadDataRepositoryFromPKML}
\alias{.loadDataRepositoryFromPKML}
\title{Loads a data repository (typically observed data) from a PKML file and returns the loaded observed data.
The pkml files are typically exported from PK-Sim or MoBi}
\usage{
.loadDataRepositoryFromPKML(filePath)
}
\arguments{
\item{filePath}{Full path of pkml file containing the observed data to load}
}
\description{
Loads a data repository (typically observed data) from a PKML file and returns the loaded observed data.
The pkml files are typically exported from PK-Sim or MoBi
}
