% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.moleculeTypeToProfileCategory}
\alias{.moleculeTypeToProfileCategory}
\title{Map a molecule type string returned by \code{MoleculesBuildingBlock$moleculeTypeFor()}
to the corresponding \link{ExpressionProfileCategories} entry.}
\usage{
.moleculeTypeToProfileCategory(moleculeType)
}
\description{
Map a molecule type string returned by \code{MoleculesBuildingBlock$moleculeTypeFor()}
to the corresponding \link{ExpressionProfileCategories} entry.
}
\keyword{internal}
