% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.normRange}
\alias{.normRange}
\title{Normal Range}
\usage{
.normRange(x, nsd = 1, na.rm = FALSE, ...)
}
\arguments{
\item{x}{numeric vector to compute normal range from}

\item{nsd}{optional argument defining the number of standard deviation to add
and substract to the mean}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds.}

\item{...}{further arguments passed to mean and sd functions.}
}
\value{
numeric vector of length 3 representing the min, mean and max of the
normal range.
}
\description{
Normal Range
}
\keyword{internal}
