% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter-value.R
\name{.parameterValueListFrom}
\alias{.parameterValueListFrom}
\title{Converts a list of .NET \code{ParameterValue} into a list with 2 entries: \code{paths}, \code{values}.
A 3rd optional entry \code{units} will be defined if the parameter \code{addUnits} is set to \code{TRUE}.
Note: Units are only available for .NET object of type \code{ParameterValueWithUnit}}
\usage{
.parameterValueListFrom(netParameterValues, addUnits = FALSE)
}
\arguments{
\item{netParameterValues}{List of.NET \code{ParameterValue} or \code{ParameterValueWithUnit}}

\item{addUnits}{If \code{TRUE}, a a third list will be returned containing the units in which the parameters are defined. Default is \code{FALSE}}
}
\value{
A list with 3 sublist: \code{paths}, \code{values}, and optionally \code{units} containing the corresponding values from each parameter value
}
\description{
Converts a list of .NET \code{ParameterValue} into a list with 2 entries: \code{paths}, \code{values}.
A 3rd optional entry \code{units} will be defined if the parameter \code{addUnits} is set to \code{TRUE}.
Note: Units are only available for .NET object of type \code{ParameterValueWithUnit}
}
