% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{.parseDimensionsXML}
\alias{.parseDimensionsXML}
\title{parse OSPSuite.Dimensions.xml containing dimensions and units}
\usage{
.parseDimensionsXML()
}
\value{
An XML document
}
\description{
parse OSPSuite.Dimensions.xml containing dimensions and units
}
