% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\docType{data}
\name{.presentDataTypes}
\alias{.presentDataTypes}
\title{An enum representing dataset types present in \code{DataCombined}}
\format{
An object of class \code{list} of length 3.
}
\usage{
.presentDataTypes
}
\description{
An enum representing dataset types present in \code{DataCombined}
}
\keyword{internal}
