% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecules-building-block.R
\name{.quantityTypeNetObject}
\alias{.quantityTypeNetObject}
\title{Convert a \code{MoleculeType} integer flag value to the underlying
\code{OSPSuite.Core.Domain.QuantityType} enum object that the engine accepts as a
parameter.}
\usage{
.quantityTypeNetObject(moleculeType)
}
\description{
Convert a \code{MoleculeType} integer flag value to the underlying
\code{OSPSuite.Core.Domain.QuantityType} enum object that the engine accepts as a
parameter.
}
\keyword{internal}
