% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{.removeEmptyDataFrame}
\alias{.removeEmptyDataFrame}
\title{Remove empty data frames from a list of data frames}
\usage{
.removeEmptyDataFrame(x)
}
\arguments{
\item{x}{A list of data frames.}
}
\description{
Remove empty data frames sometimes produced due to the non-existent
combination of source unit and molecular weight.
}
\examples{

# Create a list of data frames
(ls <- split(mtcars, list(mtcars$vs, mtcars$cyl)))

# Remove element data frames with 0 rows
ospsuite:::.removeEmptyDataFrame(ls)

}
\keyword{internal}
