% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{.removeUnpairableDatasets}
\alias{.removeUnpairableDatasets}
\title{Remove unpairable datasets for computing residuals}
\usage{
.removeUnpairableDatasets(data)
}
\arguments{
\item{data}{A data frame returned by \code{DataCombined$toDataFrame()}.}
}
\description{
Computing residuals by definition requires that data should be in pairs, i.e.
for every simulated dataset in a given group, there should also be a
corresponding observed dataset.

To this end, current function removes the following datasets:
\itemize{
\item Datasets which haven't been assigned to any group.
\item Datasets that are not part of a pair (i.e. a simulated dataset without
observed dataset partner, and vice versa).
}
}
\examples{

df <- dplyr::tribble(
  ~name, ~dataType, ~group,
  "Sim1", "Simulated", "GroupA",
  "Sim2", "Simulated", "GroupA",
  "Obs1", "Observed", "GroupB",
  "Obs2", "Observed", "GroupB",
  "Sim3", "Simulated", "GroupC",
  "Obs3", "Observed", "GroupC",
  "Sim4", "Simulated", "GroupD",
  "Obs4", "Observed", "GroupD",
  "Obs5", "Observed", "GroupD",
  "Sim5", "Simulated", "GroupE",
  "Sim6", "Simulated", "GroupE",
  "Obs7", "Observed", "GroupE",
  "Sim7", "Simulated", "GroupF",
  "Sim8", "Simulated", "GroupF",
  "Obs8", "Observed", "GroupF",
  "Obs9", "Observed", "GroupF",
  "Sim9", "Simulated", NA,
  "Obs10", "Observed", NA
)

# original
df

# transformed
ospsuite:::.removeUnpairableDatasets(df)

}
\keyword{internal}
