% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-results.R
\name{.saveResultsToCSV}
\alias{.saveResultsToCSV}
\title{Saves the simulation results to csv file}
\usage{
.saveResultsToCSV(results, filePath)
}
\arguments{
\item{results}{Results to export. Can be a single \code{SimulationResults} object
or a list containing exactly one \code{SimulationResults} object (typically the
output from \code{runSimulations} for a single simulation).}

\item{filePath}{Full path where the results will be saved.}
}
\description{
Saves the simulation results to csv file
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

# Add some outputs to the simulation
addOutputs("Organism|**|*", sim)

# Run the simulation
results <- runSimulations(sim)
exportResultsToCSV(results, tempfile())

}
