% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{.setEndSimulationTime}
\alias{.setEndSimulationTime}
\title{Set end time of the simulation}
\usage{
.setEndSimulationTime(simulation, endTime)
}
\arguments{
\item{simulation}{Simulation for which the end time should be set}

\item{endTime}{End time of the simulation in min}
}
\description{
Either extends or shortens the simulation time to the specified end time.
}
