% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.toObjectType}
\alias{.toObjectType}
\title{Convert \code{.NET} object to wrapper class in R}
\usage{
.toObjectType(netObject, class)
}
\arguments{
\item{netObject}{The \code{.NET} object instances (single or list) to wrap.}

\item{class}{The class definition that will be used to convert the parameter.}
}
\value{
The wrapped object (single or a list).
}
\description{
Transforms a single or a list of \code{.NET} object(s) to their
corresponding wrapper class in R. Note that if the object is a single object,
\code{NULL} will be returned if the \code{.NET} object is \code{null}. This allows semantic
equivalence between \code{.NET} and R.
}
\keyword{internal}
