% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{.validateHasUnit}
\alias{.validateHasUnit}
\title{Check if quantity can be represented in the unit}
\usage{
.validateHasUnit(quantity, unit)
}
\arguments{
\item{quantity}{\code{Quantity} object}

\item{unit}{Unit name to check for}
}
\value{
If validations are successful, \code{NULL} is returned. Otherwise, error is
signaled.
}
\description{
Check if quantity can be represented in the unit
}
