% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{.validateScalarDataCombined}
\alias{.validateScalarDataCombined}
\title{Validate that single instance of \code{DataCombined}}
\usage{
.validateScalarDataCombined(dataCombined)
}
\description{
Validate that single instance of \code{DataCombined}
}
\examples{
.validateScalarDataCombined(DataCombined$new()) # okay
# .validateScalarDataCombined(list(DataCombined$new(), DataCombined$new())) # error

}
\keyword{internal}
