% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{.wrapLoadedBB}
\alias{.wrapLoadedBB}
\title{Wrap a loaded .NET building block reference in the appropriate R6 class.
\code{Molecules} blocks are wrapped in \code{MoleculesBuildingBlock} (molecule-name
queries); \code{Individual} blocks in \code{IndividualBuildingBlock} (origin-data
fields); all other types use the generic \code{BuildingBlock} wrapper.}
\usage{
.wrapLoadedBB(netBB, type)
}
\description{
Wrap a loaded .NET building block reference in the appropriate R6 class.
\code{Molecules} blocks are wrapped in \code{MoleculesBuildingBlock} (molecule-name
queries); \code{Individual} blocks in \code{IndividualBuildingBlock} (origin-data
fields); all other types use the generic \code{BuildingBlock} wrapper.
}
\keyword{internal}
