% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumGetValue}
\alias{enumGetValue}
\title{Return the value that is stored under the given key. If the key is not present, an error is thrown.}
\usage{
enumGetValue(enum, key)
}
\arguments{
\item{enum}{enum that contains the key-value pair}

\item{key}{key under which the value is stored}
}
\value{
Value that is assigned to \code{key}
}
\description{
Return the value that is stored under the given key. If the key is not present, an error is thrown.
}
