% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumHasKey}
\alias{enumHasKey}
\title{Check if an enum has a certain key.}
\usage{
enumHasKey(key, enum)
}
\arguments{
\item{key}{Key to check for}

\item{enum}{Enum where to look for the key}
}
\value{
TRUE if a key-value pair for \code{key} exists, FALSE otherwise
}
\description{
Check if an enum has a certain key.
}
