% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumRemove}
\alias{enumRemove}
\title{Remove an entry from the enum.}
\usage{
enumRemove(keys, enum)
}
\arguments{
\item{keys}{Key(s) of entries to be removed from the enum}

\item{enum}{Enum from which the entries to be removed
WARNING: the original object is not modified!}
}
\value{
Enum without the removed entries
}
\description{
Remove an entry from the enum.
}
