% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{errorGetEntityMultipleOutputs}
\alias{errorGetEntityMultipleOutputs}
\title{Error message when single output is expected but multiple entities produced.}
\usage{
errorGetEntityMultipleOutputs(path, container, optionalMessage = NULL)
}
\arguments{
\item{path}{A vector of strings relative to the \code{container}}

\item{container}{A Container or Simulation used to find the entities}

\item{optionalMessage}{A string that will be appended to the end of the message}
}
\value{
A string representation of the error message
}
\description{
Error message when single output is expected but multiple entities produced.
}
