% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-file.R
\name{expandPath}
\alias{expandPath}
\title{Returns the expanded path for \code{path} and ensure that encoding is applied properly}
\usage{
expandPath(path)
}
\arguments{
\item{path}{Path to expand}
}
\description{
Returns the expanded path for \code{path} and ensure that encoding is applied properly
}
\keyword{internal}
