% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-population.R
\name{exportPopulationToCSV}
\alias{exportPopulationToCSV}
\title{Saves the population to csv file}
\usage{
exportPopulationToCSV(population, filePath)
}
\arguments{
\item{population}{Population to export to csv (typically imported from file using \code{loadPopulation})}

\item{filePath}{Full path where the population will be saved.}
}
\description{
Saves the population to csv file
}
\examples{
csvPath <- system.file("extdata", "pop.csv", package = "ospsuite")

# Load the population
population <- loadPopulation(csvPath)

# Exports the population
exportPopulationToCSV(population, tempfile())
}
