% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-snapshots.R
\name{exportProjectToSnapshot}
\alias{exportProjectToSnapshot}
\title{Export a project to a snapshot}
\usage{
exportProjectToSnapshot(..., output = ".")
}
\arguments{
\item{...}{character strings, path to project files (\code{.pksim5}) or a directory
containing project files to convert.}

\item{output}{character string, path to the output directory where to write
the converted snapshot files.}
}
\description{
Converts one or more project files into snapshot files (\code{.json}) and writes
them to an output directory.

Only PK-Sim projects (\code{.pksim5}) are supported for now. Support for MoBi
projects is planned.
}
\examples{
\dontrun{
exportProjectToSnapshot("path/to/project.pksim5", output = "path/to/output")
}
}
