% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ExpressionProfileCategories}
\alias{ExpressionProfileCategories}
\title{Expression Profile Categories}
\format{
An object of class \code{list} of length 3.
}
\usage{
ExpressionProfileCategories
}
\description{
Allowed categories for expression profiles
}
\keyword{datasets}
