% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{extendInitialConditions}
\alias{extendInitialConditions}
\title{Extend an Initial Conditions Building Block (BB) with new entries for molecules
from a molecules BB in all physical containers of a spatial structure BB.}
\usage{
extendInitialConditions(
  initialConditionsBuildingBlock,
  spatialStructureBB,
  moleculesBB,
  moleculeNames = NULL
)
}
\arguments{
\item{initialConditionsBuildingBlock}{A \code{BuildingBlock} object of type \verb{Initial Conditions}.}

\item{spatialStructureBB}{A \code{BuildingBlock} object of type \verb{Spatial Structure}.
Entries will be created for the selected molecules in all physical containers of this
spatial structure.}

\item{moleculesBB}{A \code{BuildingBlock} object of type \code{Molecules}. The entries will be
created for all molecules from this building block, or for a subset of molecules
defined in the \code{moleculeNames} argument.}

\item{moleculeNames}{Optional list of molecule names. If provided, only the molecules
with these names will be added to the \code{initialConditionsBuildingBlock}.}
}
\value{
Paths of entries added to the building block.
}
\description{
If an initial condition for the combination {container, molecule} already exists
in the IC-BB, the value will be kept AS IS.
}
