% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{extractNameAndUnit}
\alias{extractNameAndUnit}
\title{Returns a named list with two entries (name, unit) corresponding to the name and unit
extracted out of the \code{text} provided as parameter}
\usage{
extractNameAndUnit(text)
}
\description{
Returns a named list with two entries (name, unit) corresponding to the name and unit
extracted out of the \code{text} provided as parameter
}
\examples{
res <- extractNameAndUnit("Value [mg]")
# res$name == "Value", res$unit == "mg"

res <- extractNameAndUnit("Value")
# res$name == "Value", res$unit == ""
}
