% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getAllContainerPathsIn}
\alias{getAllContainerPathsIn}
\title{Retrieves the path of all containers defined in the container and all its children}
\usage{
getAllContainerPathsIn(container)
}
\arguments{
\item{container}{A Container or Simulation used to find the parameters}
}
\value{
An array with one entry per container defined in the container
}
\description{
Retrieves the path of all containers defined in the container and all its children
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Returns the path of all molecules defined in the simulation
moleculePaths <- getAllContainerPathsIn(sim)
}
\seealso{
\code{\link[=loadSimulation]{loadSimulation()}}, \code{\link[=getContainer]{getContainer()}} and \code{\link[=getAllContainersMatching]{getAllContainersMatching()}} to retrieve objects of type Container or Simulation
}
