% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getAllContainersMatching}
\alias{getAllContainersMatching}
\title{Retrieve all sub containers of a parent container (simulation or container instance) matching the given path criteria}
\usage{
getAllContainersMatching(path, container)
}
\arguments{
\item{path}{A vector of string relative to the container}

\item{container}{A Container or Simulation used to find the sub containers}
}
\value{
A list of containers matching the path criteria
}
\description{
Retrieve all sub containers of a parent container (simulation or container instance) matching the given path criteria
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Intracellular` containers define in a all direct containers of the organism
containers <- getAllContainersMatching(c("Organism", "*", "Intracellular"), sim)

# Returns all `Intracellular` containers defined in `Organism` and all its subcontainers
containers <- getAllContainersMatching(c("Organism", "**", "Intracellular"), sim)
}
