% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\name{getAllEntityPathsIn}
\alias{getAllEntityPathsIn}
\title{Retrieves all path of entities defined within the container (simulation or container instance)}
\usage{
getAllEntityPathsIn(container, entityType, method = NULL)
}
\arguments{
\item{container}{A Container or Simulation used to find the entities}

\item{entityType}{Type of entity for which the path should be returned.}

\item{method}{Method to call in the underlying .NET class. (optional). If unspecified, the method will be estimated from entity type}
}
\value{
An array of paths (one for each entity found under the container and its sub containers)
The list is empty if no entities matching were found.
}
\description{
Retrieves all path of entities defined within the container (simulation or container instance)
}
\seealso{
\code{\link[=loadSimulation]{loadSimulation()}}, \code{\link[=getContainer]{getContainer()}} and \code{\link[=getAllContainersMatching]{getAllContainersMatching()}} to create objects of type Container or Simulation
}
