% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{getAllMoleculePathsIn}
\alias{getAllMoleculePathsIn}
\title{Retrieves the path of all molecules defined in the container and all its children}
\usage{
getAllMoleculePathsIn(container)
}
\arguments{
\item{container}{A Container or Simulation used to find the parameters}
}
\value{
An array with one entry per molecule defined in the container
}
\description{
Retrieves the path of all molecules defined in the container and all its children
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Returns the path of all molecules defined in the simulation
moleculePaths <- getAllMoleculePathsIn(sim)

}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to retrieve objects of type Container or Simulation
}
