% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{getAllMoleculesMatching}
\alias{getAllMoleculesMatching}
\title{Retrieve all molecules of a container (simulation or container instance) matching the given path criteria}
\usage{
getAllMoleculesMatching(paths, container)
}
\arguments{
\item{paths}{A vector of strings representing the paths relative to the \code{container}}

\item{container}{A Container or Simulation used to find the molecules}
}
\value{
A list of molecules matching the path criteria. The list is empty if no molecules matching were found.
}
\description{
Retrieve all molecules of a container (simulation or container instance) matching the given path criteria
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Drug` molecules defined in all direct containers of the organism
params <- getAllMoleculesMatching("Organism|*|Volume", sim)

# Return all `Drug` molecules defined in all direct containers of the organism
# and the parameter 'Weight (tissue)' of the container 'Liver'
paths <- c("Organism|*|Volume", "Organism|Liver|Weight (tissue)")
params <- getAllMoleculesMatching(paths, sim)

# Returns all `Drug` molecules defined in `Organism` and all its subcontainers
params <- getAllMoleculesMatching("Organism|**|Volume", sim)
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to retrieve objects of type Container or Simulation
}
