% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getAllObserverPathsIn}
\alias{getAllObserverPathsIn}
\title{Retrieves the path of all observers defined in the container and all its children}
\usage{
getAllObserverPathsIn(container)
}
\arguments{
\item{container}{A Container or Simulation used to find the observers}
}
\value{
An array with one entry per observer defined in the container
}
\description{
Retrieves the path of all observers defined in the container and all its children
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Returns the path of all quantities defined in the simulation
observerPaths <- getAllObserverPathsIn(sim)
}
\seealso{
\code{\link[=loadSimulation]{loadSimulation()}}, \code{\link[=getContainer]{getContainer()}} and \code{\link[=getAllContainersMatching]{getAllContainersMatching()}} to retrieve objects of type Container or Simulation
}
