% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getAllParameterPathsIn}
\alias{getAllParameterPathsIn}
\title{Retrieves the path of all parameters defined in the container and all its children}
\usage{
getAllParameterPathsIn(container)
}
\arguments{
\item{container}{A Container or Simulation used to find the parameters}
}
\value{
An array with one entry per parameter defined in the container
}
\description{
Retrieves the path of all parameters defined in the container and all its children
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Returns the path of all quantities defined in the simulation
parameterPaths <- getAllParameterPathsIn(sim)

}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to retrieve objects of type Container or Simulation
}
