% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{getAllParametersForSensitivityAnalysisMatching}
\alias{getAllParametersForSensitivityAnalysisMatching}
\title{Retrieve all parameters of the given simulation matching the given path criteria and also potential candidate
for sensitivity variation}
\usage{
getAllParametersForSensitivityAnalysisMatching(paths, simulation)
}
\arguments{
\item{paths}{A vector of strings representing the path of the parameters (potentially using wildcards)}

\item{simulation}{Simulation used to find the parameters}
}
\value{
A list of parameters matching the path criteria and also candiates for a sensitivity analysis.
The list is empty if no parameters matching were found.
}
\description{
Retrieve all parameters of the given simulation matching the given path criteria and also potential candidate
for sensitivity variation
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Volume` parameters defined in all direct containers of the organism
params <- getAllParametersForSensitivityAnalysisMatching("Organism|*|Volume", sim)
}
