% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getAllQuantitiesMatching}
\alias{getAllQuantitiesMatching}
\title{Retrieve all quantities of a container (simulation or container instance) matching the given path criteria}
\usage{
getAllQuantitiesMatching(paths, container)
}
\arguments{
\item{paths}{A vector of strings relative to the \code{container}}

\item{container}{A Container or Simulation used to find the parameters}
}
\value{
A list of quantities matching the path criteria. The list is empty if no quantity matching were found.
}
\description{
Retrieve all quantities of a container (simulation or container instance) matching the given path criteria
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Volume` quantities defined in all direct containers of the organism
quantities <- getAllQuantitiesMatching("Organism|*|Volume", sim)

# Return all `Volume` quantities defined in all direct containers of the organism
# and the parameter 'Weight (tissue)' of the container 'Liver'
paths <- c("Organism|*|Volume", "Organism|Liver|Weight (tissue)")
quantities <- getAllQuantitiesMatching(paths, sim)

# Returns all `Volume` quantities defined in `Organism` and all its subcontainers
quantities <- getAllQuantitiesMatching("Organism|**|Volume", sim)
}
\seealso{
\code{\link[=loadSimulation]{loadSimulation()}}, \code{\link[=getContainer]{getContainer()}} and \code{\link[=getAllContainersMatching]{getAllContainersMatching()}} to retrieve objects of type Container or Simulation
}
