% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{getAllStateVariableParametersPaths}
\alias{getAllStateVariableParametersPaths}
\title{Get the paths of all state variable parameters of the simulation}
\usage{
getAllStateVariableParametersPaths(simulation)
}
\arguments{
\item{simulation}{\code{Simulation} object}
}
\value{
A list of paths
}
\description{
Get the paths of all state variable parameters of the simulation
}
\details{
List of paths of all state variable parameters.
}
