% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getBaseUnit}
\alias{getBaseUnit}
\title{Get base unit of a dimension}
\usage{
getBaseUnit(quantityOrDimension)
}
\arguments{
\item{quantityOrDimension}{Instance of a quantity from which the dimension will be retrieved or name of dimension}
}
\value{
String name of the base unit.
}
\description{
Get base unit of a dimension
}
