% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getContainer}
\alias{getContainer}
\title{Retrieve a single container by path under the given container}
\usage{
getContainer(path, container)
}
\arguments{
\item{path}{Path of the container}

\item{container}{The parent container used to find the container}
}
\value{
The [Container] with the given path or null if not found
}
\description{
Retrieve a single container by path under the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getContainer(c("Organism", "Liver"), sim)
}
