% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getDimensionForUnit}
\alias{getDimensionForUnit}
\title{Returns the name of dimension that can be used to support the given unit or null if the dimension cannot be found}
\usage{
getDimensionForUnit(unit)
}
\arguments{
\item{unit}{Unit used to find the corresponding dimension}
}
\description{
Returns the name of dimension that can be used to support the given unit or null if the dimension cannot be found
}
\examples{
dim <- getDimensionForUnit("mg")
}
