% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getDimensionsEnum}
\alias{getDimensionsEnum}
\title{Function to return an enum of all available dimensions}
\usage{
getDimensionsEnum()
}
\value{
enum of all dimensions
}
\description{
Function to return an enum of all available dimensions
}
\examples{

ospsuite:::getDimensionsEnum()

}
\keyword{internal}
