% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getDimensionTask}
\alias{getDimensionTask}
\title{Return an instance of the \code{.NET} Task \code{DimensionTask}
This is purely for optimization purposes}
\usage{
getDimensionTask()
}
\value{
An instance of the Task
}
\description{
Return an instance of the \code{.NET} Task \code{DimensionTask}
This is purely for optimization purposes
}
