% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\name{getEntity}
\alias{getEntity}
\title{Retrieve a single entity by path in the given container}
\usage{
getEntity(path, container, entityType, stopIfNotFound = TRUE)
}
\arguments{
\item{path}{A string representing the path relative to the \code{container}}

\item{container}{A Container or Simulation used to find the entities}

\item{entityType}{Class of the type that should be returned. Supported types are Container, Quantity, and Parameter}

\item{stopIfNotFound}{Boolean. If TRUE and no parameter exist for the given path,
an error is thrown. Default is TRUE.}
}
\value{
The \code{Entity} with the given path coerced to the \code{entityType}.
If the entity for the path does not exist, an error is thrown in case of
\code{stopIfNotFound} is TRUE (default), otherwise \code{NULL}
}
\description{
Retrieve a single entity by path in the given container
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to create objects of type Container or Simulation
}
