% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{getEnumKey}
\alias{getEnumKey}
\title{Get the key mapped to the given value in an \code{enum}}
\usage{
getEnumKey(enum, value)
}
\arguments{
\item{enum}{The enum where the key-value pair is stored}

\item{value}{The value that is mapped to the key}
}
\value{
Key under which the value is stored. If the value is not in the enum, \code{NULL} is returned
}
\description{
Get the key mapped to the given value in an \code{enum}
}
