% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{getMolecule}
\alias{getMolecule}
\title{Retrieve a single molecule by path in the given container}
\usage{
getMolecule(path, container, stopIfNotFound = TRUE)
}
\arguments{
\item{path}{A string representing the path relative to the \code{container}}

\item{container}{A Container or Simulation used to find the molecules}

\item{stopIfNotFound}{Boolean. If TRUE and no molecule exist for the given path,
an error is thrown. Default is TRUE.}
}
\value{
The \code{Molecule} with the given path. If the molecule for the path
does not exist, an error is thrown if \code{stopIfNotFound} is TRUE (default),
otherwise \code{NULL}
}
\description{
Retrieve a single molecule by path in the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
molecule <- getMolecule("**|Intrcelulalr|A", sim)
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to retrieve objects of type Container or Simulation
}
