% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getMolWeightFor}
\alias{getMolWeightFor}
\title{Retrieve molecular weight for a quantity's molecule}
\usage{
getMolWeightFor(quantity, unit = NULL, stopIfNotFound = FALSE)
}
\arguments{
\item{quantity}{A \code{Quantity} object.}

\item{unit}{Optional. Target unit for the molecular weight. Defaults to \verb{kg/\\u00b5mol}.}

\item{stopIfNotFound}{Logical. If \code{TRUE}, throws an error when the molecular
weight cannot be retrieved. If \code{FALSE}, returns \code{NA}. Default is \code{FALSE}.}
}
\value{
The molecular weight in the specified unit or \code{NA} if not found.
}
\description{
Returns the molecular weight of the molecule for a given quantity.
If no unit is provided, the value is returned in the base unit (\code{kg/µmol}).
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
parameterPath <- "Organism|Lumen|Stomach|Dapagliflozin"
quantity <- getQuantity(parameterPath, container = sim)
getMolWeightFor(quantity, unit = "g/mol")
}
