% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{getNetTask}
\alias{getNetTask}
\alias{getContainerTask}
\title{Get an Instance of the Specified \code{.NET} Task}
\usage{
getNetTask(taskName)

getContainerTask()
}
\arguments{
\item{taskName}{The name of the task to retrieve (\strong{without} \code{Get} prefix).}
}
\value{
\itemize{
\item \code{getNetTask()} returns an instance of of the specified \code{.NET} task.
\item \code{getContainerTask()} returns a reference to the container task for
optimization purposes only.
}
}
\description{
Get an Instance of the Specified \code{.NET} Task
}
\examples{

if (FALSE) {
  simulationLoader <- getNetTask("SimulationLoader")
}

ospsuite:::getContainerTask()

}
\keyword{internal}
