% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-env.R
\name{getOSPSuiteSetting}
\alias{getOSPSuiteSetting}
\title{Get the value of a global ospsuite-R setting.}
\usage{
getOSPSuiteSetting(settingName)
}
\arguments{
\item{settingName}{String name of the setting}
}
\value{
Value of the setting stored in ospsuiteEnv. If the setting does not exist, an error is thrown.
}
\description{
Get the value of a global ospsuite-R setting.
}
\examples{
getOSPSuiteSetting("suiteVersion")
getOSPSuiteSetting("sensitivityAnalysisConfig")$totalSensitivityThreshold
}
