% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getParameter}
\alias{getParameter}
\title{Retrieve a single parameter by path in the given container}
\usage{
getParameter(path, container)
}
\arguments{
\item{path}{Path of the parameter}

\item{container}{The container used to find the parameter}
}
\value{
[Parameter] The parameter with the given path or null if not found
}
\description{
Retrieve a single parameter by path in the given container
}
\examples{
\dontrun{
getParameter(c("Organism", "Liver", "Volume"), sim))
}
}
