% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getParameterDisplayPaths}
\alias{getParameterDisplayPaths}
\title{Retrieves the display path of the parameters defined by paths in the simulation}
\usage{
getParameterDisplayPaths(paths, simulation)
}
\arguments{
\item{paths}{A single string or array of paths path relative to the \code{container}}

\item{simulation}{A simulation used to find the entities}
}
\value{
a display path for each parameter in paths
}
\description{
Retrieves the display path of the parameters defined by paths in the simulation
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
displayPath <- getParameterDisplayPaths("Organism|Liver|Volume", sim)
}
