% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getParameters}
\alias{getParameters}
\title{Retrieve all parameters of a container (simulation or container instance) matching the given path criteria}
\usage{
getParameters(path, container)
}
\arguments{
\item{path}{A vector of string relative to the container}

\item{container}{A Container or Simulation used to find the parameters}
}
\value{
List[Parameter] a list of parameters matching the path criteria
}
\description{
Retrieve all parameters of a container (simulation or container instance) matching the given path criteria
}
\examples{
\dontrun{
getParameters(c("Organism", "*", "Volume"), sim)) will return all `Volume` parameters define in a all direct container of the organism
getParameters(c("Organism", "**", "Volume"), sim)) will return all `Volume` parameters defined in `Organism` and all its subcontainers
}
}
