% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dot-net.R
\name{getPropertyValue}
\alias{getPropertyValue}
\title{Returns the value of property named \code{propertyName} from .NET object instance \code{netObject}}
\usage{
getPropertyValue(netObject, propertyName)
}
\arguments{
\item{netObject}{.NET object}

\item{propertyName}{Property name that should be retrieved from the \code{netObject}}
}
\description{
Returns the value of property named \code{propertyName} from .NET object instance \code{netObject}
}
\keyword{internal}
