% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dot-net.R
\name{getPropertyValues}
\alias{getPropertyValues}
\title{Returns a list containing all properties \code{propertyName} from the .NET objects \code{netObjects}}
\usage{
getPropertyValues(netObjects, propertyName)
}
\arguments{
\item{netObjects}{List of .NET object}

\item{propertyName}{Property name that should be retrieved from the \code{netObjects}}
}
\description{
Returns a list containing all properties \code{propertyName} from the .NET objects \code{netObjects}
}
\keyword{internal}
