% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getQuantity}
\alias{getQuantity}
\title{Retrieve a single quantty by path in the given container}
\usage{
getQuantity(path, container)
}
\arguments{
\item{path}{A string representing the path relative to the \code{container}}

\item{container}{A Container or Simulation used to find the parameters}
}
\value{
The \code{Quantity} with the given path or \code{NULL} if not found
}
\description{
Retrieve a single quantty by path in the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
quantity <- getQuantity("Organism|Liver|Volume", sim)
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to create objects of type Container or Simulation
}
