% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getQuantityDisplayPaths}
\alias{getQuantityDisplayPaths}
\title{Retrieves the display path of the quantity defined by path in the simulation}
\usage{
getQuantityDisplayPaths(paths, simulation)
}
\arguments{
\item{paths}{A single string or array of paths path relative to the \code{simulation}}

\item{simulation}{A simulation used to find the entities}
}
\value{
a display path for each entry in paths
}
\description{
Retrieves the display path of the quantity defined by path in the simulation
}
