% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-type.R
\name{getQuantityTypeAsString}
\alias{getQuantityTypeAsString}
\title{getQuantityTypeAsString}
\usage{
getQuantityTypeAsString(quantityInt)
}
\arguments{
\item{quantityInt}{Integer defining the quantity type as returned by the .NET
object}
}
\value{
String representation of the quantity type
}
\description{
Convert an integer returned by the .NET class representing the quantity type
to the string representation
}
