% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{getQuantityValuesByPath}
\alias{getQuantityValuesByPath}
\title{Get the values of quantities in the simulation by path}
\usage{
getQuantityValuesByPath(
  quantityPaths,
  simulation,
  units = NULL,
  stopIfNotFound = TRUE
)
}
\arguments{
\item{quantityPaths}{A single or a list of absolute quantity paths}

\item{simulation}{Simulation containing the quantities}

\item{units}{A string or a list of strings defining the units of returned
values. If \code{NULL} (default), values are returned in base units. If not
\code{NULL}, must have the same length as \code{quantityPaths}. Single entries may be
\code{NULL}.}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no quantity exists for
the given path, an error is thrown. If \code{FALSE}, a warning is shown to the
user.}
}
\description{
Get the values of quantities in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
getQuantityValuesByPath(
  list("Organism|Liver|Volume", "Organism|Liver|A"),
  sim, list("ml", NULL)
)
}
