% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-results.R
\name{getResultsValues}
\alias{getResultsValues}
\title{Get simulated values of molecules or parameters.}
\usage{
getResultsValues(simulationResults, quantitiesOrPaths)
}
\arguments{
\item{simulationResults}{Object of type \code{SimulationResults} produced by calling \code{runSimulation}
on a \code{Simulation} object.}

\item{quantitiesOrPaths}{Quantity instances (element or vector) (typically retrieved using
\code{getAllQuantitiesMatching}) or quantity path (element or vector of strings) for which the results are to be returned.
When providing the paths, only absolute full paths are supported (i.e., no matching with '*' possible)}
}
\value{
A list of results, each entry having the name of a quantity path and consisting of the element \code{x} being the time values (in minutes)
and the element \code{y} a \code{Nx1} matrix of simulated values, with \code{N} being the number of entries (equals the number of output time steps).
If no results were generated for a given path, the value of the element is \code{NULL}.
}
\description{
Return values for given entities generated in a individual simulation.
}
\details{
The function receives an object of simulation results generated by running the simulation
and returns time-values profiles for the chosen quantities. If the results contain more then one individual,
an error is thrown. In this case, \code{getPopulationResultsValues} should be used.
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
simResults <- runSimulation(sim)

paths <- "Organism|Volume"

resultsValues <- getResultsValues(simResults, paths)
plot(resultsValues[[paths]])
}
\seealso{
\code{\link{getPopulationResultsValues}}
}
