% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{getStandardMoleculeParameters}
\alias{getStandardMoleculeParameters}
\title{Returns a list containing all standard global parameters defined in a \code{simulation} for given \code{moleculeName}.
These parameters are typically located directly under the container named after the \code{moleculeName}.
For the list of standard parameters}
\usage{
getStandardMoleculeParameters(moleculeName, simulation)
}
\arguments{
\item{moleculeName}{Name of molecule (Enzyme, Transporter etc..) for which global parameters should be returned}

\item{simulation}{Simulation to query for molecule parameters}
}
\value{
A list of all standard global parameters defined for \code{moleculeName} if the molecule exists in the \code{simulation}.
Otherwise an empty list is returned
}
\description{
Returns a list containing all standard global parameters defined in a \code{simulation} for given \code{moleculeName}.
These parameters are typically located directly under the container named after the \code{moleculeName}.
For the list of standard parameters
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim1 <- loadSimulation(simPath)

parameters <- getStandardMoleculeParameters("CYP3A4", sim1)
}
\seealso{
\link{MoleculeParameter}
}
