% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dimensions.R
\name{getUnitConversionFactor}
\alias{getUnitConversionFactor}
\title{Get conversion factor from one unit to another}
\usage{
getUnitConversionFactor(fromUnit, toUnit, dimension)
}
\arguments{
\item{fromUnit}{String name of the unit to convert from}

\item{toUnit}{String name of the unit to convert to}

\item{dimension}{Dimension of the units}
}
\value{
Numerical value. A value in \code{fromUnit} multiplied  by the factor
corresponds to the value in \code{toUnit}
}
\description{
Get conversion factor from one unit to another
}
