% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getUnitsEnum}
\alias{getUnitsEnum}
\title{Loop through dimensions and build a list containing an enum of all units available for each dimension}
\usage{
getUnitsEnum()
}
\value{
enum of all units for each dimension
}
\description{
Loop through dimensions and build a list containing an enum of all units available for each dimension
}
\keyword{internal}
