% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getUnitsForDimension}
\alias{getUnitsForDimension}
\title{Returns a vector containing all units defined in the dimension}
\usage{
getUnitsForDimension(dimension)
}
\arguments{
\item{dimension}{Name of dimension for which units should be returned}
}
\description{
Returns a vector containing all units defined in the dimension
}
\examples{
units <- getUnitsForDimension("Mass")
}
