% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{Convenience function to avoid testing for null. It returns the first object if it is not null otherwise the second object}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Object that will be returned if not NULL}

\item{rhs}{Object that will be returned if \code{lhs} is NULL. It maybe well be NULL}
}
\value{
The first parameter if it is not NULL otherwise the second parameter
}
\description{
Convenience function to avoid testing for null. It returns the first object if it is not null otherwise the second object
}
