% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dimensions.R
\name{hasDimension}
\alias{hasDimension}
\title{Dimension existence}
\usage{
hasDimension(dimension)
}
\arguments{
\item{dimension}{String name of the dimension.}
}
\description{
Dimension existence
}
\details{
Returns \code{TRUE} if the provided dimension is supported otherwise \code{FALSE}
}
