% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{hasUnit}
\alias{hasUnit}
\title{Unit existence}
\usage{
hasUnit(unit, dimension)
}
\arguments{
\item{unit}{String name of the unit}

\item{dimension}{String name of the dimension.}
}
\description{
Unit existence
}
\details{
Check if the unit is valid for the dimension.
}
