% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/enums.R
=======
% Please edit documentation in R/pk-sim.R
>>>>>>> c4fbe917 (Fixes #1834 refactor DataCombined plot examples)
\name{HumanPopulation}
\alias{HumanPopulation}
\title{Default human population defined in PK-Sim}
\usage{
HumanPopulation
}
\description{
Default human population defined in PK-Sim
}
