% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{importSensitivityAnalysisResultsFromCSV}
\alias{importSensitivityAnalysisResultsFromCSV}
\title{Imports the simulation analysis results from one or more csv files}
\usage{
importSensitivityAnalysisResultsFromCSV(simulation, filePaths)
}
\arguments{
\item{simulation}{Instance of a simulation for which the sensitivity analysis was performed}

\item{filePaths}{Full path of sensitivity analysis result files to import. Typically only one
file is provided but a list of files is sometimes available when the sensitivity analysis run
was parallelized and computed on different machines}
}
\description{
Imports the simulation analysis results from one or more csv files
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
resultPath <- system.file("extdata", "sa.csv", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

# Run the simulation
results <- importSensitivityAnalysisResultsFromCSV(sim, resultPath)
}
