% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-building-block.R
\docType{class}
\name{IndividualBuildingBlock}
\alias{IndividualBuildingBlock}
\title{IndividualBuildingBlock}
\description{
An \code{Individual} building block. Subclass of \link{BuildingBlock} that
exposes the demographic properties of the underlying individual as read-only
fields (species, population, gender, age, gestational age, height, weight).
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ObjectBase}} -> \code{\link[ospsuite:BuildingBlock]{BuildingBlock}} -> \code{IndividualBuildingBlock}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{species}}{Species of the individual. Read-only.}

    \item{\code{population}}{Population of the individual. May be \code{NULL} for
non-human species. Read-only.}

    \item{\code{gender}}{Gender of the individual. Read-only.}

    \item{\code{age}}{Age of the individual. Read-only.}

    \item{\code{gestationalAge}}{Gestational age of the individual. Read-only.}

    \item{\code{height}}{Height of the individual. Read-only.}

    \item{\code{weight}}{Weight of the individual. Read-only.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-IndividualBuildingBlock-initialize}{\code{IndividualBuildingBlock$new()}}
    \item \href{#method-IndividualBuildingBlock-print}{\code{IndividualBuildingBlock$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printClass"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printClass'><code>rSharp::NetObject$.printClass()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printLine"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printLine'><code>rSharp::NetObject$.printLine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticFields'><code>rSharp::NetObject$getStaticFields()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticMethods'><code>rSharp::NetObject$getStaticMethods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticProperties'><code>rSharp::NetObject$getStaticProperties()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndividualBuildingBlock-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-IndividualBuildingBlock-initialize}{}}}
\subsection{\code{IndividualBuildingBlock$new()}}{
  Initialize a new instance of the class. Origin-data values are read once
from the underlying object and cached on the R6 instance; subsequent
field access does not call back into the simulation engine.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{IndividualBuildingBlock$new(netObject)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{netObject}}{Reference to the underlying \code{Individual} building block.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{IndividualBuildingBlock} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndividualBuildingBlock-print"></a>}}
\if{latex}{\out{\hypertarget{method-IndividualBuildingBlock-print}{}}}
\subsection{\code{IndividualBuildingBlock$print()}}{
  Print the object to the console.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{IndividualBuildingBlock$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Rest arguments.}
    }
    \if{html}{\out{</div>}}
  }
}

}
