% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-individual.R
\docType{data}
\name{IndividualDiseaseStates}
\alias{IndividualDiseaseStates}
\title{Available disease states}
\format{
An object of class \code{list} of length 3.
}
\usage{
IndividualDiseaseStates
}
\description{
MUST WE DEFINE IT HERE, OR CAN WE GET IT FROM .NET SOMEHOW?
}
\keyword{datasets}
