% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-individual.R
\name{individualsBBToDataFrame}
\alias{individualsBBToDataFrame}
\title{Convert an Individual Building Block to a data frame.}
\usage{
individualsBBToDataFrame(individualBuildingBlock)
}
\arguments{
\item{individualBuildingBlock}{A \code{BuildingBlock} object of type \code{Individual}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \verb{Container Path}: Full path to the container where the parameter is located.
\item \verb{Parameter Name}: Name of the parameter.
\item \code{Value}: Value of the parameter. For values that are defined by a formula, the return value can be \code{NaN}.
\item \code{Unit}: Unit of the parameter value.
\item \verb{Value Origin}: Origin of the parameter value.
}
}
\description{
Convert an Individual Building Block to a data frame.
}
\examples{
individual <- createIndividualBuildingBlock(
  species = Species$Human,
  population = HumanPopulation$European_ICRP_2002
)
df <- individualsBBToDataFrame(individual)
}
