% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{initialConditionsBBToDataFrame}
\alias{initialConditionsBBToDataFrame}
\title{Convert an Initial Conditions Building Block to a data frame.}
\usage{
initialConditionsBBToDataFrame(initialConditionsBuildingBlock)
}
\arguments{
\item{initialConditionsBuildingBlock}{A \code{BuildingBlock} object of type \verb{Initial Conditions}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \verb{Container Path}: Full path to the container where the molecule is located.
\item \verb{Molecule Name}: Name of the molecule.
\item \verb{Is Present}: Boolean indicating if the molecule is present.
\item \code{Value}: Initial value of the molecule. For values that are defined by a formula, the return value can be \code{NaN}.
\item \code{Unit}: Unit of the initial value.
\item \verb{Scale Divisor}: Scale divisor for the initial value.
\item \verb{Neg. Values Allowed}: Boolean indicating if negative values are allowed.
}
}
\description{
Convert an Initial Conditions Building Block to a data frame.
}
\examples{
\dontrun{
module <- loadModuleFromPKML("path/to/module.pkml")
icBB <- module$getInitialConditionsBBs()[[1]]
df <- initialConditionsBBToDataFrame(icBB)
}
}
