% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pksim.R
\name{initPKSim}
\alias{initPKSim}
\title{Loads the \code{PKSim.R} dll that will enable create individual and create population workflows.}
\usage{
initPKSim()
}
\description{
Loads the \code{PKSim.R} dll that will enable create individual and create population workflows.
}
\note{
PKSim dlls are included in the package.
}
