% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{isExplicitFormulaByPath}
\alias{isExplicitFormulaByPath}
\title{Is the value defined by an explicit formula}
\usage{
isExplicitFormulaByPath(path, simulation, stopIfNotFound = TRUE)
}
\arguments{
\item{path}{Path to the quantity}

\item{simulation}{A \code{Simulation} object that contains the quantity}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no quantity exists
for the given path, an error is thrown. If \code{FALSE}, \code{FALSE} is returned.}
}
\value{
\code{TRUE} if the value is an explicit formula, \code{FALSE} otherwise.
Also returns \code{FALSE} if no quantity with the given path is found and
\code{stopInfNotFound} is set to \code{FALSE}.
}
\description{
Is the value defined by an explicit formula
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
isExplicitFormulaByPath("Organism|Liver|Volume", sim) # FALSE
}
