% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{isSupportedUnit}
\alias{isSupportedUnit}
\title{Check if unit is supported in the OSPSuite platform}
\usage{
isSupportedUnit(unit)
}
\arguments{
\item{unit}{String name of the unit}
}
\value{
Logical: \code{TRUE} if the unit is supported, \code{FALSE} otherwise.
}
\description{
Check if unit is supported in the OSPSuite platform
}
\details{
Returns \code{TRUE} if the provided unit is supported in the OSPSuite platform, otherwise \code{FALSE}
}
\examples{
isSupportedUnit("kg")
isSupportedUnit("invalid_unit")

}
