% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-population.R
\name{loadAgingDataFromCSV}
\alias{loadAgingDataFromCSV}
\title{Loads aging data (typically generated from PK-Sim) i}
\usage{
loadAgingDataFromCSV(filePath)
}
\arguments{
\item{filePath}{Full path containing an aging data table.}
}
\description{
Loads aging data (typically generated from PK-Sim) i
}
\examples{
csvPath <- system.file("extdata", "aging_data.csv", package = "ospsuite")

agingData <- loadAgingDataFromCSV(tempfile())
}
