% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-importer-configuration.R
\name{loadDataImporterConfiguration}
\alias{loadDataImporterConfiguration}
\title{Load \code{DataImporterConfiguration} from XML file.}
\usage{
loadDataImporterConfiguration(configurationFilePath)
}
\arguments{
\item{configurationFilePath}{Path to the XML file with stored configuration
(e.g. created in PK-Sim or MoBi).}
}
\value{
A new \code{DataImporterConfiguration} object to be used with
\code{loadDataSetsFromExcel()}.
}
\description{
Load \code{DataImporterConfiguration} from XML file.
}
\examples{

configurationFilePath <- system.file(
  "extdata", "dataImporterConfiguration.xml",
  package = "ospsuite"
)

importerConfiguration <- loadDataImporterConfiguration(configurationFilePath)

# Specifying which sheet to load
importerConfiguration$sheets <- "TestSheet_1"
xlsFilePath <- system.file("extdata", "CompiledDataSet.xlsx", package = "ospsuite")
dataSets <- loadDataSetsFromExcel(
  xlsFilePath = xlsFilePath,
  importerConfigurationOrPath = importerConfiguration,
  importAllSheets = FALSE
)

}
