% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-set.R
\name{loadDataSetFromPKML}
\alias{loadDataSetFromPKML}
\title{Loads data (typically observed data) from a PKML file and creates a \code{DataSet} from it.
The pkml files are typically exported from PK-Sim or MoBi.}
\usage{
loadDataSetFromPKML(filePath)
}
\arguments{
\item{filePath}{Full path of pkml file containing the observed data to load}
}
\description{
Loads data (typically observed data) from a PKML file and creates a \code{DataSet} from it.
The pkml files are typically exported from PK-Sim or MoBi.
}
\examples{
filePath <- system.file("extdata", "obs_data.pkml", package = "ospsuite")

obsData <- loadDataSetFromPKML(filePath)
}
