% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-project.R
\name{loadMoBiProject}
\alias{loadMoBiProject}
\title{Load a MoBi project}
\usage{
loadMoBiProject(filePath)
}
\arguments{
\item{filePath}{Path of 'mbp3' MoBi project file to load.}
}
\value{
An object of the \code{MoBiProject} type.
}
\description{
Load a MoBi project
}
\examples{
projectPath <- system.file("extdata", "simple.mbp3", package = "ospsuite")

myProject <- loadMoBiProject(projectPath)
}
