% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-module.R
\name{loadModuleFromPKML}
\alias{loadModuleFromPKML}
\title{Load a MoBi module from pkml
TODO not implemented https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1590}
\usage{
loadModuleFromPKML(path)
}
\arguments{
\item{path}{Path to the pkml file}
}
\value{
A \code{MoBiModule} object
}
\description{
Load a MoBi module from pkml
TODO not implemented https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1590
}
