% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{loadObservedData}
\alias{loadObservedData}
\title{Loads a population from a csv file and returns the population.}
\usage{
loadObservedData(filePath)
}
\arguments{
\item{filePath}{Full path of pkml file containing the observed data to load}
}
\description{
Loads a population from a csv file and returns the population.
}
\examples{
filePath <- system.file("extdata", "obsData.pkml", package = "ospsuite")

obsData <- loadObservedData(filePath)
}
