% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-population.R
\name{loadPopulation}
\alias{loadPopulation}
\title{Loads a population from a csv file and returns the population.}
\usage{
loadPopulation(csvPopulationFile)
}
\arguments{
\item{csvPopulationFile}{Full path of csv population file to load.}
}
\description{
Loads a population from a csv file and returns the population.
}
\examples{
csvPath <- system.file("extdata", "pop.csv", package = "ospsuite")

population <- loadPopulation(csvPath)
}
