% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-snapshots.R
\name{loadProjectFromSnapshot}
\alias{loadProjectFromSnapshot}
\title{Load a project from a snapshot}
\usage{
loadProjectFromSnapshot(..., output = ".", runSimulations = FALSE)
}
\arguments{
\item{...}{character strings, path to snapshot files (\code{.json}) or a directory
containing snapshot files to convert.}

\item{output}{character string, path to the output directory where to write
the converted project files.}

\item{runSimulations}{logical, whether to run the simulations during
conversion (default = \code{FALSE}).}
}
\description{
Converts one or more snapshot files (\code{.json}) into project files and writes
them to an output directory.

Only PK-Sim projects (\code{.pksim5}) are supported for now. Support for MoBi
projects is planned.
}
\examples{
\dontrun{
loadProjectFromSnapshot("path/to/snapshot.json", output = "path/to/output")
}
}
