% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{loadSimulation}
\alias{loadSimulation}
\title{Loads a simulation from a pkml file and returns the simulation}
\usage{
loadSimulation(pkmlSimulationFile)
}
\arguments{
\item{pkmlSimulationFile}{Full path of pkml simulation file to load.}
}
\value{
The \code{Simulation}individual results (one entry per Individual)
}
\description{
Loads a simulation from a pkml file and returns the simulation
}
