% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{MergeBehavior}
\alias{MergeBehavior}
\title{Merge Behaviors}
\usage{
MergeBehavior
}
\description{
Merge behaviors available for modules in MoBi. Loaded at
package init from \code{OSPSuite.Core.Domain.MergeBehavior}. Used by
\code{MoBiModule$mergeBehavior}.
}
