% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecules-building-block.R
\docType{class}
\name{MoleculesBuildingBlock}
\alias{MoleculesBuildingBlock}
\title{MoleculesBuildingBlock}
\description{
A \code{Molecules} building block. Subclass of \link{BuildingBlock}.
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ObjectBase}} -> \code{\link[ospsuite:BuildingBlock]{BuildingBlock}} -> \code{MoleculesBuildingBlock}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MoleculesBuildingBlock-initialize}{\code{MoleculesBuildingBlock$new()}}
    \item \href{#method-MoleculesBuildingBlock-allMoleculeNames}{\code{MoleculesBuildingBlock$allMoleculeNames()}}
    \item \href{#method-MoleculesBuildingBlock-allFloatingMoleculeNames}{\code{MoleculesBuildingBlock$allFloatingMoleculeNames()}}
    \item \href{#method-MoleculesBuildingBlock-allStationaryMoleculeNames}{\code{MoleculesBuildingBlock$allStationaryMoleculeNames()}}
    \item \href{#method-MoleculesBuildingBlock-allMoleculeNamesOfType}{\code{MoleculesBuildingBlock$allMoleculeNamesOfType()}}
    \item \href{#method-MoleculesBuildingBlock-allXenobioticFloatingMoleculeNames}{\code{MoleculesBuildingBlock$allXenobioticFloatingMoleculeNames()}}
    \item \href{#method-MoleculesBuildingBlock-allEndogenousStationaryMoleculeNames}{\code{MoleculesBuildingBlock$allEndogenousStationaryMoleculeNames()}}
    \item \href{#method-MoleculesBuildingBlock-moleculeTypeFor}{\code{MoleculesBuildingBlock$moleculeTypeFor()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printClass"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printClass'><code>rSharp::NetObject$.printClass()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printLine"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printLine'><code>rSharp::NetObject$.printLine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticFields'><code>rSharp::NetObject$getStaticFields()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticMethods'><code>rSharp::NetObject$getStaticMethods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticProperties'><code>rSharp::NetObject$getStaticProperties()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ospsuite" data-topic="BuildingBlock" data-id="print"><a href='../../ospsuite/html/BuildingBlock.html#method-BuildingBlock-print'><code>BuildingBlock$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-initialize}{}}}
\subsection{\code{MoleculesBuildingBlock$new()}}{
  Initialize a new instance of the class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$new(netObject)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{netObject}}{Reference to the underlying \code{Molecules} building block.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{MoleculesBuildingBlock} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allMoleculeNames}{}}}
\subsection{\code{MoleculesBuildingBlock$allMoleculeNames()}}{
  Returns the names of all molecules defined in the building block.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allMoleculeNames()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allFloatingMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allFloatingMoleculeNames}{}}}
\subsection{\code{MoleculesBuildingBlock$allFloatingMoleculeNames()}}{
  Returns the names of all floating molecules in the building block
(molecules with \code{IsFloating = TRUE}, e.g. drugs, metabolites).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allFloatingMoleculeNames()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allStationaryMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allStationaryMoleculeNames}{}}}
\subsection{\code{MoleculesBuildingBlock$allStationaryMoleculeNames()}}{
  Returns the names of all stationary molecules in the building block
(molecules with \code{IsFloating = FALSE}, e.g. enzymes, transporters).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allStationaryMoleculeNames()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allMoleculeNamesOfType"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allMoleculeNamesOfType}{}}}
\subsection{\code{MoleculesBuildingBlock$allMoleculeNamesOfType()}}{
  Returns the names of all molecules of the given type. Pass
\code{MoleculeType$Protein} to obtain all proteins (the union of
\code{Enzyme}, \code{Transporter}, and \verb{Binding Partner}).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allMoleculeNamesOfType(moleculeType)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{moleculeType}}{One of the values defined in the \code{MoleculeType} enum
(e.g. \code{Drug}, \code{Metabolite}, \code{Enzyme}, \code{Transporter}, \verb{Binding Partner},
\code{Complex}, \code{Protein}).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allXenobioticFloatingMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allXenobioticFloatingMoleculeNames}{}}}
\subsection{\code{MoleculesBuildingBlock$allXenobioticFloatingMoleculeNames()}}{
  Returns the names of all xenobiotic floating molecules in the building
block (e.g. drugs, drug complexes).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allXenobioticFloatingMoleculeNames()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-allEndogenousStationaryMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-allEndogenousStationaryMoleculeNames}{}}}
\subsection{\code{MoleculesBuildingBlock$allEndogenousStationaryMoleculeNames()}}{
  Returns the names of all endogenous stationary molecules in the
building block (e.g. enzymes, transporters, other proteins).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$allEndogenousStationaryMoleculeNames()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of molecule names.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoleculesBuildingBlock-moleculeTypeFor"></a>}}
\if{latex}{\out{\hypertarget{method-MoleculesBuildingBlock-moleculeTypeFor}{}}}
\subsection{\code{MoleculesBuildingBlock$moleculeTypeFor()}}{
  Returns the type of a molecule by name as it is recorded in the
building block (e.g. \code{"Drug"}, \code{"Enzyme"}, \code{"Transporter"},
\code{"Binding Partner"}). Throws an error if the molecule is not present.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MoleculesBuildingBlock$moleculeTypeFor(moleculeName)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{moleculeName}}{Name of the molecule to look up.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character string with the molecule type.
  }
}

}
