% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{MoleculeType}
\alias{MoleculeType}
\title{Molecule Types}
\usage{
MoleculeType
}
\description{
Allowed types for molecules in a \code{Molecules} building block.
Values are \code{Drug}, \code{Metabolite}, \code{Enzyme}, \code{Transporter}, \verb{Binding Partner},
\code{Complex}, and the umbrella \code{Protein} (= \code{Enzyme} + \code{Transporter} +
\verb{Binding Partner}). Used to filter molecule names via
\code{MoleculesBuildingBlock$allMoleculeNamesOfType()}.
}
