% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{netEnumName}
\alias{netEnumName}
\title{Retrieves the name of the constant in the specified enumeration that has the specified value.}
\usage{
netEnumName(enumType, enumValue)
}
\arguments{
\item{enumType}{a CLR object, System.Type or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{enumValue}{The value of a particular enumerated constant in terms of its underlying type. Typically an integer.}
}
\value{
A string containing the name of the enumerated constant in enumType whose value is enumValue; or null if no such constant is found.
}
\description{
Retrieves the name of the constant in the specified enumeration that has the specified value.
}
