% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin-data.R
\docType{class}
\name{OriginData}
\alias{OriginData}
\title{OriginData}
\usage{
OriginData
}
\description{
Simulation Interval (typically associated with an instace of \code{OutputSchema})
}
\section{Fields}{

\describe{
\item{\code{startTime}}{Start time of interval (instance of \code{Parameter})}

\item{\code{endTime}}{End time of interval (instance of \code{Parameter})}

\item{\code{resolution}}{Resolution of interval in pts/min (instance of \code{Parameter})}
}}

\keyword{datasets}
