% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{ospDimensions}
\alias{ospDimensions}
\title{Supported dimensions defined as a named list}
\usage{
ospDimensions
}
\description{
Supported dimensions defined as a named list
}
\details{
ospDimensions$Mass => "Mass"
}
