% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospPlotConfiguration.R
\name{ospPlotConfiguration}
\alias{ospPlotConfiguration}
\title{Plot configuration for OSP plots}
\description{
R6 class defining the configuration of a \code{{ospsuite}} plots.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ospPlotConfiguration$new()}}
\item \href{#method-clone}{\code{ospPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ospPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ospPlotConfiguration$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  legend = NULL,
  xAxis = NULL,
  yAxis = NULL,
  background = NULL,
  plotArea = NULL,
  panelArea = NULL,
  xGrid = NULL,
  yGrid = NULL,
  watermark = NULL,
  lines = NULL,
  points = NULL,
  ribbons = NULL,
  errorbars = NULL,
  export = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{character or \code{Label} object defining plot title}

\item{\code{subtitle}}{character or \code{Label} object defining plot subtitle}

\item{\code{xlabel}}{character or \code{Label} object defining plot xlabel}

\item{\code{ylabel}}{character or \code{Label} object defining plot ylabel}

\item{\code{legend}}{\code{LegendConfiguration} object defining legend properties}

\item{\code{xAxis}}{\code{XAxisConfiguration} object defining x-axis properties}

\item{\code{yAxis}}{\code{YAxisConfiguration} object defining y-axis properties}

\item{\code{background}}{\code{BackgroundConfiguration} object defining background properties}

\item{\code{plotArea}}{\code{BackgroundElement} object defining properties of plot area}

\item{\code{panelArea}}{\code{BackgroundElement} object defining properties of panel area}

\item{\code{xGrid}}{\code{LineElement} object defining properties of x-grid background}

\item{\code{yGrid}}{\code{LineElement} object defining properties of y-grid background}

\item{\code{watermark}}{\code{Label} object defining watermark}

\item{\code{lines}}{\code{ThemeAestheticSelections} object or list defining how lines are plotted}

\item{\code{points}}{\code{ThemeAestheticSelections} object or list defining how points are plotted}

\item{\code{ribbons}}{\code{ThemeAestheticSelections} object or list defining how ribbons are plotted}

\item{\code{errorbars}}{\code{ThemeAestheticSelections} object or list defining how errorbars are plotted}

\item{\code{export}}{R6 class \code{SaveConfiguration} defining saving properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ospPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
