% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{ospsuite_deprecated}
\alias{ospsuite_deprecated}
\alias{pkAnalysesAsDataFrame}
\alias{populationAsDataFrame}
\title{Deprecated functions}
\usage{
pkAnalysesAsDataFrame(...)

populationAsDataFrame(...)
}
\arguments{
\item{...}{Arguments to the deprecated function.}
}
\description{
Deprecated functions
}
\details{
\itemize{
\item \code{pkAnalysesAsDataFrame} is now \code{\link{pkAnalysesToDataFrame}}.
\item \code{populationAsDataFrame} is now \code{\link{populationToDataFrame}}.
}
}
