% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-env.R
\name{ospsuiteSettingNames}
\alias{ospsuiteSettingNames}
\title{Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteSetting()}}
\usage{
ospsuiteSettingNames
}
\description{
Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteSetting()}
}
