% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\docType{data}
\name{ospUnits}
\alias{ospUnits}
\title{Supported units defined as a named list of lists}
\format{
An object of class \code{list} of length 91.
}
\usage{
ospUnits
}
\description{
ospUnits$Mass$kg => "kg"
}
\keyword{datasets}
