% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter-value.R
\name{parameterValueListFrom}
\alias{parameterValueListFrom}
\title{Converts a list of .NET \code{ParameterValue} into a list with 3 entries: \code{paths}, \code{values}, and \code{percentiles}.}
\usage{
parameterValueListFrom(netParameterValues)
}
\arguments{
\item{netParameterValues}{List of.NET \code{ParameterValue}}
}
\value{
A list with 3 sublist: \code{paths}, \code{values}, and \code{percentiles} containing the corresponding values from each parameter value
}
\description{
Converts a list of .NET \code{ParameterValue} into a list with 3 entries: \code{paths}, \code{values}, and \code{percentiles}.
}
