% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{parameterValuesToDataFrame}
\alias{parameterValuesToDataFrame}
\title{Convert a Parameter Values Building Block to a data frame.}
\usage{
parameterValuesToDataFrame(parameterValuesBuildingBlock)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \verb{Container Path}: Full path to the container where the parameter is located.
\item \verb{Parameter Name}: Name of the parameter.
\item \code{Value}: Value of the parameter. For values that are defined by a formula, the return value can be \code{NaN}.
\item \code{Unit}: Unit of the parameter value.
}
}
\description{
Convert a Parameter Values Building Block to a data frame.
}
\examples{
\dontrun{
module <- loadModuleFromPKML("path/to/module.pkml")
pvBB <- module$getParameterValuesBBs()[[1]]
df <- parameterValuesToDataFrame(pvBB)
}
}
