% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{PartitionCoefficientMethods}
\alias{PartitionCoefficientMethods}
\title{Available methods for calculation of partition coefficients.}
\usage{
PartitionCoefficientMethods
}
\description{
Available methods for calculation of partition coefficients.
}
