% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-analyses.R
\name{pkAnalysesAsDataFrame}
\alias{pkAnalysesAsDataFrame}
\title{Convert the pk-Analysis to data frame}
\usage{
pkAnalysesAsDataFrame(pkAnalyses)
}
\arguments{
\item{pkAnalyses}{pK-Analyses to convert to data frame (typically calculated using \code{calculatePKAnalyses} or imported from file)}
}
\description{
Convert the pk-Analysis to data frame
}
