% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-parameter.R
\name{pkParameterByName}
\alias{pkParameterByName}
\title{Returns an instance of a PK-Parameter by name or NULL if the parameter by name is not found}
\usage{
pkParameterByName(name, stopIfNotFound = TRUE)
}
\arguments{
\item{name}{Name of PK-Parameter to update}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no pk parameter exist for the given name,
an error is thrown. If \code{FALSE}, \code{NULL} is returned.}
}
\description{
Returns an instance of a PK-Parameter by name or NULL if the parameter by name is not found
}
\examples{

pkParameter <- pkParameterByName(name = "t_max")
}
