% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Create a plot grid}
\usage{
plotGrid(ospPlotGridConfiguration)
}
\arguments{
\item{ospPlotGridConfiguration}{A \code{ospPlotGridConfiguration} object, which
is an \code{R6} class object that defines properties of a plot grid (like number
of rows, columns, labels, etc.). You can create instance of this object
using \code{createPlotGridConfiguration()} function.}
}
\description{
Create a plot grid using the \code{patchwork::wrap_plots()} function. The required
arguments are supplied through the \code{ospPlotGridConfiguration} object.
}
