% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual-time-profile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Time-profile plot of individual data}
\usage{
plotIndividualTimeProfile(
  dataCombined,
  defaultPlotConfiguration = NULL,
  showLegendPerDataset = FALSE
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class containing both
observed and simulated datasets to be compared.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{showLegendPerDataset}{Logical flag to display separate legend entries
for observed and simulated datasets, if available. This is experimental and
may not work reliably when both observed and simulated datasets > 1. Defaults
to \code{FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# Create a new instance of `DefaultPlotConfiguration` class
myPlotConfiguration <- DefaultPlotConfiguration$new()
myPlotConfiguration$title <- "My Plot Title"
myPlotConfiguration$subtitle <- "My Plot Subtitle"
myPlotConfiguration$caption <- "My Sources"

# plot
plotIndividualTimeProfile(dataCombinedAciclovir, myPlotConfiguration)

}
\seealso{
Other plotting:
\code{\link{DefaultPlotConfiguration}},
\code{\link[=plotObservedVsSimulated]{plotObservedVsSimulated()}},
\code{\link[=plotPopulationTimeProfile]{plotPopulationTimeProfile()}},
\code{\link[=plotResidualsVsSimulated]{plotResidualsVsSimulated()}},
\code{\link[=plotResidualsVsTime]{plotResidualsVsTime()}}
}
\concept{plotting}
