% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-observed-vs-simulated.R
\name{plotObservedVsSimulated}
\alias{plotObservedVsSimulated}
\title{Observed versus predicted/simulated scatter plot}
\usage{
plotObservedVsSimulated(
  dataCombined,
  defaultPlotConfiguration = NULL,
  foldDistance = NULL
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class containing both
observed and simulated datasets to be compared.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{foldDistance}{A vector for plotting lines at required fold distances
around the identity line (\code{x=y}). Set to NULL (default) to only draw identity
line. Set to FALSE to not draw any lines.
The vector can include only fold distance values \verb{>1}. An \code{x}-fold distance
is defined as all simulated values within the range between \code{x}-fold
(depicted by the upper fold range line) and \code{1/x}-fold
(depicted by the lower fold range line) of observed values. The identity line
can be interpreted as the \code{1}-fold range.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# Create a new instance of `DefaultPlotConfiguration` class
myPlotConfiguration <- DefaultPlotConfiguration$new()
myPlotConfiguration$title <- "My Plot Title"
myPlotConfiguration$subtitle <- "My Plot Subtitle"
myPlotConfiguration$caption <- "My Sources"

# plot
plotObservedVsSimulated(dataCombinedAciclovir, myPlotConfiguration)

}
\seealso{
Other plotting:
\code{\link{DefaultPlotConfiguration}},
\code{\link[=plotIndividualTimeProfile]{plotIndividualTimeProfile()}},
\code{\link[=plotPopulationTimeProfile]{plotPopulationTimeProfile()}},
\code{\link[=plotResidualsVsSimulated]{plotResidualsVsSimulated()}},
\code{\link[=plotResidualsVsTime]{plotResidualsVsTime()}}
}
\concept{plotting}
