% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-population.R
\name{populationFromDataFrame}
\alias{populationFromDataFrame}
\title{Creates a \code{Population} object from a data.frame}
\usage{
populationFromDataFrame(dataFrame)
}
\arguments{
\item{dataFrame}{A data.frame containing population data. Each column represents
a parameter path or covariate, with one row per individual. If no \code{IndividualId}
column is present, one will be automatically generated with 0-based sequential IDs.}
}
\value{
A \code{Population} object
}
\description{
Creates a \code{Population} object from a data.frame
}
\examples{
csvPath <- system.file("extdata", "pop.csv", package = "ospsuite")

population <- loadPopulation(csvPath)
df <- populationToDataFrame(population)
populationFromDf <- populationFromDataFrame(df)
}
