% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{potentialVariableParameterPathsFor}
\alias{potentialVariableParameterPathsFor}
\title{Returns an array of parameter path with one entry for each parameter that is used in the simulation
and can potentially be used for sensitivity analysis}
\usage{
potentialVariableParameterPathsFor(simulation)
}
\arguments{
\item{simulation}{Instance of a simulation for which variable parameters should be retrieved}
}
\description{
Returns an array of parameter path with one entry for each parameter that is used in the simulation
and can potentially be used for sensitivity analysis
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

parameterPaths <- potentialVariableParameterPathsFor(sim)
}
