% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printable.R
\docType{class}
\name{Printable}
\alias{Printable}
\title{Printable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Printable
}
\description{
Base class that implements some basic functionalities for printing to console
}
\keyword{datasets}
