% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-type.R
\docType{data}
\name{QuantityType}
\alias{QuantityType}
\title{All possible quantity types as defined in OSPSuite}
\format{An object of class \code{list} of length 10.}
\usage{
QuantityType
}
\description{
All possible quantity types as defined in OSPSuite
}
\keyword{datasets}
