% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{removeSimulationFromCache}
\alias{removeSimulationFromCache}
\title{Removes a simulation from simulations cache.}
\usage{
removeSimulationFromCache(simulation)
}
\arguments{
\item{simulation}{Simulation to  be removed from the cache}
}
\value{
TRUE if the simulation was cached and could be removed from cache.
FALSE otherwise, usually indicating that the specific simulation was not cached.
}
\description{
Removes a simulation from simulations cache.
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim1 <- loadSimulation(simPath)
sim2 <- loadSimulation(simPath, loadFromCache = FALSE, addToCache = FALSE)

removeSimulationFromCache(sim1) #returns TRUE
removeSimulationFromCache(sim2) #returns FALSE
}
