% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Runs a simulation and returns a list of \code{IndividualResults}}
\usage{
runSimulation(simulation)
}
\arguments{
\item{simulation}{Instance of a simulation to simulate.}
}
\value{
List of individual results (one entry per Individual)
}
\description{
Runs a simulation and returns a list of \code{IndividualResults}
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
results <- runSimulation(sim)
}
