% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runSimulationBatches}
\alias{runSimulationBatches}
\title{Run simulation batches}
\usage{
runSimulationBatches(
  simulationBatches,
  simulationRunOptions = NULL,
  silentMode = FALSE
)
}
\arguments{
\item{simulationBatches}{List of \code{SimulationBatch} objects with added parameter and initial values}

\item{simulationRunOptions}{Optional instance of a \code{SimulationRunOptions} used during the simulation run.}

\item{silentMode}{If \code{TRUE}, no warnings are displayed if a simulation fails.
Default is \code{FALSE}.}
}
\value{
Nested list of \code{SimulationResults} objects. The first level of the list are the IDs of the simulations of SimulationBatches, containing a list of \code{SimulationResults} for each set of parameter/initial values. If a simulation with a parameter/initial values set fails, the result for this run is \code{NULL}
}
\description{
Run simulation batches
}
\details{
Runs a set of simulation batches. The simulation batches must be populated
with sets of parameter and start values with \code{SimulationBatch$addRunValues()}
prior to running. After the run, the list of parameter and start values is cleared.
}
