% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runSimulationsConcurrently}
\alias{runSimulationsConcurrently}
\title{Runs a set of simulations.}
\usage{
runSimulationsConcurrently(
  simulations,
  simulationRunOptions = NULL,
  silentMode = FALSE
)
}
\arguments{
\item{simulations}{A list of \code{Simulation} objects to simulate.}

\item{simulationRunOptions}{Optional instance of a \code{SimulationRunOptions} used during the simulation run.}

\item{silentMode}{If \code{TRUE}, no warnings are displayed if a simulation fails.
Default is \code{FALSE}.}
}
\value{
A list of \code{SimulationResults} objects with names being the IDs of the simulations. If a simulation fails, the result for this simulation is \code{NULL}
}
\description{
Runs a set of simulations.
}
\details{
Runs a set of simulations (only individual simulations) and returns
a named list of \code{SimulationResults}. The names of the entries are the IDs of the
corresponding simulation (i.e. \code{simulation$id}).
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
sim2 <- loadSimulation(simPath)
sim3 <- loadSimulation(simPath)
results <- runSimulationsConcurrently(list(sim, sim2, sim3))
}
