% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-set.R
\name{saveDataSetToPKML}
\alias{saveDataSetToPKML}
\title{Save the \code{DataSet} to pkml}
\usage{
saveDataSetToPKML(dataSet, filePath)
}
\arguments{
\item{dataSet}{The \code{DataSet} object}

\item{filePath}{Path where the pkml file will be created}
}
\description{
Save the \code{DataSet} to pkml
}
\details{
Save the \code{DataSet} to a pkml file that can be loaded by MoBi
}
\examples{
\dontrun{
dataSet <- DataSet$new(name = "NewDataSet")
dataSet$setValues(xValues = c(1, 2, 3, 4, 5), yValues = c(10, 20, 30, 40, 50))
dataSet$saveToPKML(filePath = "../ObsData.pkml")
}
}
