% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-expression-profile.R
\name{saveExpressionProfileToPKML}
\alias{saveExpressionProfileToPKML}
\title{Save an Expression Profile Building Block to pkml}
\usage{
saveExpressionProfileToPKML(expressionProfileBuildingBlock, filePath)
}
\arguments{
\item{expressionProfileBuildingBlock}{A \code{BuildingBlock} object of type
\verb{Expression Profile}, as returned by \code{createExpressionProfileBuildingBlock()}.}

\item{filePath}{Path where the pkml file will be created. Must end with the
\code{.pkml} extension.}
}
\value{
\code{filePath}, invisibly.
}
\description{
Exports an \verb{Expression Profile} building block to a pkml file that can be
loaded by MoBi.
}
\examples{
\dontrun{
expressionProfile <- createExpressionProfileBuildingBlock(
  type = ExpressionProfileCategories$`Metabolizing Enzyme`,
  moleculeName = "CYP3A4",
  speciesName = "Human"
)
saveExpressionProfileToPKML(expressionProfile, "CYP3A4-Human.pkml")
}
}
