% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-individual.R
\name{saveIndividualToPKML}
\alias{saveIndividualToPKML}
\title{Save an Individual Building Block to pkml}
\usage{
saveIndividualToPKML(individualBuildingBlock, filePath)
}
\arguments{
\item{individualBuildingBlock}{A \code{BuildingBlock} object of type \code{Individual},
as returned by \code{createIndividualBuildingBlock()}.}

\item{filePath}{Path where the pkml file will be created. Must end with the
\code{.pkml} extension.}
}
\value{
\code{filePath}, invisibly.
}
\description{
Exports an \code{Individual} building block to a pkml file that can be loaded by
MoBi.
}
\examples{
\dontrun{
individual <- createIndividualBuildingBlock(
  species = Species$Human,
  population = HumanPopulation$European_ICRP_2002
)
saveIndividualToPKML(individual, "Individual.pkml")
}
}
