% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{saveInitialConditionsToPKML}
\alias{saveInitialConditionsToPKML}
\title{Save an Initial Conditions Building Block to pkml}
\usage{
saveInitialConditionsToPKML(initialConditionsBuildingBlock, filePath)
}
\arguments{
\item{initialConditionsBuildingBlock}{A \code{BuildingBlock} object of type
\verb{Initial Conditions}.}

\item{filePath}{Path where the pkml file will be created. Must end with the
\code{.pkml} extension.}
}
\value{
\code{filePath}, invisibly.
}
\description{
Exports an \verb{Initial Conditions} building block to a pkml file that can be
loaded by MoBi.
}
\examples{
\dontrun{
simulation <- loadSimulation("path/to/simulation.pkml")
icBB <- simulation$configuration$modules[[1]]$getInitialConditionsBBs()[[1]]
saveInitialConditionsToPKML(icBB, "InitialConditions.pkml")
}
}
