% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{saveParameterValuesToPKML}
\alias{saveParameterValuesToPKML}
\title{Save a Parameter Values Building Block to pkml}
\usage{
saveParameterValuesToPKML(parameterValuesBuildingBlock, filePath)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type
\verb{Parameter Values}.}

\item{filePath}{Path where the pkml file will be created. Must end with the
\code{.pkml} extension.}
}
\value{
\code{filePath}, invisibly.
}
\description{
Exports a \verb{Parameter Values} building block to a pkml file that can be
loaded by MoBi.
}
\examples{
\dontrun{
simulation <- loadSimulation("path/to/simulation.pkml")
pvBB <- simulation$configuration$modules[[1]]$getParameterValuesBBs()[[1]]
saveParameterValuesToPKML(pvBB, "ParameterValues.pkml")
}
}
