% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{saveSensitivityAnalysisResultsToCSV}
\alias{saveSensitivityAnalysisResultsToCSV}
\title{Saves the simulation analysis results to csv file}
\usage{
saveSensitivityAnalysisResultsToCSV(results, filePath)
}
\arguments{
\item{results}{Results to export (typically calculated using \code{runSensitivityAnalysis} or imported from file)}

\item{filePath}{Full path where the results will be saved.}
}
\description{
Saves the simulation analysis results to csv file
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

# Create a new sensitivity object for the simulation
sensitivity <- SensitivityAnalysis$new(sim)

# Runs the sensitivity analysis
results <- runSensitivityAnalysis(sensitivity)

# Export the results to csv file
exportSensitivityAnalysisResultsToCSV(results, tempfile())
}
