% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{saveSimulation}
\alias{saveSimulation}
\title{Saves a simulation to pkml file}
\usage{
saveSimulation(simulation, pkmlSimulationFile)
}
\arguments{
\item{simulation}{Instance of a simulation to save.}

\item{pkmlSimulationFile}{Full path of where the simulation will be saved.}
}
\description{
Saves a simulation to pkml file
}
