% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{scaleParameterValues}
\alias{scaleParameterValues}
\title{Scale current values of parameters using a factor}
\usage{
scaleParameterValues(parameters, factor)
}
\arguments{
\item{parameters}{A single or a list of \code{Parameter}}

\item{factor}{A numeric value that will be used to scale all parameters}
}
\description{
Scale current values of parameters using a factor
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getParameter("Organism|Liver|Volume", sim)
scaleParameterValues(param, 1)
params <- getAllParametersMatching("Organism|**|Volume", sim)
scaleParameterValues(params, 1.5)
}
\seealso{
\code{\link{getParameter}} and \code{\link{getAllParametersMatching}} to create objects of type Parameter
}
