% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{scaleQuantityValues}
\alias{scaleQuantityValues}
\title{Scale current values of quantities using a factor}
\usage{
scaleQuantityValues(quantities, factor)
}
\arguments{
\item{quantities}{A single or a list of \code{Quantity}}

\item{factor}{A numeric value that will be used to scale all quantities}
}
\description{
Scale current values of quantities using a factor
}
