% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-expression-profile.R
\name{setExpressionProfileParameters}
\alias{setExpressionProfileParameters}
\title{Set Parameters of an Expression Profile Building Block}
\usage{
setExpressionProfileParameters(
  expressionProfileBuildingBlock,
  quantityPaths,
  quantityValues,
  units
)
}
\arguments{
\item{expressionProfileBuildingBlock}{A \code{BuildingBlock} object as returned
by \code{createExpressionProfileBuildingBlock()}.}

\item{quantityPaths}{A character vector of quantity paths to set.}

\item{quantityValues}{A numeric vector of values to assign. Must have the
same length as \code{quantityPaths}.}

\item{units}{A character vector of units corresponding to the quantity values. Must have the same length as \code{quantityPaths}. Units must be compatible with the dimensions of the corresponding quantities in the building block. If a single string is provided, it will be used for all quantities.}
}
\value{
\code{expressionProfileBuildingBlock}, invisibly.
}
\description{
Sets one or more parameter values in an expression profile building block
by providing the corresponding quantity paths and values. The number of
paths and values must be equal.
}
\details{
Note: Setting initial conditions of an expression profile is currently not supported.
}
\examples{
expressionProfile <- createExpressionProfileBuildingBlock(
  type = ExpressionProfileCategories$`Metabolizing Enzyme`,
  moleculeName = "CYP3A4",
  speciesName = "Human"
)
setExpressionProfileParameters(
  expressionProfile,
  quantityPaths = c("Organism|Kidney|Intracellular|CYP3A4|Relative expression"),
  quantityValues = c(1.0),
  units = ""
)
}
