% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{setInitialConditionsBB}
\alias{setInitialConditionsBB}
\title{Set or add initial conditions to an existing Initial Conditions building block.}
\usage{
setInitialConditionsBB(
  initialConditionsBuildingBlock,
  quantityPaths,
  quantityValues,
  scaleDivisors = 1,
  isPresent = TRUE,
  negativeValuesAllowed = FALSE
)
}
\arguments{
\item{initialConditionsBuildingBlock}{A \code{BuildingBlock} object of type \verb{Initial Conditions}.
The entries will be added to or set in this building block.}

\item{quantityPaths}{A list of full paths of the quantities (usually molecules). Should contain
all path elements and the molecule name, separated by \code{|}.}

\item{quantityValues}{A list of values for the quantities.
The length of this list should be equal to the length of \code{quantityPaths}. The values must be in base unit (µmol).}

\item{scaleDivisors}{Either a single value or a list of scale divisors for the quantities. By default, the value is set to 1 for all quantities.
If only single value is provided, the value will be set for all quantities.
If a list is provided, the length of this list should be equal to the length of \code{quantityPaths}.}

\item{isPresent}{Either a single value (\code{TRUE} or \code{FALSE}) or a list of boolean
values indicating whether the quantity is present or not. If a list is provided,
the length of this list should be equal to the length of \code{quantityPaths}.}

\item{negativeValuesAllowed}{A single boolean value or a list of boolean values
indicating whether negative values are allowed for the quantities. If a list is provided,
the length of this list should be equal to the length of \code{quantityPaths}.}
}
\value{
Invisibly returns the updated \code{initialConditionsBuildingBlock} object.
}
\description{
This functions allows adding or modifying initial condition entries. Only constant values are allowed. For setting or adding initial conditions defined by formulas, use the \code{setInitialConditionsFormulas} function.
}
\examples{
\dontrun{
module <- loadModuleFromPKML("path/to/module.pkml")
icBB <- module$getInitialConditionsBBs()[[1]]
setInitialConditionsBB(
  icBB,
  quantityPaths = c("Organ|Tissue|Molecule1", "Organ|Tissue|Molecule2"),
  quantityValues = c(1, 0),
  scaleDivisors = c(1, 10),
  isPresent = c(TRUE, FALSE),
  negativeValuesAllowed = c(TRUE, FALSE)
)
}
}
