% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{setMoleculeScaleDivisors}
\alias{setMoleculeScaleDivisors}
\title{Set molecule scale divisors}
\usage{
setMoleculeScaleDivisors(molecules, values)
}
\arguments{
\item{molecules}{A single or a list of \code{Molecule}}

\item{values}{A numeric value that should be assigned to the molecule scale factor or a vector
of numeric values, if the scale factor of more than one molecule should be changed. Must have the same
length as \code{molecules}}
}
\description{
Set molecule scale divisors
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
molecule <- getMolecule("Organism|Liver|A", sim)
setMoleculeScaleDivisors(molecule, 0.001)
molecules <- getAllMoleculesMatching("Organism|**|A", sim)
setMoleculeScaleDivisors(molecules, c(0.002, 0.003))
}
\seealso{
\code{\link[=getMolecule]{getMolecule()}} and \code{\link[=getAllMoleculesMatching]{getAllMoleculesMatching()}} to retrieve objects of type Molecule
}
