% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-schema.R
\name{setOutputInterval}
\alias{setOutputInterval}
\title{Clears the output interval from the simulation and adds a new one.}
\usage{
setOutputInterval(simulation, startTime, endTime, resolution,
  intervalName = NULL)
}
\arguments{
\item{simulation}{Simulation for which a new interval should be created}

\item{startTime}{Start time of the interval in min}

\item{endTime}{End time of the interval in min}

\item{resolution}{resolution in points/min}

\item{intervalName}{Optional Name of interval. If not specified, a unique name wil be assigned.}
}
\value{
Returns the interval created.
}
\description{
Clears the output interval from the simulation and adds a new one.
}
\note{
This is essentialy a shortcut for \code{clearOutputIntervals} followed by \code{addOutputInterval}
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath, addToCache = FALSE, loadFromCache = FALSE)

# Adds a new interval starting at 1h and ending at 10h with a resolution of 10 points per hour
setOutputInterval(sim, 1 * 60, 10 * 60, 1 / 6)
}
