% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-output-selections.R
\name{setOutputs}
\alias{setOutputs}
\title{Set outputs}
\usage{
setOutputs(quantitiesOrPaths, simulation)
}
\arguments{
\item{quantitiesOrPaths}{Quantity instances (element or vector) (typically
retrieved using \code{getAllQuantitiesMatching}) or quantity path (element or
vector) to add.}

\item{simulation}{Instance of a simulation for which output selection should
be updated.}
}
\description{
Sets the quantities as output into the  \code{simulation}.
The quantities can either be specified using explicit instances or using paths.
This function clears the output selection before adding the new quantities.
See \code{addOutputs} for adding quantities without clearing the output selection.
See \code{clearOutputs} for clearing the output selection without adding new quantities.
@export
}
