% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{setParameterValuesByPath}
\alias{setParameterValuesByPath}
\title{Set the values of parameters in the simulation by path}
\usage{
setParameterValuesByPath(
  parameterPaths,
  values,
  simulation,
  stopIfNotFound = TRUE
)
}
\arguments{
\item{parameterPaths}{A single or a list of parameter path}

\item{values}{A numeric value that should be assigned to the parameters or a vector
of numeric values, if the value of more than one parameter should be changed. Must have the same
length as 'parameterPaths'}

\item{simulation}{Simulation uses to retrieve parameter instances from given paths.}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no parameter exists for the given path,
an error is thrown. If \code{FALSE}, a warning is shown to the user}
}
\description{
Set the values of parameters in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
setParameterValuesByPath("Organism|Liver|Volume", 1, sim)

setParameterValuesByPath(c("Organism|Liver|Volume", "Organism|Volume"), c(2, 3), sim)
}
