% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{setParameterValuesFormulas}
\alias{setParameterValuesFormulas}
\title{Set or add parameter values to an existing Parameter Values building block with values defined by formulas.}
\usage{
setParameterValuesFormulas(
  parameterValuesBuildingBlock,
  quantityPaths,
  formulas
)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.
The entries will be added to or set in this building block.}

\item{quantityPaths}{A list of full paths of the quantities (usually parameters). Should contain
all path elements and the parameter name, separated by \code{|}.}

\item{formulas}{A list of \code{Formula} objects that will be set for the quantities. The length of this list should be equal to the length of \code{quantityPaths}.
The dimension of the parameter will be set to the dimension of the formula.}

\item{quantityValues}{}
}
\description{
Set or add parameter values to an existing Parameter Values building block with values defined by formulas.
}
