% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{setParameterValuesInBB}
\alias{setParameterValuesInBB}
\title{Set or add parameter values to an existing Parameter Values building block.}
\usage{
setParameterValuesInBB(
  parameterValuesBuildingBlock,
  quantityPaths,
  quantityValues,
  units,
  dimensions = NULL
)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.
The entries will be added to or set in this building block.}

\item{quantityPaths}{A list of full paths of the quantities (usually parameters). Should contain
all path elements and the parameter name, separated by \code{|}.}

\item{quantityValues}{A list of values for the quantities in the provided \code{units}.
The length of this list should be equal to the length of \code{quantityPaths}.}

\item{units}{A single unit string or a list of unit strings for the quantity values.
If a single string is provided, it will be used for all quantities.}

\item{dimensions}{A single dimension or a list of dimensions (string names)
of parameter values. Supported dimensions are listed in \code{ospDimensions}. If \code{NULL}
(default), dimensions are derived from the provided \code{units} using \code{getDimensionForUnit()}.}
}
\value{
Invisibly returns the updated \code{parameterValuesBuildingBlock} object.
}
\description{
This function allows adding or modifying parameter value entries. Only constant
values are allowed. Values are expected in the provided \code{units} and will be
converted to base units internally.
}
\examples{
sim <- loadSimulation(system.file(
  "extdata",
  "simple.pkml",
  package = "ospsuite"
))
module <- sim$configuration$modules[[1]]
pvBB <- module$getParameterValuesBBs()[[1]]
setParameterValuesInBB(
  pvBB,
  quantityPaths = "Organism|Organ|Parameter",
  quantityValues = 1.33,
  units = "ml/min/kg"
)
}
