% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{setQuantityValues}
\alias{setQuantityValues}
\title{Set values of quantity}
\usage{
setQuantityValues(quantities, values)
}
\arguments{
\item{quantities}{A single or a list of \code{Quantity}}

\item{values}{A numeric value that should be assigned to the quantity or a vector
of numeric values, if the value of more than one quantity should be changed. Must have the same
length as 'quantities'}
}
\description{
Set values of quantity
}
