% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{setSimulationParameterValues}
\alias{setSimulationParameterValues}
\title{Set the values of parameters in the simulation by path}
\usage{
setSimulationParameterValues(parameterPaths, values, simulation)
}
\arguments{
\item{parameterPaths}{A single or a list of parameter path}

\item{values}{A numeric value that should be assigned to the parameters or a vector
of numeric values, if the value of more than one parameter should be changed. Must have the same
length as 'parameterPaths'}
}
\description{
Set the values of parameters in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
setSimulationParameterValues("Organism|Liver|Volume", 1, sim)

setSimulationParameterValues(c("Organism|Liver|Volume", "Organism|Kidney|Volume") , c(2, 3))
}
