% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-results.R
\name{simulationResultsToDataFrame}
\alias{simulationResultsToDataFrame}
\alias{simulationResultsToTibble}
\title{Converts a \code{SimulationResults} objects to a data.frame}
\usage{
simulationResultsToDataFrame(
  simulationResults,
  quantitiesOrPaths = NULL,
  population = NULL,
  individualIds = NULL
)

simulationResultsToTibble(
  simulationResults,
  quantitiesOrPaths = NULL,
  population = NULL,
  individualIds = NULL
)
}
\arguments{
\item{simulationResults}{Object of type \code{SimulationResults} or a list of
\code{SimulationResults} objects produced by calling \code{runSimulations} on a
\code{Simulation} object. When a list is provided, the function processes each
result and returns a named list of outputs.}

\item{quantitiesOrPaths}{Quantity instances (element or vector) typically
retrieved using \code{getAllQuantitiesMatching} or quantity path (element or
vector of strings) for which the results are to be returned. (optional)
When providing the paths, only absolute full paths are supported (i.e., no
matching with '*' possible). If quantitiesOrPaths is \code{NULL} (default
value), returns the results for all output defined in the results.}

\item{population}{population used to calculate the \code{simulationResults}
(optional). This is used only to add the population covariates to the
resulting data table.}

\item{individualIds}{\code{numeric} IDs of individuals for which the results
should be extracted. By default, all individuals from the results are
considered. If the individual with the provided ID is not found, the ID is
ignored.}
}
\value{
If a single \code{SimulationResults} object is provided, returns a data.frame
with columns IndividualId, Time, paths, simulationValues, unit, dimension,
TimeUnit, TimeDimension, molWeight.

If a list of \code{SimulationResults} is provided, returns a named list of
data.frames, one for each \code{SimulationResults} object.
}
\description{
Converts a \code{SimulationResults} objects to a data.frame
}
\examples{
library(ospsuite)

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Running an individual simulation
# results is an instance of `SimulationResults`
results <- runSimulations(sim)[[1]]

# convert to a dataframe
simulationResultsToDataFrame(results)

# Working with a list of SimulationResults
results <- runSimulations(sim)
dfList <- simulationResultsToDataFrame(results)
}
