% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-population.R
\name{splitPopulationFile}
\alias{splitPopulationFile}
\title{Loads a population from the \code{csvPopulationFile} and split the loaded population according to
\code{numberOfCores}.}
\usage{
splitPopulationFile(csvPopulationFile, numberOfCores, outputFolder,
  outputFileName)
}
\arguments{
\item{csvPopulationFile}{Full path of csv population file to split.}

\item{numberOfCores}{Number of cores used for parallelisation computing. The population will be split accross all cores.}

\item{outputFolder}{Folder where all split files will be created}

\item{outputFileName}{File names will be constructed using this parameter concatenated with the core index.}
}
\value{
A tring vector containing the full path of the population files created. Note that there might be less files than cores
}
\description{
Loads a population from the \code{csvPopulationFile} and split the loaded population according to
\code{numberOfCores}.
}
