% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{StandardContainer}
\alias{StandardContainer}
\title{Standard containers typically available in a PBPK simulation}
\usage{
StandardContainer
}
\description{
Standard containers typically available in a PBPK simulation
}
