% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/enums.R
=======
% Please edit documentation in R/pk-sim.R
>>>>>>> c4fbe917 (Fixes #1834 refactor DataCombined plot examples)
\name{StandardOntogeny}
\alias{StandardOntogeny}
\title{List of ontogeny supported in PK-Sim}
\usage{
StandardOntogeny
}
\description{
List of ontogeny supported in PK-Sim
}
