% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toBaseUnit}
\alias{toBaseUnit}
\title{Converts a value given in a specified unit into the base unit of a quantity}
\usage{
toBaseUnit(
  quantityOrDimension,
  values,
  unit,
  molWeight = NULL,
  molWeightUnit = NULL
)
}
\arguments{
\item{quantityOrDimension}{Instance of a quantity from which the dimension will be retrieved or name of dimension}

\item{values}{Value in unit (single or vector)}

\item{unit}{Unit of value}

\item{molWeight}{Optional molecule weight to use when converting, for example,  from molar to mass amount or concentration. If \code{molWeightUnit} is not specified, \code{molWeight} is assumed to be in kg/\u00b5mol}

\item{molWeightUnit}{Unit of the molecular weight value. If \code{NULL} (default), kg/\u00b5mol is assumed.}
}
\description{
Converts a value given in a specified unit into the base unit of a quantity
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in unit (1000 ml) to the base unit (l) => 1
valueInBaseUnit <- toBaseUnit(par, 1000, "ml")

valuesInBaseUnit <- toBaseUnit(par, c(1000, 2000, 3000), "ml")
}
