% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toDisplayUnit}
\alias{toDisplayUnit}
\title{Convert base unit to display unit}
\usage{
toDisplayUnit(quantity, values, unit = NULL)
}
\arguments{
\item{quantity}{Instance of a quantity from which the base unit will be retrieved}

\item{values}{Value (single or vector). If \code{unit} is not specified, values are assumed to be in base unit}

\item{unit}{Optional Name of the unit to convert from. If \code{NULL} (default), the values are assumed to be in base unit.}
}
\description{
Converts a value given in base unit of a quantity into the display unit of a quantity
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in base unit (1L) to display unit
valueInMl <- toDisplayUnit(par, 1)

valuesInDisplayUnit <- toDisplayUnit(par, c(1, 5, 5))

# Converts the value in ml to display unit
valueInDisplayUnit <- toDisplayUnit(par, 1000, "ml")
}
