% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-path.R
\name{toPathArray}
\alias{toPathArray}
\title{Convert a path defined as string to a path array}
\usage{
toPathArray(path)
}
\arguments{
\item{path}{A string representation of a path, with path entries separated
by '|'}
}
\value{
An array containing one element for each path entry.
}
\description{
Convert a path defined as string to a path array
}
\examples{
toPathArray("Organism|Organ|Liver")

}
