% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toUnit}
\alias{toUnit}
\title{Converts values from one unit to another}
\usage{
toUnit(
  quantityOrDimension,
  values,
  targetUnit,
  sourceUnit = NULL,
  molWeight = NULL,
  molWeightUnit = NULL
)
}
\arguments{
\item{quantityOrDimension}{Instance of a quantity from which the dimension will be retrieved or name of dimension}

\item{values}{Values to convert (single or vector). If \code{sourceUnit} is not specified, \code{values} are in the base unit of the dimension}

\item{targetUnit}{Unit to convert to}

\item{sourceUnit}{Optional Name of the unit to convert from. If \code{NULL} (default), the values are assumed to be in base unit.}

\item{molWeight}{Optional molecular weight to use when converting, for example,  from molar to mass amount or concentration. If \code{molWeightUnit} is not specified, \code{molWeight} is assumed to be in kg/\u00b5mol}

\item{molWeightUnit}{Optional Unit of the molecular weight value. If \code{NULL} (default), kg/\u00b5mol is assumed.}
}
\description{
Converts values from one unit to another
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in base unit (1L) to ml => 1000
valueInMl <- toUnit(par, 1, "ml")

valuesInMl <- toUnit(par, c(1, 5, 5), "ml")

# Converts a numerical value in from mmol/l to mg/dl
valuesInMgDl <- toUnit(ospDimensions$`Concentration (molar)`, 5,
  targetUnit = "mmol/l",
  sourceUnit = "mg/dl", molWeight = 180, molWeightUnit = "g/mol"
)
}
