% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConverter.R
\name{unitConverter}
\alias{unitConverter}
\title{Convert to common units}
\usage{
unitConverter(data, xUnit = NULL, yUnit = NULL)
}
\arguments{
\item{data}{A data frame (or a tibble).}

\item{xUnit, yUnit}{Target units for \code{xValues} and \code{yValues}, respectively. If
not specified (\code{NULL}), one of the existing units in the respective columns
(\code{xUnit} and \code{yUnit}) will be selected as the common unit. For available
dimensions and units, see \code{ospsuite::ospDimensions} and
\code{ospsuite::ospUnits}, respectively.}
}
\value{
A data frame with measurement columns transformed to have common units.
}
\description{
When multiple (observed and/or simulated) datasets are present in a data
frame, they are likely to have different units. This function helps to
convert them to a common unit specified by the user.

This is especially helpful while plotting since the quantities from different
datasets to be plotted on the x-and y-axis need to have same units to be
meaningfully compared.
}
\examples{

# small dataframe to illustrate the conversion
(df <- dplyr::tibble(
  xValues = c(15, 30, 60), xUnit = "min", xDimension = "Time",
  yValues = c(0.25, 45, 78), yUnit = c("", "\%", "\%"), yDimension = c("Fraction", "Fraction", "Fraction"),
  molWeight = c(12.5, 10, 5)
))

unitConverter(df)

unitConverter(df, xUnit = ospUnits$Time$h)

unitConverter(df, yUnit = ospUnits$Fraction$`\%`)

unitConverter(df, xUnit = ospUnits$Time$h, yUnit = ospUnits$Fraction$`\%`)

}
\seealso{
toUnit
}
