% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-parameter.R
\name{updatePKParameter}
\alias{updatePKParameter}
\title{Updates some properties of a PK-Parameter (displayName and displayUnit)}
\usage{
updatePKParameter(name, displayName = NULL, displayUnit = NULL)
}
\arguments{
\item{name}{Name of PK-Parameter to update}

\item{displayName}{Optional display name}

\item{displayUnit}{Optional display unit. Note that the unit should be defined in unit of the dimension}
}
\description{
Updates some properties of a PK-Parameter (displayName and displayUnit)
}
\examples{

updatePKParameter("t_max", "MyTmax", "min")
}
