% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{validateDimension}
\alias{validateDimension}
\title{Validate dimension}
\usage{
validateDimension(dimension)
}
\arguments{
\item{dimension}{String name of the dimension.}
}
\description{
Validate dimension
}
\details{
Check if the provided dimension is supported. If not, throw an error
}
