% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{validateEnumValue}
\alias{validateEnumValue}
\title{Check if \code{value} is in the given {enum}. If not, stops with an error.}
\usage{
validateEnumValue(value, enum, nullAllowed = FALSE)
}
\arguments{
\item{value}{\code{value} to search for in the \code{enum}}

\item{enum}{\code{enum} where the \code{value} should be contained}

\item{nullAllowed}{If TRUE, \code{value} can be \code{NULL} and the test always passes.
If \code{FALSE} (default), NULL is not accepted and the test fails.}
}
\description{
Check if \code{value} is in the given {enum}. If not, stops with an error.
}
