% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validateIsNamedList}
\alias{validateIsNamedList}
\title{Validate that an object is a named list}
\usage{
validateIsNamedList(x, varName)
}
\arguments{
\item{x}{}
}
\value{
invisible TRUE if the validation passed, otherwise an error is thrown.
}
\description{
Validate that an object is a named list
}
\examples{
validateIsNamedList(list(a = 1, b = 2), "myVar") # passes
# validateIsNamedList(list(1, 2), "myVar") # throws an error
}
