% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{validateIsOfType}
\alias{validateIsOfType}
\title{Check if the provided object is of certain type. If not, stop with an error.}
\usage{
validateIsOfType(object, type, nullAllowed = FALSE)
}
\arguments{
\item{object}{An object or a list of objects}

\item{type}{String representation of the type that should be checked for}

\item{nullAllowed}{Boolean flag if \code{NULL} is accepted for the \code{object}. If \code{TRUE},
\code{NULL} is always valid, otherwise the error is thrown. Default is \code{FALSE}}
}
\description{
Check if the provided object is of certain type. If not, stop with an error.
}
