% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{validateUnit}
\alias{validateUnit}
\title{Validate unit}
\usage{
validateUnit(unit, dimension)
}
\arguments{
\item{unit}{String name of the unit}

\item{dimension}{String name of the dimension.}
}
\description{
Validate unit
}
\details{
Check if the unit is valid for the dimension. If not, throw an error
}
