% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{addOutputs}
\alias{addOutputs}
\title{Adds the quantities as output into the  \code{simulation}. The quantiites can either be specified using explicit instances or using paths.}
\usage{
addOutputs(quantities = NULL, paths = NULL, simulation)
}
\arguments{
\item{quantities}{Instance of quantities to add (typically retrieved using \code{getAllQuantitiesMatching})}

\item{paths}{List of quantity path to add}

\item{simulation}{Instance of a simulation for which output selection should be updated.}
}
\description{
Adds the quantities as output into the  \code{simulation}. The quantiites can either be specified using explicit instances or using paths.
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
outputs <- c("Organism|VenousBlood|Plasma|Caffeine", "Organism|ArterialBlood|**|Caffeine")
addOutputs(paths=outputs, sim)

}
