% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R
\name{addProteinExpressionToParameterValuesBB}
\alias{addProteinExpressionToParameterValuesBB}
\title{Extend a Parameter Values Building Block (BB) with protein expression parameters
for selected protein molecules in the selected organs.}
\usage{
addProteinExpressionToParameterValuesBB(
  parameterValuesBuildingBlock,
  spatialStructureBB,
  organPaths = NULL,
  moleculesBB,
  moleculeNames = NULL
)
}
\arguments{
\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.}

\item{spatialStructureBB}{A \code{BuildingBlock} object of type \verb{Spatial Structure}.
Entries will be created for the selected molecules in all physical containers of this
spatial structure.}

\item{organPaths}{A vector of paths to the organs for which the expression paramters
will be created. If any of the provided path is not an organ, an error is thrown.
If \code{NULL} (default), the function will use all organs from the spatial structure.}

\item{moleculesBB}{A \code{BuildingBlock} object of type \code{Molecules}. The entries will be
created for all proteins from this building block, or for a subset of protein molecules
defined in the \code{moleculeNames} argument.}

\item{moleculeNames}{Optional list of protein molecule names. If provided, only the molecules
with these names will be added to the \code{parameterValuesBuildingBlock}.}
}
\value{
Path of entries added to the building block.
}
\description{
TBD: is a spatial structure required? Option 1 - define the organs as paths
Option 2 - provide a spatial structure and then define organ paths. Entries will
be created for all sub-organs.
}
\details{
TBD: Is a molecules BB required? Or just molecule names?
}
