% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{allAvailableDimensions}
\alias{allAvailableDimensions}
\title{List all available dimensions in the \code{OSPSuite} platform}
\usage{
allAvailableDimensions()
}
\value{
Returns the names of all available dimensions defined in the \code{OSPSuite}
platform.
}
\description{
List all available dimensions in the \code{OSPSuite} platform
}
\examples{

allAvailableDimensions()

}
