% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\docType{data}
\name{AllMatchingMethod}
\alias{AllMatchingMethod}
\title{Names of the \code{.NET} container tasks of the type \code{"AllXXXMatching"}}
\format{
An object of class \code{list} of length 4.
}
\usage{
AllMatchingMethod
}
\description{
Names of the \code{.NET} container tasks of the type \code{"AllXXXMatching"}
}
\keyword{internal}
