% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-config.R
\docType{class}
\name{ApiConfig}
\alias{ApiConfig}
\title{ApiConfig}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiConfig
}
\description{
Global configuration for the OSPSuite .NET API
}
\keyword{datasets}
