% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-block.R
\docType{data}
\name{BuildingBlockTypes}
\alias{BuildingBlockTypes}
\title{Types of building blocks supported.}
\format{
An object of class \code{list} of length 4.
}
\usage{
BuildingBlockTypes
}
\description{
Types of building blocks supported.
}
\keyword{datasets}
