% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-structure.R
\docType{class}
\name{Cache}
\alias{Cache}
\title{Cache}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cache
}
\description{
Structure for caching objects
}
\section{Fields}{

\describe{
\item{\code{type}}{Type of the object that are cached. All objects must be of the same type.}

\item{\code{keys}}{Keys of the cached objects}
}}

\keyword{datasets}
