% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-analysis.R
\name{calculatePKAnalyses}
\alias{calculatePKAnalyses}
\title{Calculates the pkAnalyses for all output values available in \code{results}.}
\usage{
calculatePKAnalyses(results)
}
\arguments{
\item{results}{Results of simulation. Typically the \code{results} are calculated
using \code{runSimulations} or imported from csv file via \code{importResults}.}
}
\value{
An instance of \code{SimulationPKAnalyses} class.
}
\description{
Calculates the pkAnalyses for all output values available in \code{results}.
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

addOutputs("Organism|VenousBlood|*|Caffeine", sim)
results <- runSimulations(sim)
pkAnalyses <- calculatePKAnalyses(results)
}
