% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-analyses.R
\name{calculatePKAnalyses}
\alias{calculatePKAnalyses}
\title{Calculates the pkAnalyses for all output values available in \code{results}.}
\usage{
calculatePKAnalyses(results, simulation)
}
\arguments{
\item{results}{Results of simulation. Typically the \code{results} are calculated using \code{runSimulation} or imported from csv file via \code{importResults}}

\item{simulation}{Instance of a simulation for which pk-analyses should be calculated}
}
\value{
A list of QuantityPKParameter (one object for each output-pk parameter combination)
}
\description{
Calculates the pkAnalyses for all output values available in \code{results}.
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

addOutputs("Organism|VenousBlood|*|Caffeine", sim)
results <- runSimulation(sim)
pkAnalyses <- calculatePKAnalyses(results, sim)
}
