% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{calculateResiduals}
\alias{calculateResiduals}
\title{Calculate residuals for datasets in \code{DataCombined}}
\usage{
calculateResiduals(
  dataCombined,
  xUnit = NULL,
  yUnit = NULL,
  scaling = tlf::Scaling$lin
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class.}

\item{xUnit, yUnit}{Target units for \code{xValues} and \code{yValues}, respectively. If
not specified (\code{NULL}), first of the existing units in the respective
columns (\code{xUnit} and \code{yUnit}) will be selected as the common unit. For
available dimensions and units, see \code{ospsuite::ospDimensions} and
\code{ospsuite::ospUnits}, respectively.}

\item{scaling}{A character specifying scale: either \code{tlf::Scaling$lin}
(linear) or \code{tlf::Scaling$log} (logarithmic).}
}
\description{
Calculate residuals for datasets in \code{DataCombined}
}
\details{
To compute residuals, for every simulated dataset in a given group, there
should also be a corresponding observed dataset. If this is not the case, the
corresponding observed or simulated datasets will be removed.

When multiple (observed and/or simulated) datasets are present in
\code{DataCombined}, they are likely to have different units. The \code{xUnit} and
\code{yUnit} arguments help you specify a common unit to convert them to.
}
\examples{
# simulated data
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)
simResults <- runSimulation(sim)
outputPath <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"

# observed data
obsData <- lapply(
  c("ObsDataAciclovir_1.pkml", "ObsDataAciclovir_2.pkml", "ObsDataAciclovir_3.pkml"),
  function(x) loadDataSetFromPKML(system.file("extdata", x, package = "ospsuite"))
)
names(obsData) <- lapply(obsData, function(x) x$name)


# Create a new instance of `DataCombined` class
myDataCombined <- DataCombined$new()

# Add simulated results
myDataCombined$addSimulationResults(
  simulationResults = simResults,
  quantitiesOrPaths = outputPath,
  groups = "Aciclovir PVB"
)

# Add observed data set
myDataCombined$addDataSets(obsData$`Vergin 1995.Iv`, groups = "Aciclovir PVB")

calculateResiduals(myDataCombined)
}
