% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{CalculationMethodCategories}
\alias{CalculationMethodCategories}
\title{Categories of molecule calculation methods that can be overridden on a
per-molecule basis in a simulation configuration.}
\format{
An object of class \code{list} of length 2.
}
\usage{
CalculationMethodCategories
}
\description{
Categories of molecule calculation methods that can be overridden on a
per-molecule basis in a simulation configuration.
}
\keyword{datasets}
