% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationConfiguration.R
\docType{data}
\name{CellularPermeabilityMethods}
\alias{CellularPermeabilityMethods}
\title{Available methods for calculation of cellular permeabilities.}
\format{
An object of class \code{list} of length 3.
}
\usage{
CellularPermeabilityMethods
}
\description{
MUST WE DEFINE IT HERE, OR CAN WE GET IT FROM .NET SOMEHOW?
}
\keyword{datasets}
