% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\docType{data}
\name{CompareBy}
\alias{CompareBy}
\title{How should comparison of entities be performed}
\format{
An object of class \code{list} of length 3.
}
\usage{
CompareBy
}
\description{
How should comparison of entities be performed
}
\keyword{datasets}
