% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{class}
\name{Container}
\alias{Container}
\title{Container}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Container
}
\description{
Contains other entities such as Parameter or containers
}
\section{Methods}{

\describe{
  \item{containerType}{Type of container}
  }
}

\keyword{datasets}
