% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{class}
\name{Container}
\alias{Container}
\title{Container}
\usage{
Container
}
\description{
Contains other entities such as Parameter or containers
}
\section{Fields}{

\describe{
\item{\code{containerType}}{Type of container}
}}

\keyword{datasets}
