% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{convertSnapshot}
\alias{convertSnapshot}
\title{Convert between snapshot and project formats}
\usage{
convertSnapshot(..., format, output = ".")
}
\arguments{
\item{...}{character strings, path to files or a directory containing files to convert}

\item{format, }{character string, target format either "snapshot" or "project".}

\item{output}{character string, path to the output directory where to write the converted files}
}
\description{
Convert between snapshot and project formats
}
