% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-snapshots.R
\name{convertSnapshot}
\alias{convertSnapshot}
\title{Convert between snapshot and project formats}
\usage{
convertSnapshot(..., format, output = ".", runSimulations = FALSE)
}
\arguments{
\item{...}{character strings, path to files or a directory containing files to convert}

\item{format, }{character string, target format either "snapshot" or "project".}

\item{output}{character string, path to the output directory where to write the converted files}

\item{runSimulations}{logical, whether to run simulations during conversion (default = FALSE).
Only when converting from snapshot to project.}
}
\description{
Convert between snapshot and project formats
}
\examples{
\dontrun{
convertSnapshot("path/to/snapshot.json", format = "project")
convertSnapshot("path/to/project.pksim5", format = "snapshot")
}
}
