% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-individual.R
\name{createIndividual}
\alias{createIndividual}
\title{Creates an individual using the PKSim Database.}
\usage{
createIndividual(originData, useDistribution = FALSE,
  moleculeOntogenies = NULL)
}
\arguments{
\item{path}{A string representation of a path, with path entries separated}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates an individual using the PKSim Database.
}
\examples{

}
