% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-individual.R
\name{createIndividualBuildingBlock}
\alias{createIndividualBuildingBlock}
\title{Create a MoBi Individual Building Block}
\usage{
createIndividualBuildingBlock(
  species,
  population = NULL,
  gender = NULL,
  weight = NULL,
  weightUnit = "kg",
  height = NULL,
  heightUnit = "cm",
  age = NULL,
  ageUnit = "year(s)",
  gestationalAge = 40,
  gestationalAgeUnit = "week(s)",
  seed = NULL
)
}
\arguments{
\item{species}{Species of the individual as defined in PK-Sim (see Species enum)}

\item{population}{Population to use to create the individual. This is required only when the species is Human. (See HumanPopulation enum)}

\item{gender}{Gender to use to create the individual. (See Gender enum)}

\item{weight}{Weight of the created individual}

\item{weightUnit}{Unit in which the weight value is defined. Default is kg}

\item{height}{Height of the created individual (for human species only)}

\item{heightUnit}{Unit in which the height value is defined. Default is cm}

\item{age}{Age of the created individual (for human species only)}

\item{ageUnit}{Unit in which the age value is defined. Default is year(s)}

\item{gestationalAge}{Gestational age of the created individual (for human species only using the Preterm population). Default is 40 Weeks}

\item{gestationalAgeUnit}{Unit in which the gestational age value is defined. Default is week(s)}

\item{seed}{Optional seed parameter to use to generate start values for the created individual algorithm.}
}
\value{
An object of type \code{BuildingBlock} representing the created individual.
}
\description{
Creates an individual building block in MoBi for a given species and
optional demographic characteristics. Currently, disease states are not supported.
}
\examples{
individual <- createIndividualBuildingBlock(
  species = Species$Human,
  population = HumanPopulation$European_ICRP_2002,
  gender = Gender$Male,
  weight = 73,
  age = 30
)
}
