% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-individual.R
\name{createIndividualBuildingBlock}
\alias{createIndividualBuildingBlock}
\title{Create an Individual Building Block}
\usage{
createIndividualBuildingBlock(
  species,
  population = NULL,
  gender = NULL,
  weight = NULL,
  weightUnit = "kg",
  height = NULL,
  heightUnit = "cm",
  age = NULL,
  ageUnit = "year(s)",
  gestationalAge = 40,
  gestationalAgeUnit = "week(s)",
  moleculeOntogenies = NULL,
  seed = NULL,
  diseaseState = IndividualDiseaseStates$None
)
}
\arguments{
\item{seed}{}
}
\value{
An object of type \code{BuildingBlock} representing an individual
}
\description{
Create an Individual Building Block
}
