% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-individual.R
\name{createIndividualCharacteristics}
\alias{createIndividualCharacteristics}
\title{Creates an individual using the PK-Sim Database.}
\usage{
createIndividualCharacteristics(
  species,
  population = NULL,
  gender = NULL,
  weight = NULL,
  weightUnit = "kg",
  height = NULL,
  heightUnit = "cm",
  age = NULL,
  ageUnit = "year(s)",
  gestationalAge = 40,
  gestationalAgeUnit = "week(s)",
  moleculeOntogenies = NULL,
  seed = NULL
)
}
\arguments{
\item{species}{Species of the individual as defined in PK-Sim (see Species enum)}

\item{population}{Population to use to create the individual. This is required only when the species is Human. (See HumanPopulation enum)}

\item{gender}{Gender to use to create the individual. (See Gender enum)}

\item{weight}{Weight of the created individual}

\item{weightUnit}{Unit in which the weight value is defined. Default is kg}

\item{height}{Height of the created individual (for human species only)}

\item{heightUnit}{Unit in which the height value is defined. Default is cm}

\item{age}{Age of the created individual (for human species only)}

\item{ageUnit}{Unit in which the age value is defined. Default is year(s)}

\item{gestationalAge}{Gestational age of the created individual (for human species only using the Preterm population). Default is 40 Weeks}

\item{gestationalAgeUnit}{Unit in which the gestational age value is defined. Default is week(s)}

\item{moleculeOntogenies}{Optional list of \code{MoleculeOntogeny} that will be used to retrieve ontogeny information for molecules.
A \code{MoleculeOntogeny} is an object with a \code{molecule} property (e.g. the name of the molecule as defined in your simulation)
and an \code{ontogeny} property (e.g. the name of the predefined ontogeny to use for this molecule). The list of all available ontogenies
can be accessed programmatically using the enum \code{StandardOntogeny}}

\item{seed}{Optional seed parameter to use to generate start values for the created individual algorithm.}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates an individual using the PK-Sim Database.
}
