% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-individual.R
\name{createIndividualCharacteristics}
\alias{createIndividualCharacteristics}
\title{Creates an individual using the PKSim Database.}
\usage{
createIndividualCharacteristics(species, population = NULL,
  gender = NULL, weight, weightUnit = "kg", height = NULL,
  heightUnit = "cm", age = NULL, ageUnit = "year(s)",
  gestationalAge = 40, gestationalAgeUnit = "week(s)",
  moleculeOntogenies = NULL)
}
\arguments{
\item{individualCharacteristics}{Characteristics of the individual to create as an instance of \code{OriginData}}

\item{useDistribution}{Boolean. Indicates wheather the function should returns all parameters defining an individual on only those parameters
that are actually}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates an individual using the PKSim Database.
}
\examples{

}
