% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-individual.R
\name{createIndividualCharacteristics}
\alias{createIndividualCharacteristics}
\title{Creates an individual using the PKSim Database.}
\usage{
createIndividualCharacteristics(species, population = NULL,
  gender = NULL, weight, weightUnit = "kg", height = NULL,
  heightUnit = "cm", age = NULL, ageUnit = "year(s)",
  gestationalAge = 40, gestationalAgeUnit = "week(s)",
  moleculeOntogenies = NULL)
}
\arguments{
\item{species}{Species of the individual as defined in PK-Sim (see Species enum)}

\item{population}{Population to use to create the individual. This is required only when the species is Human. (See HumanPopulation enum)}

\item{gender}{Gender to use to create the individual. (See Gender enum)}

\item{weight}{Weight of the created individual}

\item{weightUnit}{Unit in which the weight value is defined. Default is kg}

\item{height}{Height of the created individual (for human species only)}

\item{heightUnit}{Unit in which the height value is defined. Default is cm}

\item{age}{Age of the created individual (for human species only)}

\item{ageUnit}{Unit in which the age value is defined. Default is year(s)}

\item{gestationalAge}{Gestational age of the created individual (for human species only using the Preterm population). Default is 40 Weeks}

\item{gestationalAgeUnit}{Unit in which the gestational age value is defined. Default is week(s)}

\item{moleculeOntogenies}{Optional list of \code{MoleculeOntogeny} that will be used to retrieve ontogeny information for molecules.}
}
\value{
An array of \code{ParameterValue} containing the value of each individual parameter
}
\description{
Creates an individual using the PKSim Database.
}
