% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-expression-profile.R
\name{createMoBiExpressionProfileBuildingBlock}
\alias{createMoBiExpressionProfileBuildingBlock}
\title{Create a MoBi Expression Profile Building Block}
\usage{
createMoBiExpressionProfileBuildingBlock(category, moleculeName, speciesName)
}
\arguments{
\item{category}{Category of the expression profile.}

\item{moleculeName}{Name of the molecule.}

\item{speciesName}{Name of the species.}
}
\value{
A \code{BuildingBlock} object representing the created expression profile.
}
\description{
Creates an expression profile building block in MoBi for a given molecule,
species, and category by calling the MoBi \code{ExpressionProfileTask}.
}
\details{
Create a MoBi Expression Profile Building Block
}
\examples{
\dontrun{
expressionProfile <- createMoBiExpressionProfileBuildingBlock(
  category = "MyCategory",
  moleculeName = "CYP3A4",
  speciesName = "Human"
)
}
}
