% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plot-configuration.R
\name{createPlotConfiguration}
\alias{createPlotConfiguration}
\title{Create an instance of \code{ospPlotConfiguration} class}
\usage{
createPlotConfiguration(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  legendPosition = "outsideRight",
  legendTitle = NULL,
  legendTitleSize = NULL,
  legendTitleColor = NULL,
  legendTitleFontFamily = NULL,
  legendTitleFontFace = NULL,
  legendTitleAngle = NULL,
  legendCaptionSize = NULL,
  legendCaptionColor = NULL,
  legendCaptionFontFamily = NULL,
  legendCaptionFontFace = NULL,
  legendCaptionAngle = NULL,
  xAxisLimits = NULL,
  xAxisScale = "lin",
  xAxisTicks = NULL,
  xAxisTickLabels = NULL,
  xAxisLabelSize = NULL,
  xAxisLabelColor = NULL,
  xAxisLabelFontFamily = NULL,
  xAxisLabelFontFace = NULL,
  xAxisLabelAngle = NULL,
  yAxisLimits = NULL,
  yAxisScale = "lin",
  yAxisTicks = NULL,
  yAxisTickLabels = NULL,
  yAxisLabelSize = NULL,
  yAxisLabelColor = NULL,
  yAxisLabelFontFamily = NULL,
  yAxisLabelFontFace = NULL,
  yAxisLabelAngle = NULL,
  watermark = NULL,
  watermarkSize = NULL,
  watermarkColor = NULL,
  watermarkFontFamily = NULL,
  watermarkFontFace = NULL,
  watermarkAngle = NULL,
  plotBackgroundFill = NULL,
  plotBackgroundColor = NULL,
  plotBackgroundSize = NULL,
  plotBackgroundLinetype = NULL,
  plotPanelBackgroundFill = NULL,
  plotPanelBackgroundColor = NULL,
  plotPanelBackgroundSize = NULL,
  plotPanelBackgroundLinetype = NULL,
  xAxisColor = NULL,
  xAxisSize = NULL,
  xAxisLinetype = NULL,
  yAxisColor = NULL,
  yAxisSize = NULL,
  yAxisLinetype = NULL,
  xGridColor = NULL,
  xGridSize = NULL,
  xGridLinetype = NULL,
  yGridColor = NULL,
  yGridSize = NULL,
  yGridLinetype = NULL,
  linesColor = NULL,
  linesFill = NULL,
  linesShape = NULL,
  linesSize = NULL,
  linesLinetype = NULL,
  linesAlpha = NULL,
  pointsColor = NULL,
  pointsFill = NULL,
  pointsShape = NULL,
  pointsSize = NULL,
  pointsLinetype = NULL,
  pointsAlpha = NULL,
  ribbonsColor = NULL,
  ribbonsFill = NULL,
  ribbonsShape = NULL,
  ribbonsSize = NULL,
  ribbonsLinetype = NULL,
  ribbonsAlpha = NULL,
  errorbarsColor = NULL,
  errorbarsFill = NULL,
  errorbarsShape = NULL,
  errorbarsSize = NULL,
  errorbarsLinetype = NULL,
  errorbarsAlpha = NULL,
  plotSaveFileFormat = "png",
  plotSaveFileWidth = 16,
  plotSaveFileHeight = 9,
  plotSaveFileDimensionUnits = "cm"
)
}
\arguments{
\item{title, subtitle, xlabel, ylabel, legendTitle, watermark}{A character string
providing plot annotations for plot title, subtitle, x-axis label, y-axis
label, plot legend, watermark, respectively.}

\item{legendPosition}{A character string defining the legend position
(default: \code{"outsideRight"}). Available options
are:\code{"none"},\code{"insideTop"},\code{"insideTopLeft"},
\code{"insideLeft"},\code{"insideBottomLeft"},\code{"insideBottom"},\code{"insideBottomRight"},
\code{"insideRight"},\code{"insideTopRight"},\code{"outsideTop"},\code{"outsideTopLeft"},
\code{"outsideLeft"},\code{"outsideBottomLeft"},\code{"outsideBottom"},\code{"outsideBottomRight"},
\code{"outsideRight"},\code{"outsideTopRight"}.}

\item{legendTitleSize, legendTitleColor, legendTitleFontFamily, legendTitleFontFace, legendTitleAngle, legendCaptionSize}{Aesthetic properties for the legend title.}

\item{legendCaptionColor, legendCaptionFontFamily, legendCaptionFontFace, legendCaptionAngle}{Aesthetic properties for the legend caption.}

\item{xAxisLimits, yAxisLimits}{A numeric vector of axis limits for the x-and
y-axis, respectively.}

\item{xAxisScale, yAxisScale}{A character string defining axis scale (default:
\code{"lin"}). Available options are: \code{"lin"}, \code{"log"}, \code{"ln"}, \code{"discrete"},
\code{"reverse"}, \code{"sqrt"}, \code{"time"}, \code{"date"}.}

\item{xAxisTicks, yAxisTicks}{A numeric vector or a function defining where to
position x-and y-axis ticks, respectively.}

\item{xAxisTickLabels, xAxisLabelSize, xAxisLabelColor, xAxisLabelFontFamily, xAxisLabelFontFace, xAxisLabelAngle}{Aesthetic properties for the x-axis label.}

\item{yAxisTickLabels, yAxisLabelSize, yAxisLabelColor, yAxisLabelFontFamily, yAxisLabelFontFace, yAxisLabelAngle}{Aesthetic properties for the y-axis label.}

\item{watermarkSize, watermarkColor, watermarkFontFamily, watermarkFontFace, watermarkAngle}{A character string specifying the aesthetic properties for the watermark.}

\item{plotBackgroundFill, plotBackgroundColor, plotBackgroundSize, plotBackgroundLinetype}{A character string specifying the aesthetic properties for the plot background.}

\item{plotPanelBackgroundFill, plotPanelBackgroundColor, plotPanelBackgroundSize, plotPanelBackgroundLinetype}{A character string specifying the aesthetic properties for the plot panel (inside of plot) background.}

\item{xAxisColor, xAxisSize, xAxisLinetype}{A character string specifying the aesthetic properties for the x-axis.}

\item{yAxisColor, yAxisSize, yAxisLinetype}{A character string specifying the aesthetic properties for the y-axis.}

\item{xGridColor, xGridSize, xGridLinetype}{A character string specifying the aesthetic properties for the x-axis grid.}

\item{yGridColor, yGridSize, yGridLinetype}{A character string specifying the aesthetic properties for the y-axis grid.}

\item{linesColor, linesFill, linesShape, linesSize, linesLinetype, linesAlpha}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for lines.}

\item{pointsColor, pointsFill, pointsShape, pointsSize, pointsLinetype, pointsAlpha}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for points.}

\item{ribbonsColor, ribbonsFill, ribbonsShape, ribbonsSize, ribbonsLinetype, ribbonsAlpha}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for ribbons.}

\item{errorbarsColor, errorbarsFill, errorbarsShape, errorbarsSize, errorbarsLinetype, errorbarsAlpha}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for errorbars.}

\item{plotSaveFileFormat, plotSaveFileWidth, plotSaveFileHeight, plotSaveFileDimensionUnits}{File format to which the plot needs to be saved, and the units and dimensions for saving the plot.}
}
\description{
Create an instance of \code{ospPlotConfiguration} class
}
