% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-population.R
\name{createPopulation}
\alias{createPopulation}
\title{Creates an population using the PKSim Database}
\usage{
createPopulation(populationCharacteristics)
}
\arguments{
\item{populationCharacteristics}{Characteristics of the population to create as an instance of \code{OriginData}
that are actually distributed parameters}
}
\value{
An list with two entries:
The \code{population} An instance of a population object.
The \code{derivedParameters} containing the parameter values modified indirectly by the algorithm. Those parameters are typically formula parameters.
}
\description{
Creates an population using the PKSim Database
}
