% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-population.R
\name{createPopulation}
\alias{createPopulation}
\title{Creates an population using the PKSim Database}
\usage{
createPopulation(populationCharacteristics)
}
\arguments{
\item{populationCharacteristics}{Characteristics of the population to create as an instance of \code{OriginData}
that are actually distributed parameters}
}
\value{
An instance of a population object
}
\description{
Creates an population using the PKSim Database
}
