% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{createSimulation}
\alias{createSimulation}
\title{Create a new simulation from a simulation configuration}
\usage{
createSimulation(
  simulationName,
  simulationConfiguration,
  createAllProcessRateParameters = FALSE,
  showWarnings = FALSE
)
}
\arguments{
\item{simulationName}{Name of the simulation.}

\item{simulationConfiguration}{An instance of \code{SimulationConfiguration} that defines the simulation.}

\item{createAllProcessRateParameters}{If \code{TRUE}, process rate parameters will be created for all reactions and transport processes.}

\item{showWarnings}{If \code{TRUE}, warnings generated during simulation creation will be shown as R warnings. Default is \code{FALSE}.}
}
\value{
A \code{Simulation} object
}
\description{
Create a new simulation from a simulation configuration
}
