% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-configuration.R
\name{createSimulationConfiguration}
\alias{createSimulationConfiguration}
\title{Create a simulation configuration from modules.}
\usage{
createSimulationConfiguration(
  modules,
  individual = NULL,
  expressionProfiles = NULL,
  selectedInitialConditions = NULL,
  selectedParameterValues = NULL
)
}
\arguments{
\item{modules}{A list of \code{Module} objects from which to create in simulation.
The order of modules defines the order in which the modules will be combined to a simulation!}

\item{individual}{Optional, an individual building block}

\item{expressionProfiles}{Optional, a list of expression profiles to apply to the simulation.}

\item{selectedInitialConditions}{By default, the first Initial Conditions
(IC) building block (BB) of each module will be selected. If a module has multiple
IC BBs, it is possible to specify which IC BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the IC BB.
By setting the value to \code{NULL}, no IC BB from the specified module will be applied.}

\item{selectedParameterValues}{By default, the first Parameter Values
(PV) building block (BB) of each module will be selected. If a module has multiple
PV BBs, it is possible to specify which PV BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the PV BB.
By setting the value to \code{NULL}, no PV BB from the specified module will be applied.}
}
\value{
A \code{SimulationConfiguration} object.
}
\description{
Create a simulation configuration from modules.
}
