% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{DataAggregationMethods}
\alias{DataAggregationMethods}
\title{Names of aggregation available for plotPopulationTimeProfile()}
\usage{
DataAggregationMethods
}
\description{
Names of aggregation available for plotPopulationTimeProfile()
}
