% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{DataAggregationMethods}
\alias{DataAggregationMethods}
\title{Names of aggregation available for plotTimeProfile()}
\usage{
DataAggregationMethods
}
\description{
Names of aggregation available for plotTimeProfile()
}
