% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\docType{data}
\name{DataAggregationMethods}
\alias{DataAggregationMethods}
\title{Names of aggregation available for plotPopulationTimeProfile()}
\format{
An object of class \code{list} of length 3.
}
\usage{
DataAggregationMethods
}
\description{
Names of aggregation available for plotPopulationTimeProfile()
}
\keyword{datasets}
