% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-column.R
\docType{class}
\name{DataColumn}
\alias{DataColumn}
\title{DataColumn}
\description{
One column defined in a \code{DataRepository}
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{DataColumn}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{values}}{Returns the values defined in the column}

\item{\code{name}}{Returns the name of the column  (Read-Only)}

\item{\code{unit}}{The base unit in which the values are defined (Read-Only)}

\item{\code{displayUnit}}{The unit in which the values should be displayed}

\item{\code{dimension}}{The dimension of the values}

\item{\code{molWeight}}{Molecular weight of associated observed data in internal unit
In no molecular weight is defined, the value is \code{NULL}}

\item{\code{LLOQ}}{Lower Limit Of Quantification.
In no LLOQ is defined, the value is \code{NULL}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{DataColumn$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="finalize">}\href{../../ospsuite/html/DotNetWrapper.html#method-finalize}{\code{ospsuite::DotNetWrapper$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize">}\href{../../ospsuite/html/DotNetWrapper.html#method-initialize}{\code{ospsuite::DotNetWrapper$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataColumn$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
