% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-combined.R
\docType{class}
\name{DataCombined}
\alias{DataCombined}
\title{Object combining simulated and observed data}
\description{
A class for storage of simulated and/or observed data, which can be further
used to extract as a dataframe or for visualization methods.
}
\examples{

# load the simulation
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)
simulationResults <- runSimulations(simulations = sim)

# create a new dataset object
dataSet <- DataSet$new()

# created object with datasets combined
myCombDat <- DataCombined$new()
myCombDat$addSimulationResults(simulationResults)
myCombDat$addDataSet(dataSet)

# print the object
myCombDat
}
\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{DataCombined}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataCombined$new()}}
\item \href{#method-addSimulationResults}{\code{DataCombined$addSimulationResults()}}
\item \href{#method-addDataSet}{\code{DataCombined$addDataSet()}}
\item \href{#method-toDataFrame}{\code{DataCombined$toDataFrame()}}
\item \href{#method-print}{\code{DataCombined$print()}}
\item \href{#method-clone}{\code{DataCombined$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of \code{DataCombined} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSimulationResults"></a>}}
\if{latex}{\out{\hypertarget{method-addSimulationResults}{}}}
\subsection{Method \code{addSimulationResults()}}{
Add simulated data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$addSimulationResults(
  simulationResults,
  grouping = NULL,
  paths = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationResults}}{Instance of the \code{SimulationResults} object.}

\item{\code{grouping}}{TODO:}

\item{\code{paths}}{Quantity paths (element or vector of strings) for which the
results are to be returned. When providing the paths, only absolute full
paths are supported (i.e., no matching with '*' possible). If \code{NULL}
(default value), returns the results for all output defined in the results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{DataCombined} object containing simulated data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDataSet"></a>}}
\if{latex}{\out{\hypertarget{method-addDataSet}{}}}
\subsection{Method \code{addDataSet()}}{
Add observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$addDataSet(dataSet, grouping = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSet}}{Instance (or a list of instances) of the \code{DataSet} object(s).}

\item{\code{grouping}}{TODO:}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{DataCombined} object containing observed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
A dataframe of simulated and observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$toDataFrame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
If dataSet provided was a list of DataSet objects, this can be quite long
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCombined$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
