% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-set.R
\docType{data}
\name{DataErrorType}
\alias{DataErrorType}
\title{Supported types of the error}
\format{
An object of class \code{list} of length 2.
}
\usage{
DataErrorType
}
\description{
Supported types of the error
}
\keyword{datasets}
