% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-repository.R
\docType{class}
\name{DataRepository}
\alias{DataRepository}
\title{DataRepository}
\description{
An object typically holding observed data
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{DataRepository}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{The name of the object.}

\item{\code{baseGrid}}{Returns the base column for the data repository (typically time column).}

\item{\code{columns}}{Returns all columns (including baseGrid) defined in the data repository.}

\item{\code{allButBaseGrid}}{Returns all columns excluding baseGrid defined on the data repository.}

\item{\code{metaData}}{Returns a named list of meta data defined for the data repository.
where the name is the name of the metaData and the value is the meta data value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataRepository-addColumn}{\code{DataRepository$addColumn()}}
\item \href{#method-DataRepository-new}{\code{DataRepository$new()}}
\item \href{#method-DataRepository-print}{\code{DataRepository$print()}}
\item \href{#method-DataRepository-addMetaData}{\code{DataRepository$addMetaData()}}
\item \href{#method-DataRepository-removeMetaData}{\code{DataRepository$removeMetaData()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRepository-addColumn"></a>}}
\if{latex}{\out{\hypertarget{method-DataRepository-addColumn}{}}}
\subsection{Method \code{addColumn()}}{
Adds a column to the data repository
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRepository$addColumn(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column}}{Column to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRepository-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataRepository-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRepository$new(ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Optional underlying DataRepository. If it is not provided, a new instance will be created}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DataRepository} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRepository-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataRepository-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRepository$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRepository-addMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-DataRepository-addMetaData}{}}}
\subsection{Method \code{addMetaData()}}{
Adds a new entry to meta data list or changes its value if the name is already present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRepository$addMetaData(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of new meta data list entry}

\item{\code{value}}{Value of new meta data list entry}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRepository-removeMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-DataRepository-removeMetaData}{}}}
\subsection{Method \code{removeMetaData()}}{
Removes the meta data entry in the list if one is defined with this name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRepository$removeMetaData(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of meta data entry to delete}
}
\if{html}{\out{</div>}}
}
}
}
