% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-set.R
\docType{class}
\name{DataSet}
\alias{DataSet}
\title{DataSet}
\description{
A wrapper around DataRepository exposing convenience methods to use and manipulate dataSets
(typically observed data) containing an X column, a Y column and potentially an Error columns
}
\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{DataSet}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{The name of the DataSet}

\item{\code{xDimension}}{Dimension in which the xValues are defined}

\item{\code{xUnit}}{Unit in which the xValues are defined}

\item{\code{xValues}}{Values stored in the xUnit dimension (not necessarily in the base unit of the dimension)}

\item{\code{yDimension}}{Dimension in which the xValues are defined}

\item{\code{yUnit}}{Unit in which the yValues are defined}

\item{\code{yValues}}{Values stored in the yUnit dimension (not necessarily in the base unit of the dimension)}

\item{\code{yErrorType}}{Type of the error - geometric or arithmetic}

\item{\code{yErrorUnit}}{Unit in which the yErrorValues are defined}

\item{\code{yErrorValues}}{Values of error stored in the yErrorUnit unit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSet$new()}}
\item \href{#method-print}{\code{DataSet$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$new(dataRepository = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataRepository}}{Instance of the \code{DataRepository} object to wrap}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DotNetWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
