% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-set.R
\name{dataSetToDataFrame}
\alias{dataSetToDataFrame}
\title{Converts a list of DataSet objects to a data.frame}
\usage{
dataSetToDataFrame(dataSets)
}
\arguments{
\item{dataSets}{A list of DataSet objects or a single DataSet}
}
\value{
DataSet objects as dataframe with columns name, xValue, yValue, yErrorValues,
xDimension, xUnit, yDimension, yUnit, yErrorType, yErrorUnit, yMolWeight
}
\description{
Converts a list of DataSet objects to a data.frame
}
