% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-plot-configuration.R
\name{DefaultPlotConfiguration}
\alias{DefaultPlotConfiguration}
\title{Convenience class for plot configuration for OSP plots}
\description{
R6 class defining the configuration for \code{{ospsuite}} plots. It holds values
for all relevant plot properties.

Note that the values this objects contains are of general-purpose utility. In
other words, the public members of this class instance can be used to specify
arguments for base plots, ggplot plots, or any other plotting framework.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{xUnit, yUnit}}{Units for quantities plotted on x- and y-axes, respectively.}

\item{\code{title, subtitle, xLabel, yLabel, legendTitle, watermark}}{A character string
providing plot annotations for plot title, subtitle, x-axis label, y-axis
label, plot legend, watermark, respectively.}

\item{\code{titleColor, titleSize, titleFontFace, titleFontFamily, titleAngle}}{Aesthetic properties for the plot title.}

\item{\code{subtitleColor, subtitleSize, subtitleFontFace, subtitleFontFamily, subtitleAngle}}{Aesthetic properties for the plot subtitle.}

\item{\code{xLabelColor, xLabelSize, xLabelFontFace, xLabelFontFamily, xLabelAngle}}{Aesthetic properties for the plot xLabel.}

\item{\code{yLabelColor, yLabelSize, yLabelFontFace, yLabelFontFamily, yLabelAngle}}{Aesthetic properties for the plot yLabel.}

\item{\code{legendPosition}}{A character string defining the legend position.
Available options can be seen using \code{tlf::LegendPositions} list.}

\item{\code{legendTitleSize, legendTitleColor, legendTitleFontFamily, legendTitleFontFace, legendTitleAngle, legendCaptionSize}}{Aesthetic properties for the legend title.}

\item{\code{legendCaptionColor, legendCaptionFontFamily, legendCaptionFontFace, legendCaptionAngle}}{Aesthetic properties for the legend caption.}

\item{\code{xAxisTickLabels, xAxisLabelSize, xAxisLabelColor, xAxisLabelFontFamily, xAxisLabelFontFace, xAxisLabelAngle}}{Aesthetic properties for the x-axis label.}

\item{\code{yAxisTickLabels, yAxisLabelSize, yAxisLabelColor, yAxisLabelFontFamily, yAxisLabelFontFace, yAxisLabelAngle}}{Aesthetic properties for the y-axis label.}

\item{\code{xAxisLimits, yAxisLimits}}{A numeric vector of axis limits for the x-and
y-axis, respectively.}

\item{\code{xAxisTicks, yAxisTicks}}{A numeric vector or a function defining where to
position x-and y-axis ticks, respectively.}

\item{\code{xAxisScale, yAxisScale}}{A character string defining axis scale.
Available options can be seen using \code{tlf::Scaling} list.}

\item{\code{watermarkSize, watermarkColor, watermarkFontFamily, watermarkFontFace, watermarkAngle}}{A character string specifying the aesthetic properties for the watermark.}

\item{\code{plotBackgroundFill, plotBackgroundColor, plotBackgroundSize, plotBackgroundLinetype}}{A character string specifying the aesthetic properties for the plot background.}

\item{\code{plotPanelBackgroundFill, plotPanelBackgroundColor, plotPanelBackgroundSize, plotPanelBackgroundLinetype}}{A character string specifying the aesthetic properties for the plot panel (inside of plot) background.}

\item{\code{xAxisColor, xAxisSize, xAxisLinetype}}{A character string specifying the aesthetic properties for the x-axis.}

\item{\code{yAxisColor, yAxisSize, yAxisLinetype}}{A character string specifying the aesthetic properties for the y-axis.}

\item{\code{xGridColor, xGridSize, xGridLinetype}}{A character string specifying the aesthetic properties for the x-axis grid.}

\item{\code{yGridColor, yGridSize, yGridLinetype}}{A character string specifying the aesthetic properties for the y-axis grid.}

\item{\code{linesColor, linesFill, linesSize, linesLinetype, linesAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for lines.}

\item{\code{pointsColor, pointsFill, pointsShape, pointsSize, pointsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for points.}

\item{\code{ribbonsColor, ribbonsFill, ribbonsSize, ribbonsLinetype, ribbonsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for ribbons.}

\item{\code{errorbarsColor, errorbarsFill, errorbarsSize, errorbarsLinetype, errorbarsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for errorbars.}

\item{\code{plotSaveFileName, plotSaveFileFormat, plotSaveFileWidth, plotSaveFileHeight, plotSaveFileDimensionUnits, plotSaveFileDpi}}{File name (without extension) format to which the plot needs to be saved, and the specifications for saving the plot.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{xUnit, yUnit}}{Units for quantities plotted on x- and y-axes, respectively.}

\item{\code{title, subtitle, xLabel, yLabel, legendTitle, watermark}}{A character string
providing plot annotations for plot title, subtitle, x-axis label, y-axis
label, plot legend, watermark, respectively.}

\item{\code{titleColor, titleSize, titleFontFace, titleFontFamily, titleAngle}}{Aesthetic properties for the plot title.}

\item{\code{subtitleColor, subtitleSize, subtitleFontFace, subtitleFontFamily, subtitleAngle}}{Aesthetic properties for the plot subtitle.}

\item{\code{xLabelColor, xLabelSize, xLabelFontFace, xLabelFontFamily, xLabelAngle}}{Aesthetic properties for the plot xLabel.}

\item{\code{yLabelColor, yLabelSize, yLabelFontFace, yLabelFontFamily, yLabelAngle}}{Aesthetic properties for the plot yLabel.}

\item{\code{legendTitleSize, legendTitleColor, legendTitleFontFamily, legendTitleFontFace, legendTitleAngle, legendCaptionSize}}{Aesthetic properties for the legend title.}

\item{\code{legendCaptionColor, legendCaptionFontFamily, legendCaptionFontFace, legendCaptionAngle}}{Aesthetic properties for the legend caption.}

\item{\code{xAxisTickLabels, xAxisLabelSize, xAxisLabelColor, xAxisLabelFontFamily, xAxisLabelFontFace, xAxisLabelAngle}}{Aesthetic properties for the x-axis label.}

\item{\code{yAxisTickLabels, yAxisLabelSize, yAxisLabelColor, yAxisLabelFontFamily, yAxisLabelFontFace, yAxisLabelAngle}}{Aesthetic properties for the y-axis label.}

\item{\code{xAxisLimits, yAxisLimits}}{A numeric vector of axis limits for the x-and
y-axis, respectively.}

\item{\code{xAxisTicks, yAxisTicks}}{A numeric vector or a function defining where to
position x-and y-axis ticks, respectively.}

\item{\code{xAxisScale, yAxisScale}}{A character string defining axis scale.
Available options can be seen using \code{tlf::Scaling} list.}

\item{\code{watermarkSize, watermarkColor, watermarkFontFamily, watermarkFontFace, watermarkAngle}}{A character string specifying the aesthetic properties for the watermark.}

\item{\code{plotBackgroundFill, plotBackgroundColor, plotBackgroundSize, plotBackgroundLinetype}}{A character string specifying the aesthetic properties for the plot background.}

\item{\code{plotPanelBackgroundFill, plotPanelBackgroundColor, plotPanelBackgroundSize, plotPanelBackgroundLinetype}}{A character string specifying the aesthetic properties for the plot panel (inside of plot) background.}

\item{\code{xAxisColor, xAxisSize, xAxisLinetype}}{A character string specifying the aesthetic properties for the x-axis.}

\item{\code{yAxisColor, yAxisSize, yAxisLinetype}}{A character string specifying the aesthetic properties for the y-axis.}

\item{\code{xGridColor, xGridSize, xGridLinetype}}{A character string specifying the aesthetic properties for the x-axis grid.}

\item{\code{yGridColor, yGridSize, yGridLinetype}}{A character string specifying the aesthetic properties for the y-axis grid.}

\item{\code{linesColor, linesFill, linesSize, linesLinetype, linesAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for lines.}

\item{\code{pointsColor, pointsFill, pointsShape, pointsSize, pointsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for points.}

\item{\code{ribbonsColor, ribbonsFill, ribbonsSize, ribbonsLinetype, ribbonsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for ribbons.}

\item{\code{errorbarsColor, errorbarsFill, errorbarsSize, errorbarsLinetype, errorbarsAlpha}}{A selection key or values for choice of color, fill, shape, size, linetype, alpha, respectively, for errorbars.}

\item{\code{plotSaveFileName, plotSaveFileFormat, plotSaveFileWidth, plotSaveFileHeight, plotSaveFileDimensionUnits, plotSaveFileDpi}}{File name (without extension) format to which the plot needs to be saved, and the specifications for saving the plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DefaultPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
