% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.createAxesLabels}
\alias{.createAxesLabels}
\title{Create axes labels}
\usage{
.createAxesLabels(data, specificPlotConfiguration)
}
\arguments{
\item{data}{A data frame from \code{DataCombined$toDataFrame()}, which has
additionally been cleaned using \code{ospsuite:::.unitConverter()} to have the
same units across datasets.}

\item{specificPlotConfiguration}{The nature of labels will change depending
on the type of plot. The type of plot can be guessed from the specific
\code{PlotConfiguration} object used, since each plot has a unique corresponding
class.}
}
\description{
Create axes labels
}
\details{
If axes labels haven't been specified, create them using information about
dimensions and units present in the data frame produced by
\code{DataCombined$toDataFrame()}.
}
\examples{

df <- dplyr::tibble(
  dataType = c(rep("simulated", 3), rep("observed", 3)),
  xValues = c(0, 14.482, 28.965, 0, 1, 2),
  xUnit = "min",
  xDimension = "Time",
  yValues = c(1, 1, 1, 1, 1, 1),
  yUnit = "mol/ml",
  yDimension = ospDimensions$`Concentration (mass)`,
  yErrorValues = c(2.747, 2.918, 2.746, NA, NA, NA),
  molWeight = c(10, 10, 20, 20, 10, 10)
)

df <- ospsuite:::.unitConverter(df, xUnit = "min", yUnit = "mol/ml")

ospsuite:::.createAxesLabels(df, tlf::TimeProfilePlotConfiguration$new())

}
\seealso{
Other utilities-plotting:
\code{\link[=.addMissingGroupings]{.addMissingGroupings()}},
\code{\link[=.convertGeneralToSpecificPlotConfiguration]{.convertGeneralToSpecificPlotConfiguration()}},
\code{\link[=.extractAggregatedSimulatedData]{.extractAggregatedSimulatedData()}}
}
\concept{utilities-plotting}
\keyword{internal}
