% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.createObsVsPredData}
\alias{.createObsVsPredData}
\title{Created observed versus simulated paired data}
\usage{
.createObsVsPredData(data, tolerance = 0.001)
}
\arguments{
\item{data}{A data frame from \code{DataCombined$toDataFrame()}, which has been
further tidied using \code{.removeUnpairableDatasets()} and then
\code{.unitConverter()} functions.}

\item{tolerance}{Tolerance of comparison for observed and simulated time
points.}
}
\description{
Created observed versus simulated paired data
}
\keyword{internal}
