% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.createObsVsPredData}
\alias{.createObsVsPredData}
\title{Created observed versus simulated paired data}
\usage{
.createObsVsPredData(data, tolerance = NULL, scaling = tlf::Scaling$lin)
}
\arguments{
\item{data}{A data frame from \code{DataCombined$toDataFrame()}, which has been
further tidied using \code{.removeUnpairableDatasets()} and then
\code{.unitConverter()} functions.}

\item{tolerance}{Tolerance of comparison for observed and simulated time
points. Default is \code{NULL}, in which case the internal enum
\code{.thresholdByTimeUnit} will be used to decide on what threshold to use
based on the unit of time measurement.}

\item{scaling}{A character specifying scale: either linear (default) or
logarithmic.}
}
\description{
Created observed versus simulated paired data
}
\examples{

# create an example data frame
df <- dplyr::tibble(
  dataType = c(rep("observed", 5), rep("simulated", 3)),
  xValues = c(1, 3, 3.5, 4, 5, 0, 2, 4),
  yValues = c(1.9, 6.1, 7, 8.2, 1, 0, 4, 8),
  xUnit = ospUnits$Time$min
)

ospsuite:::.createObsVsPredData(df)

}
\seealso{
Other utilities-plotting: 
\code{\link{.addMissingGroupings}()},
\code{\link{.createAxesLabels}()},
\code{\link{.extractAggregatedSimulatedData}()},
\code{\link{.extractMatchingIndices}()},
\code{\link{.removeUnpairableDatasets}()}
}
\concept{utilities-plotting}
\keyword{internal}
