% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-combined.R
\name{.extractResidualsToTibble}
\alias{.extractResidualsToTibble}
\title{Created observed versus simulated paired data}
\usage{
.extractResidualsToTibble(data, scaling)
}
\arguments{
\item{data}{A data frame from \code{DataCombined$toDataFrame()}, which has been
further tidied using \code{.removeUnpairableDatasets()} and then
\code{.unitConverter()} functions.}
}
\description{
Created observed versus simulated paired data
}
\examples{
# create an example data frame
df <- dplyr::tibble(
  dataType = c(rep("observed", 5), rep("simulated", 3)),
  xValues = c(1, 3, 3.5, 4, 5, 0, 2, 4),
  xUnit = ospUnits$Time$min,
  xDimension = ospDimensions$Time,
  yValues = c(1.9, 6.1, 7, 8.2, 1, 0, 4, 8),
  yErrorValues = rnorm(8),
  yUnit = ospUnits$`Concentration [mass]`$`mg/l`,
  yDimension = ospDimensions$`Concentration (mass)`
)

ospsuite:::.extractResidualsToTibble(df, scaling = tlf::Scaling$lin)

}
\keyword{internal}
