% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{.getNetTask}
\alias{.getNetTask}
\title{Returns an instance of the specified \code{.NET} Task}
\usage{
.getNetTask(taskName)
}
\arguments{
\item{taskName}{The name of the task to retrieve (\strong{without} \code{Get} prefix).}
}
\value{
\itemize{
\item \code{.getNetTask()} returns an instance of of the specified \code{.NET} task.
\item \code{.getContainerTask()} returns a reference to the container task for
optimization purposes only.
}
}
\description{
Returns an instance of the specified \code{.NET} Task
}
\examples{

ospsuite:::.getNetTask("SimulationRunner")

ospsuite:::.getContainerTask()

}
\keyword{internal}
