% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pksim.R
\name{.getPathToPKSimInstallDir}
\alias{.getPathToPKSimInstallDir}
\title{Tries to find the installation path for a specific version of PK-Sim.
Searching is done in the following order:
\enumerate{
\item Search via Windows registry entry
\item Search via file system in a guessed installation folder from the base.search.folder (fallback 1)
}}
\usage{
.getPathToPKSimInstallDir(
  pksim.version = ospsuiteEnv$suiteVersion,
  base.search.folder = Sys.getenv("ProgramW6432")
)
}
\arguments{
\item{pksim.version}{The version number of Pk-Sim as a string.}

\item{base.search.folder}{The base folder for file system lookup fallback (default: 64-bit program folder)}
}
\value{
The path to the PK-Sim installation for version \code{pksim.version} or NA if no path could be found.
The path is separated with slashes (unix-style) and in compliance with base-R without a trailing slash.
If more than one matching path is found a warning is produced.
}
\description{
Tries to find the installation path for a specific version of PK-Sim.
Searching is done in the following order:
\enumerate{
\item Search via Windows registry entry
\item Search via file system in a guessed installation folder from the base.search.folder (fallback 1)
}
}
\examples{
\dontrun{
path <- .getPathToPKSimInstallDir("9")
path2 <- .getPathToPKSimInstallDir("10.1", "C:/MyOSPFolder/")
}
#' @keywords internal
}
