% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-importer-configuration.R
\docType{data}
\name{.ImporterErrorTypeToDataSetErrorType}
\alias{.ImporterErrorTypeToDataSetErrorType}
\title{Mapping of string representation for the error types supported by DataSet
to the values supported in the importer configuration}
\format{
An object of class \code{list} of length 2.
}
\usage{
.ImporterErrorTypeToDataSetErrorType
}
\description{
Mapping of string representation for the error types supported by DataSet
to the values supported in the importer configuration
}
\keyword{internal}
