% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-package.R
\name{.loadMoleculeTypeEnum}
\alias{.loadMoleculeTypeEnum}
\title{Build the \code{MoleculeType} enum by reading flag values from the underlying
\code{OSPSuite.Core.Domain.QuantityType} enum and keeping only molecule-relevant
members.}
\usage{
.loadMoleculeTypeEnum()
}
\description{
Build the \code{MoleculeType} enum by reading flag values from the underlying
\code{OSPSuite.Core.Domain.QuantityType} enum and keeping only molecule-relevant
members.
}
\keyword{internal}
