% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecules-building-block.R
\name{.quantityTypeNetObject}
\alias{.quantityTypeNetObject}
\title{Convert a \code{MoleculeType} integer flag value to a \code{.NET}
\code{OSPSuite.Core.Domain.QuantityType} enum object so that it can be passed to
.NET methods whose parameter type is \code{QuantityType}.}
\usage{
.quantityTypeNetObject(moleculeType)
}
\description{
Convert a \code{MoleculeType} integer flag value to a \code{.NET}
\code{OSPSuite.Core.Domain.QuantityType} enum object so that it can be passed to
.NET methods whose parameter type is \code{QuantityType}.
}
\keyword{internal}
