% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plotting.R
\name{.removeUnpairableDatasets}
\alias{.removeUnpairableDatasets}
\title{Remove unpairable datasets for scatter plots}
\usage{
.removeUnpairableDatasets(data)
}
\arguments{
\item{data}{A data frame returned by \code{DataCombined$toDataFrame()}.}
}
\description{
Datasets which haven't been assigned to any group will be removed from the
combined data frame.
}
\examples{

df <- dplyr::tribble(
  ~name, ~dataType, ~group,
  "Sim1", "Simulated", "GroupA",
  "Sim2", "Simulated", "GroupA",
  "Obs1", "Observed", "GroupB",
  "Obs2", "Observed", "GroupB",
  "Sim3", "Simulated", "GroupC",
  "Obs3", "Observed", "GroupC",
  "Sim4", "Simulated", "GroupD",
  "Obs4", "Observed", "GroupD",
  "Obs5", "Observed", "GroupD",
  "Sim5", "Simulated", "GroupE",
  "Sim6", "Simulated", "GroupE",
  "Obs7", "Observed", "GroupE",
  "Sim7", "Simulated", "GroupF",
  "Sim8", "Simulated", "GroupF",
  "Obs8", "Observed", "GroupF",
  "Obs9", "Observed", "GroupF",
  "Sim9", "Simulated", NA,
  "Obs10", "Observed", NA
)

# original
df

# transformed
ospsuite:::.removeUnpairableDatasets(df)

}
\keyword{internal}
