% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-net-wrapper.R
\docType{class}
\name{DotNetWrapper}
\alias{DotNetWrapper}
\title{Wrapper class for \code{.NET} objects}
\description{
Wrapper class for \code{.NET} objects

Wrapper class for \code{.NET} objects
}
\examples{

sim <- loadSimulation(system.file("extdata", "Aciclovir.pkml", package = "ospsuite"))

# looking at a reference to `.NET` simulation object
sim$ref

# create a new instance of `DotNetWrapper` class using this reference
DotNetWrapper$new(sim$ref)

}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{DotNetWrapper}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ref}}{Underlying \code{.NET} reference (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DotNetWrapper$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DotNetWrapper$new(ref)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Instance of the \code{.NET} object to wrap.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DotNetWrapper} object.
}
}
}
