% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-net-wrapper.R
\docType{class}
\name{DotNetWrapper}
\alias{DotNetWrapper}
\title{DotNetRwapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DotNetWrapper
}
\description{
Wrapper class for .net Object
}
\section{Fields}{

\describe{
\item{\code{ref}}{The actual .NET object instanced being wrapped}
}}

\section{Methods}{

\describe{
  \item{wrapProperty}{Simple way to wrap a get;set; .NET property}
  \item{wrapReadOnlyProperty}{Simple way to wrap a get; .NET readonly property}
  }
}

\keyword{datasets}
