% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-net-wrapper.R
\docType{class}
\name{DotNetWrapper}
\alias{DotNetWrapper}
\title{DotNetRwapper}
\description{
Wrapper class for .net Object
}
\section{Methods}{

\describe{
  \item{wrapProperty}{Simple way to wrap a get;set; .NET property}
  \item{wrapReadOnlyProperty}{Simple way to wrap a get; .NET readonly property}
  }
}

\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{DotNetWrapper}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ref}}{The actual .NET object instanced being wrapped}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DotNetWrapper$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DotNetWrapper$new(ref)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Instance of the .NET object to wrap.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DotNetWrapper` object.
}
}
}
