% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\docType{class}
\name{DotNetWrapper}
\alias{DotNetWrapper}
\title{DotNetRwapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DotNetWrapper
}
\description{
Wrapper class for .net Object
}
\section{Fields}{

\describe{
\item{\code{ref}}{The actual .NET object instanced being wrapped}
}}

\keyword{datasets}
