% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\docType{class}
\name{Entity}
\alias{Entity}
\title{Entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Entity
}
\description{
Abstract wrapper for an OSPSuite.Core Entity class
}
\section{Fields}{

\describe{
\item{\code{path}}{The path of the entity in the container hiearchy without the simulation name. (read-only)}

\item{\code{fullPath}}{Same as \code{path}, but with the simulation name. (read-only)}
}}

\keyword{datasets}
