% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\docType{class}
\name{Entity}
\alias{Entity}
\title{Entity}
\description{
Abstract wrapper for an OSPSuite.Core Entity class
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{Entity}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{The path of the entity in the container hiearchy without the simulation name. (read-only)}

\item{\code{fullPath}}{Same as \code{path}, but with the simulation name. (read-only)}
}
\if{html}{\out{</div>}}
}
