% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum}
\alias{enum}
\title{Create an enumeration to be used instead of arbitraty values in code.}
\usage{
enum(inputList)
}
\arguments{
\item{inputList}{}
}
\value{
the Enum created
}
\description{
Create an enumeration to be used instead of arbitraty values in code.
}
