% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum}
\alias{enum}
\title{Create an enumeration to be used instead of arbitraty values in code.}
\usage{
enum(enumValues)
}
\arguments{
\item{enumValues}{Values to use for the enum}
}
\value{
the Enum created
}
\description{
Create an enumeration to be used instead of arbitraty values in code.
}
\examples{

# Without predefined values
Color <- enum(c("Red", "Blue", "Green"))
myColor <- Color$Red

# With predefined values
Symbol <- enum(c(Diamond = 1, Triangle = 2, Circle = 2))
mySymbol <- Symbol$Diamond
}
