% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-analyses.R
\name{exportPKAnalysesToCSV}
\alias{exportPKAnalysesToCSV}
\title{Saves the pK-analyses  to csv file}
\usage{
exportPKAnalysesToCSV(pkAnalyses, filePath)
}
\arguments{
\item{pkAnalyses}{pK-Analyses to exporte (typically calculated using \code{calculatePKAnalyses} or imported from file)}

\item{filePath}{Full path where the pK-Analyses will be saved.}
}
\description{
Saves the pK-analyses  to csv file
}
