% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{exportResultsToCSV}
\alias{exportResultsToCSV}
\title{Saves the simulation results to csv file}
\usage{
exportResultsToCSV(results, simulation, filename)
}
\arguments{
\item{results}{Results to export (typically calculated using \code{runSimulation} or imported from file)}

\item{simulation}{Instance of a simulation used to calculate the results}

\item{filename}{Full path of where the results will be saved.}
}
\description{
Saves the simulation results to csv file
}
