% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-results.R
\name{exportResultsToCSV}
\alias{exportResultsToCSV}
\title{Saves the simulation results to csv file}
\usage{
exportResultsToCSV(results, filePath)
}
\arguments{
\item{results}{Results to export (typically calculated using \code{runSimulation} or imported from file)}

\item{filePath}{Full path where the results will be saved.}
}
\description{
Saves the simulation results to csv file
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load the simulation
sim <- loadSimulation(simPath)

# Add some outputs to the simulation
addOutputs("Organism|**|*", sim)

# Run the simulation
results <- runSimulation(sim)

# Export the results to csv file
exportResultsToCSV(results, tempfile())
}
