% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\docType{class}
\name{Formula}
\alias{Formula}
\title{Formula}
\usage{
Formula
}
\description{
A formula of the model (Typically related to a \code{Quantity} such as a parameter)
}
\section{Fields}{

\describe{
\item{\code{isTable}}{Is this a table formula (Read-Only)}

\item{\code{isConstant}}{Is this a constant formula (Read-Only)}
}}

\keyword{datasets}
