% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R, R/table-formula.R
\docType{class}
\name{Formula}
\alias{Formula}
\alias{TableFormula}
\title{Formula}
\usage{
Formula

TableFormula
}
\description{
A formula of the model (Typically related to a \code{Quantity} such as a parameter)

A table formula of the model (Typically related to a \code{Quantity} such as a parameter)
}
\section{Fields}{

\describe{
\item{\code{dimension}}{The dimension in which the quantity is defined  (Read-Only)}

\item{\code{isTable}}{Is this a table formula (Read-Only)}

\item{\code{isConstant}}{Is this a constant formula (Read-Only)}

\item{\code{isExplicit}}{Is this an explicit formula (Read-Only)}

\item{\code{isDistributed}}{Is this a distributed formula (Read-Only)}

\item{\code{formulaString}}{Returns the formula as a string for an \code{ExplicitFormula} or \code{NULL} otherwise (Read-Only).}

\item{\code{allPoints}}{Returns all points defined in the table formulafor a \code{TableFormula} or \code{NULL}  otherwise (Read-Only).}

\item{\code{allPoints}}{Returns all points defined in the table formulafor a \code{TableFormula} or \code{NULL}  otherwise (Read-Only).}

\item{\code{useDerivedValues}}{Indicates whether table values should be derived during solving. the ODE system. Default value is \code{TRUE}}

\item{\code{xDimension}}{The dimension in which the x values are defined (Read-Only).}
}}

\section{Methods}{

\describe{
  \item{ addPoint(x, y)}{Adds a point to the table, with x value in base unit for XDimension and y value in base unit for Dimension}
  \item{ removePoint(x, y)}{ Remove the point having the same x and y from the table}
  }
}

\keyword{datasets}
