% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getAllContainersMatching}
\alias{getAllContainersMatching}
\title{Retrieve all sub containers of a parent container (simulation or container instance) matching the given path criteria}
\usage{
getAllContainersMatching(paths, container)
}
\arguments{
\item{paths}{A vector of strings representing the paths relative to the \code{container}}

\item{container}{A Container or Simulation used to find the containers}
}
\value{
A list of containers matching the path criteria. The list is empty if no containers matching were found.
}
\description{
Retrieve all sub containers of a parent container (simulation or container instance) matching the given path criteria
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Intracellular` containers defined in all direct containers of the organism
containers <- getAllContainersMatching("Organism|*|Intracellular", sim)

# Return all `Intracellular` containers defined in all direct containers of the organism
# and the container "Interstitial" under 'Organism|Brain'
paths <- c("Organism|*|Intracellular", "Organism|Brain|Interstitial")
containers <- getAllContainersMatching(paths, sim)

# Returns all `Intracellular` containers defined in `Organism` and all its subcontainers
containers <- getAllContainersMatching("Organism|**|Intracellular", sim)
}
\seealso{
\code{\link{loadSimulation}} and \code{\link{getContainer}} to create objects of type Container or Simulation
}
