% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-entity.R
\name{getAllEntitiesMatching}
\alias{getAllEntitiesMatching}
\title{Retrieve all entities of a container (simulation or container instance) matching the given path criteria.}
\usage{
getAllEntitiesMatching(paths, container, entityType)
}
\arguments{
\item{paths}{A vector of strings representing the paths relative to the \code{container}}

\item{container}{A Container or Simulation used to find the entities}

\item{entityType}{Class of the type that should be returned. Supported types are Container, Quantity, and Parameter}
}
\value{
A list of entities matching the path criteria coerced to the \code{entityType}.
The list is empty if no entities matching were found.
}
\description{
Retrieve all entities of a container (simulation or container instance) matching the given path criteria.
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to create objects of type Container or Simulation
}
