% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getAllParametersMatching}
\alias{getAllParametersMatching}
\title{Retrieve all parameters of a container (simulation or container instance) matching the given path criteria}
\usage{
getAllParametersMatching(paths, container)
}
\arguments{
\item{paths}{A vector of strings relative to the \code{container}}

\item{container}{A Container or Simulation used to find the parameters}
}
\value{
A list of parameters matching the path criteria. The list is empty if no parameters matching were found.
}
\description{
Retrieve all parameters of a container (simulation or container instance) matching the given path criteria
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Return all `Volume` parameters defined in all direct containers of the organism
params <- getAllParametersMatching("Organism|*|Volume", sim)

# Return all `Volume` parameters defined in all direct containers of the organism
# and the parameter 'Weight (tissue)' of the container 'Liver'
paths <- c("Organism|*|Volume", "Organism|Liver|Weight (tissue)")
params <- getAllParametersMatching(paths, sim)

# Returns all `Volume` parameters defined in `Organism` and all its subcontainers
params <- getAllParametersMatching("Organism|**|Volume", sim)
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to create objects of type Container or Simulation
}
