% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getBaseUnit}
\alias{getBaseUnit}
\title{Get base unit of a dimension}
\usage{
getBaseUnit(dimension)
}
\arguments{
\item{dimension}{Dimension (string name) for which the base unit is returned.}
}
\value{
String name of the base unit.
}
\description{
Get base unit of a dimension
}
