% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getContainer}
\alias{getContainer}
\title{Retrieve a single container by path under the given container}
\usage{
getContainer(path, container, stopIfNotFound = TRUE)
}
\arguments{
\item{path}{A string representing the path relative to the \code{container}}

\item{container}{A Container or Simulation used to find the containers}

\item{stopIfNotFound}{Boolean. If TRUE and no container exist for the given path,
an error is thrown. Default is TRUE.}
}
\value{
The \code{Container} with the given path. If the container for the path
does not exist, an error is thrown if \code{stopIfNotFound} is TRUE (default),
otherwise \code{NULL}
}
\description{
Retrieve a single container by path under the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
container <- getContainer("Organism|Liver", sim)
}
\seealso{
\code{\link{loadSimulation}} and \code{\link{getContainer}} to create objects of type Container or Simulation
}
