% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-container.R
\name{getContainer}
\alias{getContainer}
\title{Retrieve a single container by path under the given container}
\usage{
getContainer(path, container)
}
\arguments{
\item{path}{A string representing the path relative to the \code{container}}

\item{container}{A Container or Simulation used to find the containers}

\item{path}{An absolute path relative to the \code{container}}
}
\value{
The \code{Container} with the given path or null if not found
}
\description{
Retrieve a single container by path under the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getContainer("Organism|Liver", sim)
}
\seealso{
\code{\link{loadSimulation}} and \code{\link{getContainer}} to create objects of type Container or Simulation
}
