% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getDimensionByName}
\alias{getDimensionByName}
\title{Returns the an instance of the dimension with the given name if found or NULL otherwise}
\usage{
getDimensionByName(name)
}
\arguments{
\item{dimension}{Name of dimension that should be retrieved}
}
\description{
Returns the an instance of the dimension with the given name if found or NULL otherwise
}
\examples{
dim <- getDimensionByName("Time")
}
