% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getDimensionForUnit}
\alias{getDimensionForUnit}
\title{Returns the dimension that can be used to support the given unit.
Returns null}
\usage{
getDimensionForUnit(unit)
}
\arguments{
\item{unit}{Unit used to find the corresponding dimension}
}
\description{
Returns the dimension that can be used to support the given unit.
Returns null
}
\examples{
dim <- getDimensionForUnit("mg")

}
