% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{getNetTask}
\alias{getNetTask}
\alias{getContainerTask}
\title{Returns an instance of the specified .NET Task}
\usage{
getNetTask(taskName)

getContainerTask()
}
\arguments{
\item{taskName}{The name of the task to retrieve (without the Get)}
}
\value{
An instance of the Task
}
\description{
Returns an instance of the specified .NET Task
}
\details{
simulationLoader <- getNetTask("SimulationLoader")
}
