% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-net-task.R
\name{getNetTask}
\alias{getNetTask}
\title{Returns an instace of the specified .NET Task}
\usage{
getNetTask(taskName)
}
\arguments{
\item{taskName}{The name of the task to retrieve (without the Get)}
}
\value{
An instance of the Task
}
\description{
Returns an instace of the specified .NET Task
}
\examples{

simulationLoader <- getNetTask("SimulationLoader")
}
