% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-results.R
\name{getOutputValues}
\alias{getOutputValues}
\title{Returns the simulated values for the selected outputs (e.g molecules or parameters)}
\usage{
getOutputValues(
  simulationResults,
  quantitiesOrPaths = NULL,
  population = NULL,
  individualIds = NULL,
  stopIfNotFound = TRUE
)
}
\arguments{
\item{simulationResults}{Object of type \code{SimulationResults} produced by calling \code{runSimulation}
on a \code{Simulation} object.}

\item{quantitiesOrPaths}{Quantity instances (element or vector) typically retrieved using
\code{getAllQuantitiesMatching} or quantity path (element or vector of strings) for which the results are to be returned. (optional)
When providing the paths, only absolute full paths are supported (i.e., no matching with '*' possible).
If quantitiesOrPaths is \code{NULL} (default value), returns the results for all output defined in the results.}

\item{population}{population used to calculate the simulationResults (optional). This is used only to add the population covariates to the resulting data table.}

\item{individualIds}{\code{numeric} IDs of individiuals for which the results should be extracted.
By default, all individuals from the results are considered. If the individual with the provided ID is not found, the ID is ignored}

\item{stopIfNotFound}{If \code{TRUE} (default) an error is thrown if no results exist for any \code{path}
If \code{FALSE}, a list of \code{NA} values is returned for the respecitve path.}
}
\description{
The function receives an object of simulation results generated by running the simulation
and returns time-values profiles for the chosen quantities. Results of a simulation
of a single individual is treated as a population simulation with only one individual.
}
