% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-results.R
\name{getOutputValuesTLF}
\alias{getOutputValuesTLF}
\title{Export values that can be plugged directly into the TLF lib (TODO)}
\usage{
getOutputValuesTLF(simulationResults, population,
  quantitiesOrPaths = NULL, individualIds = NULL)
}
\description{
Export values that can be plugged directly into the TLF lib (TODO)
}
