% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{getParameter}
\alias{getParameter}
\title{Retrieve a single parameter by path in the given container}
\usage{
getParameter(path, container)
}
\arguments{
\item{path}{A vector of strings relative to the \code{container}}

\item{container}{A Container or Simulation used to find the parameters}
}
\value{
The \code{Parameter} with the given path or \code{NULL} if not found
}
\description{
Retrieve a single parameter by path in the given container
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getParameter(c("Organism", "Liver", "Volume"), sim)
}
\seealso{
\code{\link{loadSimulation}}, \code{\link{getContainer}} and \code{\link{getAllContainersMatching}} to create objects of type Container or Simulation
}
