% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-explorer.R
\name{getSimulationTree}
\alias{getSimulationTree}
\title{Given a simulation file path or an instance of a simulation, traverses the simulation structure and returns a tree like structure
allowing for intuitive navigation in the simulation tree}
\usage{
getSimulationTree(simulationOrFilePath)
}
\arguments{
\item{simulationOrFilePath}{Full path of the simulation to load or instance of a simulation}
}
\value{
A list with a branched structure representing the path tree of Quantities in the simulation file.
At the end of each branch is a string called 'path' that is the path of the quantity represented by the branch.
}
\description{
Given a simulation file path or an instance of a simulation, traverses the simulation structure and returns a tree like structure
allowing for intuitive navigation in the simulation tree
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

tree <- getSimulationTree(sim)

liver_volume_path <- tree$Organism$Liver$Volume$path
}
