% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getUnitConversionFactor}
\alias{getUnitConversionFactor}
\title{Get conversion factor from one unit to another}
\usage{
getUnitConversionFactor(fromUnit, toUnit, dimension, MW = NULL)
}
\arguments{
\item{fromUnit}{String name of the unit to convert from}

\item{toUnit}{String name of the unit to convert to}

\item{dimension}{Dimension of the units (conversion between mass and molar possible when
\code{MW} is provided)}

\item{MW}{Optional molecule weight in g/mol to use when converting, for example, from molar to mass amount or concentration}
}
\value{
Numerical value. A value in \code{fromUnit} multiplied by the factor
corresponds to the value in \code{toUnit}
}
\description{
Get conversion factor from one unit to another
}
