% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{getUnitsForDimension}
\alias{getUnitsForDimension}
\title{Get units for a given dimension}
\usage{
getUnitsForDimension(dimension)
}
\arguments{
\item{dimension}{Name of dimension for which units should be returned}
}
\value{
Returns a vector containing all units defined in the dimension
}
\description{
Get units for a given dimension
}
\examples{

getUnitsForDimension("Mass")
}
