% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifNotNull}
\alias{ifNotNull}
\title{Mimic the ternary operator  \code{a ? x : y} behavior in other languages
If \code{condition} is not null, returns \code{outputIfNotNull} otherwise \code{outputIfNull}}
\usage{
ifNotNull(condition, outputIfNotNull, outputIfNull = NULL)
}
\arguments{
\item{condition}{The .NET object instances (single or list) to wrap}

\item{outputIfNotNull}{The class definition that will be used to convert the parameter}

\item{outputIfNull}{The class definition that will be used to convert the parameter}
}
\description{
Mimic the ternary operator  \code{a ? x : y} behavior in other languages
If \code{condition} is not null, returns \code{outputIfNotNull} otherwise \code{outputIfNull}
}
