% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifNotNull}
\alias{ifNotNull}
\title{ifnotnull}
\usage{
ifNotNull(condition, outputIfNotNull, outputIfNull = NULL)
}
\arguments{
\item{condition}{argument 1}

\item{outputIfNotNull}{argument 2}

\item{outputIfNull}{argument 3}
}
\value{
outputIfNotNull if condition is not null, outputIfNull otherwise
}
\description{
Check if condition is not null, if so output outputIfNotNull,
otherwise, output outputIfNull
}
\details{
Shortkey checking if argument 1 is not null,
output the argument 2 if not null, or output argument 3 otherwise
}
