% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-characteristics.R
\docType{class}
\name{IndividualCharacteristics}
\alias{IndividualCharacteristics}
\title{IndividualCharacteristics}
\usage{
IndividualCharacteristics
}
\description{
Characteristics of an individual describing its origin
}
\section{Fields}{

\describe{
\item{\code{species}}{Specifies the species of the individual. It should be a species available in PK-Sim (see \code{Species})}

\item{\code{population}}{For a Human species, the population of intereset. It should be a population available in PK-Sim (see \code{HumanPopulation})}

\item{\code{gender}}{Gender of the individual. It should be defined for the species in PK-Sim  (see \code{Gender})}

\item{\code{age}}{Age of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{gestationalAge}}{Gestational Age of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{weight}}{Weight of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{height}}{Height of the individual as in instance of a \code{SnapshotParameter} (optional)}
}}

\keyword{datasets}
