% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-characteristics.R
\docType{class}
\name{IndividualCharacteristics}
\alias{IndividualCharacteristics}
\title{IndividualCharacteristics}
\description{
Characteristics of an individual describing its origin
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{IndividualCharacteristics}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{species}}{Specifies the species of the individual. It should be a species available in PK-Sim (see \code{Species})}

\item{\code{population}}{For a Human species, the population of intereset. It should be a population available in PK-Sim (see \code{HumanPopulation})}

\item{\code{gender}}{Gender of the individual. It should be defined for the species in PK-Sim  (see \code{Gender})}

\item{\code{age}}{Age of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{gestationalAge}}{Gestational Age of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{weight}}{Weight of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{height}}{Height of the individual as in instance of a \code{SnapshotParameter} (optional)}

\item{\code{allMoleculeOntogenies}}{All molecule ontogenies defined for this individual characteristics.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IndividualCharacteristics$new()}}
\item \href{#method-print}{\code{IndividualCharacteristics$print()}}
\item \href{#method-addMoleculeOntogeny}{\code{IndividualCharacteristics$addMoleculeOntogeny()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="finalize">}\href{../../ospsuite/html/DotNetWrapper.html#method-finalize}{\code{ospsuite::DotNetWrapper$finalize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndividualCharacteristics$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{IndividualCharacteristics} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndividualCharacteristics$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMoleculeOntogeny"></a>}}
\if{latex}{\out{\hypertarget{method-addMoleculeOntogeny}{}}}
\subsection{Method \code{addMoleculeOntogeny()}}{
Add a molecule ontogeny \code{MoleculeOntogeny} to the individual characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndividualCharacteristics$addMoleculeOntogeny(moleculeOntogeny)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moleculeOntogeny}}{Molecule ontogeny to add}
}
\if{html}{\out{</div>}}
}
}
}
