% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pksim.R
\name{initPKSim}
\alias{initPKSim}
\title{Loads the PKSim.R that will enable create individual and create population workflows.}
\usage{
initPKSim(pksimFolderPath = NULL)
}
\arguments{
\item{pksimFolderPath}{Path where PK-Sim is installed. If this is not specified, path will be read from registry using the package version}
}
\description{
Loads the PKSim.R that will enable create individual and create population workflows.
}
\note{
This will only work on Windows machine and should not be called on any other OS.
This function should also only be called explicitely when using a portable install of the package
}
