% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\docType{class}
\name{Interval}
\alias{Interval}
\title{Interval}
\usage{
Interval
}
\description{
Simulation Interval (typically associated with an instace of \code{OutputSchema})
}
\section{Fields}{

\describe{
\item{\code{startTime}}{Start time of interval (instance of \code{Parameter})}

\item{\code{endTime}}{End time of interval (instance of \code{Parameter})}

\item{\code{resolution}}{Resolution of interval in pts/min (instance of \code{Parameter})}
}}

\keyword{datasets}
