% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-set.R
\name{loadDataSetsFromExcel}
\alias{loadDataSetsFromExcel}
\title{Load data sets from excel}
\usage{
loadDataSetsFromExcel(
  xlsFilePath,
  importerConfigurationOrPath,
  importAllSheets = FALSE
)
}
\arguments{
\item{xlsFilePath}{Path to the excel file with the data}

\item{importerConfigurationOrPath}{An object of type \code{DataImporterConfiguration} that is valid
for the excel file or a path to a XML file with stored configuration}

\item{importAllSheets}{If \code{FALSE} (default), only sheets specified in the
\code{importerConfiguration} will be loaded. If \code{TRUE}, an attempt to load all sheets
is performed. If any sheet does not comply with the configuration, an error is thrown.}
}
\value{
A named set of \code{DataSet} objects. The naming is defined by the property
\code{importerConfiguration$namingPattern}.
}
\description{
Load data sets from excel
}
\details{
Load observed data from an excel file using an importer configuration
}
\examples{
xlsFilePath <- system.file("extdata", "CompiledDataSet.xlsx", package = "ospsuite")
importerConfiguration <- createImporterConfigurationForFile(xlsFilePath)
importerConfiguration$sheets <- "TestSheet_1"

dataSets <- loadDataSetsFromExcel(
  xlsFilePath = xlsFilePath,
  importerConfigurationOrPath = importerConfiguration,
  importAllSheets = FALSE
)

importerConfigurationFilePath <- system.file("extdata", "dataImporterConfiguration.xml", package = "ospsuite")

dataSets <- loadDataSetsFromExcel(
  xlsFilePath = xlsFilePath,
  importerConfigurationOrPath = importerConfigurationFilePath,
  importAllSheets = FALSE
)
}
