% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mobi-module.R
\name{loadModuleFromPKML}
\alias{loadModuleFromPKML}
\title{Load a MoBi module from pkml.}
\usage{
loadModuleFromPKML(path)
}
\arguments{
\item{path}{Path to the pkml file with a module export}
}
\value{
A \code{MoBiModule} object
}
\description{
Load a MoBi module from pkml.
}
\examples{
filePath <- system.file("extdata", "Thyroid.pkml", package = "ospsuite")
module <- loadModuleFromPKML(filePath)
}
