% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-simulation.R
\name{loadSimulation}
\alias{loadSimulation}
\title{Load a simulation from a pkml file and returns the simulation}
\usage{
loadSimulation(pkmlSimulationFile)
}
\arguments{
\item{pkmlSimulationFile}{Full path of pkml simulation file to load.}
}
\description{
Load a simulation from a pkml file and returns the simulation
}
