% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{loadSimulation}
\alias{loadSimulation}
\title{Loads a simulation from a pkml file and returns the simulation. If the passed simulation file
has been loaded before, the simulation is not loaded again but a cached object is returned.
This behavior can be overriden.}
\usage{
loadSimulation(pkmlSimulationFile, loadFromCache = TRUE,
  addToCache = TRUE)
}
\arguments{
\item{pkmlSimulationFile}{Full path of pkml simulation file to load.}

\item{loadFromCache}{If TRUE, an already loaded pkml file will not be loaded
again, but the simulation object will be retrieved from cache. This is the
default behavior. If FALSE, new object will be created.}

\item{addToCache}{If TRUE, the loaded simulation is added to cache. If false,
the returned simulation only exists locally. Default is TRUE.}
}
\description{
Loads a simulation from a pkml file and returns the simulation. If the passed simulation file
has been loaded before, the simulation is not loaded again but a cached object is returned.
This behavior can be overriden.
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")

# Load sim1 for the first time
sim1 <- loadSimulation(simPath)

# sim2 will be loaded from cache and will represent the same object as sim1
sim2 <- loadSimulation(simPath)

parameter1 <- getParameter(toPathString(c("Organism", "Liver", "Volume")), sim1)
parameter2 <- getParameter(toPathString(c("Organism", "Liver", "Volume")), sim2)

# parameter1 and parameter2 belong to the same simulation object, so changing
# one of the them will also change another
setParametersValues(parameters = parameter2, values = 0)
parameter1$value == parameter2$value # TRUE

# sim3 will be loaded from not from cache
sim3 <- loadSimulation(simPath, loadFromCache = FALSE)
# parameter3 belong to different simulation object than parameter1 and parameter2
parameter3 <- getParameter(toPathString(c("Organism", "Liver", "Volume")), sim3)
setParametersValues(parameters = parameter3, values = 1)
parameter2$value == parameter3$value # FALSE#'
}
