% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{MergeBehavior}
\alias{MergeBehavior}
\title{Merge Behaviors}
\format{
An object of class \code{list} of length 2.
}
\usage{
MergeBehavior
}
\description{
Merge behaviors available for modules in MoBi. Loaded at
package init from \code{OSPSuite.Core.Domain.MergeBehavior}. Used by
\code{MoBiModule$mergeBehavior}.
}
\keyword{datasets}
