% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messages}
\alias{messages}
\title{List of functions and strings used to signal error messages
Extends the \code{messages} list from ospsuite.utils}
\usage{
messages
}
\description{
List of functions and strings used to signal error messages
Extends the \code{messages} list from ospsuite.utils
}
