% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{Error message when single output is expected but multiple entities produced.}
\format{An object of class \code{list} of length 5.}
\usage{
messages
}
\arguments{
\item{path}{A vector of strings relative to the \code{container}}

\item{container}{A Container or Simulation used to find the entities}

\item{optionalMessage}{A string that will be appended to the end of the message}
}
\value{
A string representation of the error message
}
\description{
Error message when single output is expected but multiple entities produced.
}
\keyword{datasets}
