% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{List of functions and strings used to signal error messages
Extends the \code{messages} list from ospsuite.utils}
\format{
An object of class \code{list} of length 54.
}
\usage{
messages
}
\description{
List of functions and strings used to signal error messages
Extends the \code{messages} list from ospsuite.utils
}
\keyword{datasets}
