% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBiModule.R
\docType{class}
\name{MoBiModule}
\alias{MoBiModule}
\title{MoBi Module}
\value{
A named list of \code{BuildingBlock} objects, with names being the names of the PV BBs.

A named list of \code{BuildingBlock} objects, with names being the names of the IC BBs.

A \code{MoleculesBuildingBlock} object exposing molecule-name
queries (e.g. \code{allMoleculeNames()}, \code{allMoleculeNamesOfType()},
\code{moleculeTypeFor()}), or \code{NULL} if the module has no Molecules BB.
}
\description{
A MoBi module, either loaded from a project or from a pkml file
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{MoBiModule}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the module (read-only)}

\item{\code{isPKSimModule}}{Whether the module is a PK-Sim module (read-only)}

\item{\code{mergeBehavior}}{Merge behavior of the module. Must be one of "Extend" or "Overwrite".}

\item{\code{parameterValuesBBnames}}{Names of the Parameter Values Building Blocks (PV BBs) in the module (read-only)}

\item{\code{initialConditionsBBnames}}{Names of the Initial Conditions Building Blocks (IC BBs) in the module (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MoBiModule-new}{\code{MoBiModule$new()}}
\item \href{#method-MoBiModule-getParameterValuesBBs}{\code{MoBiModule$getParameterValuesBBs()}}
\item \href{#method-MoBiModule-getInitialConditionsBBs}{\code{MoBiModule$getInitialConditionsBBs()}}
\item \href{#method-MoBiModule-getMoleculesBB}{\code{MoBiModule$getMoleculesBB()}}
\item \href{#method-MoBiModule-print}{\code{MoBiModule$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printClass"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printClass'><code>rSharp::NetObject$.printClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printLine"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printLine'><code>rSharp::NetObject$.printLine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticFields'><code>rSharp::NetObject$getStaticFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticMethods'><code>rSharp::NetObject$getStaticMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticProperties'><code>rSharp::NetObject$getStaticProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$new(netObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{netObject}}{Reference to \code{NetObject} .NET MoBi-module object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MoBiModule} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-getParameterValuesBBs"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-getParameterValuesBBs}{}}}
\subsection{Method \code{getParameterValuesBBs()}}{
Get the list of Parameter Values Building Blocks (PV BBs) in the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$getParameterValuesBBs(names = NULL, stopIfNotFound = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Optional names of the Parameter Values Building Block to retrieve.
If \code{NULL}, returns all PV BBs.}

\item{\code{stopIfNotFound}}{If \code{TRUE} (default), an error is thrown if any of the specified
parameter values BB is not present in the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-getInitialConditionsBBs"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-getInitialConditionsBBs}{}}}
\subsection{Method \code{getInitialConditionsBBs()}}{
Get the list of Initial Conditions Building Blocks (IC BBs) in the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$getInitialConditionsBBs(names = NULL, stopIfNotFound = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Optional names of the Initial Conditions Building Block to retrieve.
If \code{NULL}, returns all IC BBs.}

\item{\code{stopIfNotFound}}{If \code{TRUE} (default), an error is thrown if any of the specified
initial conditions BB is not present in the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-getMoleculesBB"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-getMoleculesBB}{}}}
\subsection{Method \code{getMoleculesBB()}}{
Get the \code{Molecules} Building Block of the module, if any.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$getMoleculesBB()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-print"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$print(printClassProperties = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{printClassProperties}}{Logical, whether to print class properties (default: \code{FALSE}). If \code{TRUE}, calls first the \code{print} method of the parent class.
Useful for debugging.}

\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
