% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module.R
\docType{class}
\name{MoBiModule}
\alias{MoBiModule}
\title{MoBi Module # OR SHOULD WE CALL IT AN OspModule ?}
\description{
A MoBi module, either loaded from a project or from a pkml file
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{MoBiModule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MoBiModule-new}{\code{MoBiModule$new()}}
\item \href{#method-MoBiModule-getBuildingBlocks}{\code{MoBiModule$getBuildingBlocks()}}
\item \href{#method-MoBiModule-isPkSimModule}{\code{MoBiModule$isPkSimModule()}}
\item \href{#method-MoBiModule-print}{\code{MoBiModule$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printClass"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printClass'><code>rSharp::NetObject$.printClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printLine"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printLine'><code>rSharp::NetObject$.printLine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticFields'><code>rSharp::NetObject$getStaticFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticMethods'><code>rSharp::NetObject$getStaticMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticProperties'><code>rSharp::NetObject$getStaticProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$new(netObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{netObject}}{Reference to \code{NetObject} .NET MoBi-module object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MoBiModule} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-getBuildingBlocks"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-getBuildingBlocks}{}}}
\subsection{Method \code{getBuildingBlocks()}}{
Get the building blocks from the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$getBuildingBlocks(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Optional, type of building blocks to return. Must be one of \code{BuildingBlockTypes}.
If \code{NULL} (default), all building blocks of the module are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of building block objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-isPkSimModule"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-isPkSimModule}{}}}
\subsection{Method \code{isPkSimModule()}}{
Is this module a PK-Sim module?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$isPkSimModule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiModule-print"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiModule-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiModule$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
