% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBiProject.R
\docType{class}
\name{MoBiProject}
\alias{MoBiProject}
\title{MoBi Project}
\value{
A new \code{MoBiProject} object.

A \code{Simulation} object, if the simulation with the given name is present in the
project. \code{NULL} if no such simulation is available.

A named list of \code{DataSet} objects. \code{NULL} if the project does not contain
any observed data or no specified data sets are found.

A named list of \code{MoBiModule} objects. \code{NULL} if the project does not contain
any module.

An object of the type \code{BuildingBlock}. \code{NULL} if the project does not contain
such an individual and \code{stopIfNotFound = FALSE}.

A named list of objects of the type \code{BuildingBlock}. \code{NULL} for each
specified expression profile that is not present in the project  if \code{stopIfNotFound = FALSE}.

A \code{SimulationConfiguration} object.

Path of the file to which the project is saved.
}
\description{
A MoBi project, containing modules, building blocks, simulations, etc.
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{MoBiProject}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{expressionProfilesNames}}{Names of the expression profiles that are present in the project}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sourceFile}}{Path to the file the project was loaded from (read-only)}

\item{\code{simulationNames}}{Names of the simulations that are present in the project (read-only)}

\item{\code{parameterIdentificationNames}}{Names of the parameter identifications
that are present in the project (read-only)
2DO}

\item{\code{individualsNames}}{Names of the individuals that are present in the project (read-only)}

\item{\code{expressionProfilesNames}}{Names of the expression profiles that are present in the project}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MoBiProject-new}{\code{MoBiProject$new()}}
\item \href{#method-MoBiProject-getSimulation}{\code{MoBiProject$getSimulation()}}
\item \href{#method-MoBiProject-getObservedData}{\code{MoBiProject$getObservedData()}}
\item \href{#method-MoBiProject-getModules}{\code{MoBiProject$getModules()}}
\item \href{#method-MoBiProject-getIndividual}{\code{MoBiProject$getIndividual()}}
\item \href{#method-MoBiProject-getExpressionProfiles}{\code{MoBiProject$getExpressionProfiles()}}
\item \href{#method-MoBiProject-createSimulationConfiguration}{\code{MoBiProject$createSimulationConfiguration()}}
\item \href{#method-MoBiProject-saveProject}{\code{MoBiProject$saveProject()}}
\item \href{#method-MoBiProject-print}{\code{MoBiProject$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printClass"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printClass'><code>rSharp::NetObject$.printClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id=".printLine"><a href='../../rSharp/html/NetObject.html#method-NetObject-.printLine'><code>rSharp::NetObject$.printLine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticFields'><code>rSharp::NetObject$getStaticFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticMethods'><code>rSharp::NetObject$getStaticMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getStaticProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getStaticProperties'><code>rSharp::NetObject$getStaticProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-new"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class

Should not be directly used. Instead, use function \code{loadMoBiProject()}
to load a project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$new(netObject, sourceFile = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{netObject}}{Reference to \code{NetObject} .NET MoBi-project object}

\item{\code{sourceFile}}{(Optional) File used to load the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-getSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-getSimulation}{}}}
\subsection{Method \code{getSimulation()}}{
Load a simulation from the project
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$getSimulation(simulationName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationName}}{Name of the simulation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-getObservedData"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-getObservedData}{}}}
\subsection{Method \code{getObservedData()}}{
Get observed data present in the project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$getObservedData(dataSetNames = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSetNames}}{Optional. List of names of observed data sets to retrieve
from project. If \code{NULL}, all data sets are returned. If a specified data set
is not found, the name is ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-getModules"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-getModules}{}}}
\subsection{Method \code{getModules()}}{
Get modules present in the project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$getModules()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-getIndividual"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-getIndividual}{}}}
\subsection{Method \code{getIndividual()}}{
Get a specified individual from the project
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$getIndividual(name, stopIfNotFound = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the individual}

\item{\code{stopIfNotFound}}{If \code{TRUE} (default), an error is thrown if the specified
individual is not present in the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-getExpressionProfiles"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-getExpressionProfiles}{}}}
\subsection{Method \code{getExpressionProfiles()}}{
Get specified expression profiles from the project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$getExpressionProfiles(names, stopIfNotFound = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{List of names of the expression profiles to retrieve.}

\item{\code{stopIfNotFound}}{If \code{TRUE} (default), an error is thrown if any of the specified
expression profiles is not present in the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-createSimulationConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-createSimulationConfiguration}{}}}
\subsection{Method \code{createSimulationConfiguration()}}{
Create a simulation configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$createSimulationConfiguration(
  modulesNames,
  individualName = NULL,
  expressionProfilesNames = NULL,
  selectedInitialConditions = NULL,
  selectedParameterValues = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modulesNames}}{A list of the modules from which to create in simulation.
All defined modules must be present in the project. The order of module names defines the order in which the modules will be combined to a simulation!}

\item{\code{individualName}}{Optional, name of the individual.}

\item{\code{expressionProfilesNames}}{Optional, list of expression profiles to apply to the simulation.}

\item{\code{selectedInitialConditions}}{By default, the first Initial Conditions
(IC) building block (BB) of each module will be selected. If a module has multiple
IC BBs, it is possible to specify which IC BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the IC BB.
By setting the value to \code{NULL}, no IC BB from the specified module will be applied.}

\item{\code{selectedParameterValues}}{By default, the first Parameter Values
(PV) building block (BB) of each module will be selected. If a module has multiple
PV BBs, it is possible to specify which PV BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the PV BB.
By setting the value to \code{NULL}, no PV BB from the specified module will be applied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-saveProject"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-saveProject}{}}}
\subsection{Method \code{saveProject()}}{
PRIO 2

Save the project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$saveProject(filePath = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filePath}}{Path to the file, including file name, where the project should be saved to.
If \code{NULL} (default), the project is saved to the same file it was loaded from.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoBiProject-print"></a>}}
\if{latex}{\out{\hypertarget{method-MoBiProject-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoBiProject$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
