% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{class}
\name{Molecule}
\alias{Molecule}
\title{Molecule}
\usage{
Molecule
}
\description{
A molecule defined in a compartment of the system
}
\details{
Derived from \link{Quantity}, please see base class documentation.
}
\keyword{Molecule}
