% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule-ontogeny.R
\docType{class}
\name{MoleculeOntogeny}
\alias{MoleculeOntogeny}
\title{MoleculeOntogeny}
\usage{
MoleculeOntogeny
}
\description{
Use when retrieving indvidual values using the createIndividualAlgorithm. This class is a simple pair {MoleculeName, Ontogeny}
allowing the user to retrieve potential ontogeny values.
}
\section{Fields}{

\describe{
\item{\code{molecule}}{Name of the molecule in the model}

\item{\code{ontogeny}}{Name of the ontogeny to use for the molecule}
}}

\keyword{datasets}
