% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{MoleculeParameter}
\alias{MoleculeParameter}
\title{Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim}
\usage{
MoleculeParameter
}
\description{
Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim
}
