% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-path.R
\docType{data}
\name{MoleculeParameter}
\alias{MoleculeParameter}
\title{Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim}
\format{An object of class \code{list} of length 5.}
\usage{
MoleculeParameter
}
\description{
Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim
}
\keyword{datasets}
