% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{MoleculeType}
\alias{MoleculeType}
\title{MoleculeType}
\format{
An object of class \code{list} of length 7.
}
\usage{
MoleculeType
}
\description{
Subset of the .NET \code{OSPSuite.Core.Domain.QuantityType} flag
values used to filter molecules in a \code{Molecules} building block.
Values are integer flags as defined in .NET. \code{Protein} is the umbrella
flag covering \code{Enzyme}, \code{Transporter}, and \code{OtherProtein}.
Populated at package load time from the bundled .NET assembly via
\code{.loadEnums()}; defined here as a placeholder for documentation.
}
\keyword{datasets}
