% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{MoleculeType}
\alias{MoleculeType}
\title{Molecule Types}
\format{
An object of class \code{list} of length 7.
}
\usage{
MoleculeType
}
\description{
Allowed types for molecules.
}
\keyword{datasets}
