% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\docType{class}
\name{ObjectBase}
\alias{ObjectBase}
\title{ObjectBase}
\usage{
ObjectBase
}
\description{
Abstract wrapper for an OSPSuite.Core ObjectBase.
}
\section{Fields}{

\describe{
\item{\code{id}}{The id of the .NET wrapped object. (read-only)}

\item{\code{name}}{The name of the object}
}}

\keyword{datasets}
