% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\docType{data}
\name{ospDimensions}
\alias{ospDimensions}
\title{Supported dimensions defined as a named list}
\format{
An object of class \code{list} of length 91.
}
\usage{
ospDimensions
}
\description{
Supported dimensions defined as a named list
}
\details{
ospDimensions$Mass => "Mass"
}
\keyword{datasets}
