% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospPlotConfiguration.R
\name{ospPlotConfiguration}
\alias{ospPlotConfiguration}
\title{Convenience class storing plot labels}
\description{
R6 class defining the labels for \code{{ospsuite}} plots.

Note that the values this objects contains are of general-purpose utility. In
other words, the public members of this class instance can be used to specify
arguments for base plots, ggplot plots, or any other plotting framework.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title, subtitle, xlabel, ylabel, legendTitle}}{A character string
providing plot annotations for plot title, subtitle, x-axis label, y-axis
label, plot legend, respectively.}

\item{\code{xUnit, yUnit}}{Units for x- and y-axes, respectively.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{title, subtitle, xlabel, ylabel, legendTitle}}{A character string
providing plot annotations for plot title, subtitle, x-axis label, y-axis
label, plot legend, respectively.}

\item{\code{xUnit, yUnit}}{Units for x- and y-axes, respectively.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{ospPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ospPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
