% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospPlotConfiguration.R
\name{ospPlotConfiguration}
\alias{ospPlotConfiguration}
\title{Plot configuration for OSP plots}
\description{
R6 class defining the configuration of a \code{{ospsuite}} plots.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{A character string providing plot title.}

\item{\code{subtitle}}{A character string providing plot subtitle.}

\item{\code{xlabel}}{A character string providing plot x-axis label.}

\item{\code{ylabel}}{A character string providing plot y-axis label.}

\item{\code{legend}}{\code{tlf::LegendConfiguration} object defining legend properties.}

\item{\code{xAxis}}{\code{tlf::XAxisConfiguration} object defining x-axis properties.}

\item{\code{yAxis}}{\code{tlf::YAxisConfiguration} object defining y-axis properties.}

\item{\code{watermark}}{\code{tlf::Label} object defining watermark.}

\item{\code{lines}}{\code{tlf::ThemeAestheticSelections} object or list defining how
lines are plotted.}

\item{\code{points}}{\code{tlf::ThemeAestheticSelections} object or list defining how
points are plotted.}

\item{\code{ribbons}}{\code{tlf::ThemeAestheticSelections} object or list defining
how ribbons are plotted.}

\item{\code{errorbars}}{\code{tlf::ThemeAestheticSelections} object or list defining
how errorbars are plotted.}

\item{\code{export}}{R6 class \code{tlf::SaveConfiguration} defining saving properties.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ospPlotConfiguration$new()}}
\item \href{#method-clone}{\code{ospPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ospPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ospPlotConfiguration$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  legend = NULL,
  xAxis = NULL,
  yAxis = NULL,
  watermark = NULL,
  lines = NULL,
  points = NULL,
  ribbons = NULL,
  errorbars = NULL,
  export = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title, subtitle, xlabel, ylabel}}{A character string providing plot
annotations for plot title, subtitle, x-axis label, y-axis label, plot
legend, watermark, respectively.}

\item{\code{legend}}{\code{tlf::LegendConfiguration} object defining legend properties.}

\item{\code{xAxis}}{\code{tlf::XAxisConfiguration} object defining x-axis properties.}

\item{\code{yAxis}}{\code{tlf::YAxisConfiguration} object defining y-axis properties.}

\item{\code{watermark}}{\code{tlf::Label} object defining watermark.}

\item{\code{lines}}{\code{tlf::ThemeAestheticSelections} object or list defining how
lines are plotted.}

\item{\code{points}}{\code{tlf::ThemeAestheticSelections} object or list defining how
points are plotted.}

\item{\code{ribbons}}{\code{tlf::ThemeAestheticSelections} object or list defining
how ribbons are plotted.}

\item{\code{errorbars}}{\code{tlf::ThemeAestheticSelections} object or list defining
how errorbars are plotted.}

\item{\code{export}}{R6 class \code{tlf::SaveConfiguration} defining saving
properties.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ospPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
