% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-env.R
\docType{data}
\name{ospsuiteSettingNames}
\alias{ospsuiteSettingNames}
\title{Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteSetting()}}
\format{
An object of class \code{list} of length 10.
}
\usage{
ospsuiteSettingNames
}
\description{
Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteSetting()}
}
\keyword{datasets}
