% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-schema.R
\docType{class}
\name{OutputSchema}
\alias{OutputSchema}
\title{OutputSchema}
\description{
Output schema associated with a given simulation
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{OutputSchema}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{intervals}}{All intervals defined in the schema (Read-Only)}

\item{\code{timePoints}}{All single time points defined in the schema (Read-Only)}

\item{\code{endTime}}{Returns the end time of the simulation in kernel unit (Read-Only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputSchema-clear}{\code{OutputSchema$clear()}}
\item \href{#method-OutputSchema-addInterval}{\code{OutputSchema$addInterval()}}
\item \href{#method-OutputSchema-removeInterval}{\code{OutputSchema$removeInterval()}}
\item \href{#method-OutputSchema-addTimePoints}{\code{OutputSchema$addTimePoints()}}
\item \href{#method-OutputSchema-print}{\code{OutputSchema$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize"><a href='../../ospsuite/html/DotNetWrapper.html#method-DotNetWrapper-initialize'><code>ospsuite::DotNetWrapper$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSchema-clear"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSchema-clear}{}}}
\subsection{Method \code{clear()}}{
Clears all intervals and time points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSchema$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSchema-addInterval"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSchema-addInterval}{}}}
\subsection{Method \code{addInterval()}}{
Adds an interval to the schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSchema$addInterval(interval)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interval}}{Interval to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSchema-removeInterval"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSchema-removeInterval}{}}}
\subsection{Method \code{removeInterval()}}{
Removes the interval from the schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSchema$removeInterval(interval)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interval}}{Interval to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSchema-addTimePoints"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSchema-addTimePoints}{}}}
\subsection{Method \code{addTimePoints()}}{
Adds the time points to the schema. Note that time points and intervals exists concurrently.
Use time points only if you need to ensure that specific time are used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSchema$addTimePoints(timePoints)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timePoints}}{Time points to add to the schema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSchema-print"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSchema-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSchema$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
