% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-schema.R
\docType{class}
\name{OutputSchema}
\alias{OutputSchema}
\title{OutputSchema}
\usage{
OutputSchema
}
\description{
Output schema associated with a given simulation
}
\section{Fields}{

\describe{
\item{\code{intervals}}{All intervals defined in the schema (Read-Only)}

\item{\code{timePoints}}{All single time points defined in the schema (Read-Only)}

\item{\code{endTime}}{Returns the end time of the simulation in kernel unit (Read-Only)}
}}

\section{Methods}{

\describe{
  \item{clear}{Clears all intervals and time points}
  \item{addInterval(interval)}{Adds an interval to the schema}
  \item{removeInterval(interval)}{Removes the interval from the schema}
  \item{addTimePoints(timePoints)}{Adds the time points to the schema}
  }
}

\keyword{datasets}
