% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-selections.R
\docType{class}
\name{OutputSelections}
\alias{OutputSelections}
\title{OutputSelections}
\description{
List of selected quantities selected as output for a given simulation
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{OutputSelections}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{allOutputs}}{Returns all outputs defined in the selection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputSelections-clear}{\code{OutputSelections$clear()}}
\item \href{#method-OutputSelections-addQuantity}{\code{OutputSelections$addQuantity()}}
\item \href{#method-OutputSelections-removeQuantity}{\code{OutputSelections$removeQuantity()}}
\item \href{#method-OutputSelections-print}{\code{OutputSelections$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize"><a href='../../ospsuite/html/DotNetWrapper.html#method-DotNetWrapper-initialize'><code>ospsuite::DotNetWrapper$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSelections-clear"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSelections-clear}{}}}
\subsection{Method \code{clear()}}{
Removes all selected output from the selection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSelections$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSelections-addQuantity"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSelections-addQuantity}{}}}
\subsection{Method \code{addQuantity()}}{
Adds a quantity as selected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSelections$addQuantity(quantity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{Quantity to add to the selection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSelections-removeQuantity"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSelections-removeQuantity}{}}}
\subsection{Method \code{removeQuantity()}}{
Removes a quantity from the selection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSelections$removeQuantity(quantity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{Quantity to remove from the selection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputSelections-print"></a>}}
\if{latex}{\out{\hypertarget{method-OutputSelections-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputSelections$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
