% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-selections.R
\docType{class}
\name{OutputSelections}
\alias{OutputSelections}
\title{OutputSelections}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OutputSelections
}
\description{
List of selected quantities selected as output for a given simulation
}
\keyword{datasets}
