% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{class}
\name{Parameter}
\alias{Parameter}
\title{Parameter}
\usage{
Parameter
}
\description{
A model parameter
}
\details{
Derived from \link{Quantity}, please see base class documentation.
}
\keyword{Parameter}
