% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\docType{class}
\name{Parameter}
\alias{Parameter}
\title{Parameter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Parameter
}
\description{
A model parameter with a value
}
\section{Fields}{

\describe{
\item{\code{value}}{The value of the parameter}
}}

\keyword{datasets}
