% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{PartitionCoefficientMethods}
\alias{PartitionCoefficientMethods}
\title{Available methods for calculation of partition coefficients.}
\format{
An object of class \code{list} of length 5.
}
\usage{
PartitionCoefficientMethods
}
\description{
Available methods for calculation of partition coefficients.
}
\keyword{datasets}
