% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationConfiguration.R
\docType{data}
\name{PartitionCoefficientMethods}
\alias{PartitionCoefficientMethods}
\title{Available methods for calculation of partition coefficients.}
\format{
An object of class \code{list} of length 5.
}
\usage{
PartitionCoefficientMethods
}
\description{
MUST WE DEFINE IT HERE, OR CAN WE GET IT FROM .NET SOMEHOW?
}
\keyword{datasets}
