% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter.R
\docType{class}
\name{PKParameter}
\alias{PKParameter}
\title{PKParameter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PKParameter
}
\description{
pK-Parameter values for all indiviudals of a simulation (1 or more) calculated for \code{quantityPath}
}
\section{Fields}{

\describe{
\item{\code{values}}{All values for \code{quantityPath} and \code{name}}

\item{\code{name}}{The name of the pK-Parameter (AUC, Cmax, Tmax etc...)}

\item{\code{quantityPath}}{The path of the quantity for which the values were calculated}
}}

\keyword{datasets}
