% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter.R
\docType{class}
\name{PKParameter}
\alias{PKParameter}
\title{PKParameter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PKParameter
}
\description{
Results of a simulation run (either individual or population simulation)
}
\keyword{datasets}
