% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter.R
\docType{class}
\name{PKParameter}
\alias{PKParameter}
\title{PKParameter}
\description{
Standard PK Parameters defined in the OSPSuite
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{PKParameter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the PK-Parameter}

\item{\code{displayName}}{Display Name of the PK-Parameter. If not set, Name will be used}

\item{\code{dimension}}{Dimension instance used by the PK-Parameter (Read-Only)}

\item{\code{unit}}{Unit of the PK-Parameter (Read-Only)}

\item{\code{displayUnit}}{Display Unit used for the PK-Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PKParameter-print}{\code{PKParameter$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize"><a href='../../ospsuite/html/DotNetWrapper.html#method-DotNetWrapper-initialize'><code>ospsuite::DotNetWrapper$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKParameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-PKParameter-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
