% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pk-parameter.R
\name{pkParameterByName}
\alias{pkParameterByName}
\title{Returns an instance of a PK-Parameter by name or NULL if the parameter by name is not found}
\usage{
pkParameterByName(name, stopIfNotFound = TRUE)
}
\arguments{
\item{name}{Name of PK-Parameter to update}

\item{stopIfNotFound}{Indicates whether an exception is thrown when the PK-Parameter was not found. Default is \code{TRUE}}
}
\description{
Returns an instance of a PK-Parameter by name or NULL if the parameter by name is not found
}
\examples{

pkParameter <- pkParameterByName(name = "t_max")
}
