% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter-sensitivity.R
\docType{class}
\name{PKParameterSensitivity}
\alias{PKParameterSensitivity}
\title{PKParameterSensitivity}
\usage{
PKParameterSensitivity
}
\description{
Sensitivity of a PK Parameter for one output for a given parameter
}
\section{Fields}{

\describe{
\item{\code{parameterPath}}{Unique name of parameter in sensitivity analysis}

\item{\code{pkParameterName}}{Name of PK Output (Cmax, Tmax etc...)}

\item{\code{outputPath}}{Path of underlying quantity for which pk-analyses were performed}

\item{\code{value}}{Value of sensitivity}
}}

\keyword{datasets}
