% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter-sensitivity.R
\docType{class}
\name{PKParameterSensitivity}
\alias{PKParameterSensitivity}
\title{PKParameterSensitivity}
\description{
Sensitivity of a PK Parameter for one output for a given parameter
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{PKParameterSensitivity}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameterName}}{Unique name of parameter in sensitivity analysis}

\item{\code{pkParameterName}}{Name of PK Output (Cmax, Tmax etc...)}

\item{\code{outputPath}}{Path of underlying quantity for which pk-analyses were performed}

\item{\code{value}}{Value of sensitivity}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PKParameterSensitivity-print}{\code{PKParameterSensitivity$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize"><a href='../../ospsuite/html/DotNetWrapper.html#method-DotNetWrapper-initialize'><code>ospsuite::DotNetWrapper$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKParameterSensitivity-print"></a>}}
\if{latex}{\out{\hypertarget{method-PKParameterSensitivity-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKParameterSensitivity$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
