% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Create a plot grid}
\usage{
plotGrid(PlotGridConfiguration)
}
\arguments{
\item{PlotGridConfiguration}{A \code{PlotGridConfiguration} object, which
is an \code{R6} class object that defines properties of a plot grid (like number
of rows, columns, labels, etc.).}
}
\description{
Create a plot grid using the \code{patchwork::wrap_plots()} function. The required
arguments are supplied through the \code{PlotGridConfiguration} object.
}
\examples{

library(ospsuite)
library(ggplot2)
library(tlf)

# plots to be arranged in a grid
set.seed(123)
ls_plots <- list(
  tlf::plotHistogram(x = rnorm(100)),
  tlf::plotHistogram(x = rnorm(100, mean = 3)),
  tlf::plotHistogram(x = rnorm(100, mean = 10))
)

# create an instance of plot configuration class
plotGridObj <- PlotGridConfiguration$new(plotList = ls_plots)

# specify further customizations for the plot grid
plotGridObj$title <- "my combined plot"
plotGridObj$subtitle <- "something clever"
plotGridObj$caption <- "my sources"
plotGridObj$nColumns <- 2L
plotGridObj$tagLevels <- "A"
plotGridObj$tagPrefix <- "Plot ("
plotGridObj$tagSuffix <- ")"

# plot the grid
plotGrid(plotGridObj)

}
\references{
For more, see: \url{https://patchwork.data-imaginist.com/articles/patchwork.html}
}
