% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Create a plot grid}
\usage{
plotGrid(PlotGridConfiguration)
}
\arguments{
\item{PlotGridConfiguration}{A \code{PlotGridConfiguration} object, which
is an \code{R6} class object that defines properties of a plot grid (like number
of rows, columns, labels, etc.).}
}
\description{
Create a plot grid using the \code{patchwork::wrap_plots()} function. The required
arguments are supplied through the \code{PlotGridConfiguration} object.
}
\examples{

library(tlf)
library(ospsuite)

ls_plots <- list(
  plotHistogram(x = rnorm(100)),
  plotHistogram(x = rnorm(100, mean = 3)),
  plotHistogram(x = rnorm(100, mean = 10))
)

plotGridObj <- createPlotGridConfiguration(
  plotList = ls_plots,
  title = "my combined plot",
  subtitle = "something clever",
  caption = "something dumb"
)

plotGrid(plotGridObj)

}
\references{
For more, see: \url{https://patchwork.data-imaginist.com/articles/patchwork.html}
}
