% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualTimeProfile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Concentration time profile plot}
\usage{
plotIndividualTimeProfile(
  dataCombined,
  individualTimeProfilePlotConfiguration = TimeProfilePlotConfiguration$new(),
  tlfTheme = NULL
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{individualTimeProfilePlotConfiguration}{A \code{PlotConfiguration} object.
For more, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/plot-configuration.html}.}

\item{tlfTheme}{A path to JSON file containing
\href{https://www.open-systems-pharmacology.org/TLF-Library/reference/Theme.html}{\code{Theme}}
object for \code{{tlf}} library.}
}
\description{
Concentration time profile plot
}
