% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualTimeProfile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Concentration time profile plot}
\usage{
plotIndividualTimeProfile(
  dataCombined,
  ospPlotConfiguration,
  xUnit = NULL,
  yUnit = NULL,
  tlfTheme = NULL
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{ospPlotConfiguration}{A \code{ospPlotConfiguration} object, which is an \code{R6}
class object that defines plot properties (like labels, axes scaling and
limits, legend position, etc.). You can create instance of this object
using \code{createPlotConfiguration()} function.}

\item{xUnit, yUnit}{Units for x- and y-axes, respectively.}

\item{tlfTheme}{A path to JSON file containing
\href{https://www.open-systems-pharmacology.org/TLF-Library/reference/Theme.html}{\code{Theme}}
object for \code{{tlf}} library.}
}
\description{
Concentration time profile plot
}
