% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual-time-profile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Time-profile plot of individual data}
\usage{
plotIndividualTimeProfile(dataCombined, defaultPlotConfiguration = NULL)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}
}
\description{
Time-profile plot of individual data
}
\examples{

# TODO: add example

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotPopulationTimeProfile}()}
}
\concept{plotting}
