% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualTimeProfile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Concentration time profile plot}
\usage{
plotIndividualTimeProfile(
  dataCombined,
  theme = NULL,
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  legendTitle = NULL
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{theme}{Theme to be used as default of the tlf environment}
}
\description{
Concentration time profile plot
}
