% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual-time-profile.R
\name{plotIndividualTimeProfile}
\alias{plotIndividualTimeProfile}
\title{Concentration time profile plot}
\usage{
plotIndividualTimeProfile(
  dataCombined,
  defaultPlotConfiguration = DefaultPlotConfiguration$new()
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}
}
\description{
Concentration time profile plot
}
