% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-observed-vs-simulated.R
\name{plotObservedVsSimulated}
\alias{plotObservedVsSimulated}
\title{Observed versus predicted/simulated scatter plot}
\usage{
plotObservedVsSimulated(
  dataCombined,
  defaultPlotConfiguration = NULL,
  smoother = NULL,
  foldDistance = 2
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{smoother}{Optional name of smoother function:
\itemize{
\item \code{"loess"} for loess regression
\item \code{"lm"} for linear regression
}}

\item{foldDistance}{A vector for plotting lines at required fold distances
The vector can include only fold distance values different from \code{1}. Even
if it is not specified, it will \strong{always} be included.}
}
\description{
Observed versus predicted/simulated scatter plot
}
\examples{

# TODO: add example

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotPopulationTimeProfile}()}
}
\concept{plotting}
