% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population-time-profile.R
\name{plotPopulationTimeProfile}
\alias{plotPopulationTimeProfile}
\title{Time-values profile plot for population simulations}
\usage{
plotPopulationTimeProfile(
  dataCombined,
  defaultPlotConfiguration = NULL,
  aggregation = "quantiles",
  quantiles = c(0.05, 0.5, 0.95),
  showLegendPerDataset = FALSE,
  ...
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class containing both
observed and simulated datasets to be compared.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{aggregation}{The type of the aggregation of individual data. One of
\code{quantiles} (Default), \code{arithmetic} or \code{geometric} (full list in
\code{ospsuite::DataAggregationMethods}). Will replace \code{yValues} by the median,
arithmetic or geometric average and add a set of upper and lower bounds
(\code{yValuesLower} and \code{yValuesHigher}).}

\item{quantiles}{A numerical vector with quantile values (Default: \code{c(0.05, 0.50, 0.95)}) to be plotted. Ignored if \code{aggregation} is not \code{quantiles}.}

\item{showLegendPerDataset}{Logical flag to display separate legend entries
for observed and simulated datasets, if available. This is experimental and
may not work reliably when both observed and simulated datasets > 1. Defaults
to \code{FALSE}.}

\item{...}{additionnal arguments to pass to \code{.extractAggregatedSimulatedData()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The simulated values will be aggregated across individuals for each
time point.

For \code{aggregation = quantiles} (default), the quantile values defined in the
argument \code{quantiles} will be used. In the profile plot, the middle value
will be used to draw a line, while the lower and upper values will be used
as the lower und upper ranges. For \code{aggregation = arithmetic}, arithmetic
mean with arithmetic standard deviation (SD) will be plotted. Use the
optional parameter \code{nsd} to change the number of SD to plot above and below
the mean. For \code{aggregation = geometric}, geometric mean with geometric
standard deviation (SD) will be plotted. Use the optional parameter \code{nsd} to
change the number of SD to plot above and below the mean.
}
\examples{
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)

populationResults <- importResultsFromCSV(
  simulation = sim,
  filePaths = system.file("extdata", "SimResults.csv", package = "ospsuite")
)

# Create a new instance of `DataCombined` class
myDataComb <- DataCombined$new()
myDataComb$addSimulationResults(populationResults)


# plot
plotPopulationTimeProfile(myDataComb)

# plot with other quantiles
plotPopulationTimeProfile(myDataComb, quantiles = c(0.1, 0.5, 0.9))

# plot with arithmetic mean
plotPopulationTimeProfile(myDataComb,
  aggregation = "arithmetic"
)

}
\seealso{
Other plotting:
\code{\link{DefaultPlotConfiguration}},
\code{\link[=plotIndividualTimeProfile]{plotIndividualTimeProfile()}},
\code{\link[=plotObservedVsSimulated]{plotObservedVsSimulated()}},
\code{\link[=plotResidualsVsSimulated]{plotResidualsVsSimulated()}},
\code{\link[=plotResidualsVsTime]{plotResidualsVsTime()}}
}
\concept{plotting}
