% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population-time-profile.R
\name{plotPopulationTimeProfile}
\alias{plotPopulationTimeProfile}
\title{Time-values profile plot for population simulations}
\usage{
plotPopulationTimeProfile(
  dataCombined,
  defaultPlotConfiguration = NULL,
  quantiles = c(0.05, 0.5, 0.95),
  aggregation = "quantiles"
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{quantiles}{A numerical vector with quantile values (Default: \code{c(0.05, 0.50, 0.95)}), with the quantile values defining the aggregation of
individual data. Ignored if aggregation is different than \code{quantiles}.}

\item{aggregation}{The type of the aggregation of individual data. One of
\code{quantiles} (Default), \code{arithmetic} or \code{geometric}. Quantiles will
replace \code{yValues} by the median and add a set of upper and lower bounds
(\code{yValuesLower} and \code{yValuesHigher}). \code{arithmetic} or \code{geometric} will only
replace \code{yValues} by means.}
}
\description{
Time-values profile plot for population simulations
}
\examples{
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)

populationResults <- importResultsFromCSV(
  simulation = sim,
  filePaths = system.file("extdata", "SimResults_pop.csv", package = "ospsuite")
)

# Create a new instance of `DataCombined` class
myDataComb <- DataCombined$new()
myDataComb$addSimulationResults(populationResults)


# plot
plotPopulationTimeProfile(myDataComb)

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotResidualsVsSimulated}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
