% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population-time-profile.R
\name{plotPopulationTimeProfile}
\alias{plotPopulationTimeProfile}
\title{Time-values profile plot for population simulations}
\usage{
plotPopulationTimeProfile(
  dataCombined,
  defaultPlotConfiguration = NULL,
  quantiles = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{quantiles}{A numerical vector with quantile values (Default: \code{c(0.05, 0.50, 0.95)}), with the quantile values defining the aggregation of
individual data. In the profile plot, the middle value will be used to draw
a line, while the lower and upper values will be used to create a ribbon.}
}
\description{
Time-values profile plot for population simulations
}
\examples{
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)

populationResults <- importResultsFromCSV(
  simulation = sim,
  filePaths = system.file("extdata", "SimResults_pop.csv", package = "ospsuite")
)

# Create a new instance of `DataCombined` class
myDataComb <- DataCombined$new()
myDataComb$addSimulationResults(populationResults)

# Create a new instance of `DefaultPlotConfiguration` class
myPlotConfiguration <- DefaultPlotConfiguration$new()
myPlotConfiguration$title <- "My Plot Title"
myPlotConfiguration$subtitle <- "My Plot Subtitle"
myPlotConfiguration$caption <- "My Sources"

# plot
plotPopulationTimeProfile(myDataComb, myPlotConfiguration)

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotResidualsVsSimulated}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
