% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population-time-profile.R
\name{plotPopulationTimeProfile}
\alias{plotPopulationTimeProfile}
\title{Time-values profile plot for population simulations}
\usage{
plotPopulationTimeProfile(
  dataCombined,
  defaultPlotConfiguration = NULL,
  aggregation = "quantiles",
  ...
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{aggregation}{The type of the aggregation of individual data. One of
\code{quantiles} (Default), \code{arithmetic} or \code{geometric} (full list in ospsuite::DataAggregationMethods). Will
replace \code{yValues} by the median, arithmetic or geometric average and add a set of upper and lower bounds
(\code{yValuesLower} and \code{yValuesHigher})}

\item{...}{Extra parameters to pass to aggregating functions. \code{probs} for \code{stats::quantile} or \code{n} for the number of
standard deviation to add below and above the average for \code{arithmetic} or \code{geometric}.}
}
\description{
Time-values profile plot for population simulations
}
\examples{
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)

populationResults <- importResultsFromCSV(
  simulation = sim,
  filePaths = system.file("extdata", "SimResults_pop.csv", package = "ospsuite")
)

# Create a new instance of `DataCombined` class
myDataComb <- DataCombined$new()
myDataComb$addSimulationResults(populationResults)


# plot
plotPopulationTimeProfile(myDataComb)

# plot with other quantiles
plotPopulationTimeProfile(myDataComb, probs = c(0.1, 0.5, 0.9))

# Previous quantiles argument still works
plotPopulationTimeProfile(myDataComb, quantiles = c(0.1, 0.5, 0.9))

# plot with arithmetic mean
plotPopulationTimeProfile(myDataComb,
                          aggregation = "arithmetic")
# plot with arithmetic mean and 3sd above and below average line
plotPopulationTimeProfile(myDataComb,
                          aggregation = "arithmetic",
                          n=3)

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotResidualsVsSimulated}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
