% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-with-ospsuite-plots.R
\name{plotPredictedVsObserved}
\alias{plotPredictedVsObserved}
\title{Plots predicted vs observed data, grouped by "group".}
\usage{
plotPredictedVsObserved(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  xyScale = "log",
  comparisonLineVector = ospsuite.plots::getFoldDistanceList(folds = c(2)),
  ...
)
}
\arguments{
\item{plotData}{An object of class \code{DataCombined} or a \code{data.table}.
If a \code{data.table}, it must include the following:
\itemize{
\item \code{xValues}: Numeric time points.
\item \code{yValues}: Observed or simulated values (numeric).
\item \code{group}: Grouping variable (factor or character).
\item \code{name}: Name for the dataset (factor or character).
\item \code{xUnit}: Unit of the x-axis values (character).
\item \code{yUnit}: Unit of the y-axis values (character).
\item \code{dataType}: Specifies data type—either \code{observed} or \code{simulated}.
\item Optional:
\itemize{
\item \code{yErrorType}: Type of y error, relative to \code{ospsuite::DataErrorType}.
\item \code{yErrorValues}: Numeric error values.
\item \code{yMin}, \code{yMax}: Custom ranges for y-axis instead of error types.
\item \code{IndividualId}: Used for aggregation of simulated population data.
}
}}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data. Expected structure includes information about dimensions and units for both x and y axes.}

\item{mapping}{A ggplot2 aesthetic mapping object. Default is \code{ggplot2::aes()}. This is added or replaces the default mapping constructed by the data.}

\item{xyScale}{A character string specifying the scale for the x and y-axis.
Default is 'log'.}

\item{comparisonLineVector}{A named list generated by the function
\code{ospsuite.plots::getFoldDistanceList}. This list contains fold distances, where each
entry represents a fold and its reciprocal. The identity fold (1)
will be included if specified in \code{getFoldDistanceList}.}

\item{...}{Additional arguments passed to \code{ospsuite.plots::plotPredVsObs}.}
}
\value{
A \code{ggplot2} plot object representing predicted vs observed values, including aesthetics for the x and y axes.
}
\description{
This function visualizes the relationship between predicted and observed values, allowing for easy identification of discrepancies.
}
\examples{
\dontrun{
# Generate a predicted vs observed plot for the provided data
plotPredictedVsObserved(myDataCombined)
}
}
\seealso{
Other plot functions based on ospsuite.plots: 
\code{\link{plotQuantileQuantilePlot}()},
\code{\link{plotResidualsVsObserved}()},
\code{\link{plotResidualsVsTimePoints}()},
\code{\link{plotTimeProfile}()}
}
\concept{plot functions based on ospsuite.plots}
