% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-with-ospsuite-plots.R
\name{plotQuantileQuantilePlot}
\alias{plotQuantileQuantilePlot}
\title{Plot Quantile-Quantile Plot}
\usage{
plotQuantileQuantilePlot(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  residualScale = "log",
  ...
)
}
\arguments{
\item{plotData}{An object of class \code{DataCombined} or a \code{data.table}. If a
\code{data.table}, it must include the following:
\itemize{
\item \code{xValues}: Numeric time points.
\item \code{yValues}: Observed or simulated values (numeric).
\item \code{group}: Grouping variable (factor or character).
\item \code{name}: Name for the dataset (factor or character).
\item \code{xUnit}: Unit of the x-axis values (character).
\item \code{yUnit}: Unit of the y-axis values (character).
\item \code{dataType}: Specifies data type—either \code{observed} or \code{simulated}.
\item Optional:
\itemize{
\item \code{yErrorType}: Type of y error, relative to \code{ospsuite::DataErrorType}.
\item \code{yErrorValues}: Numeric error values.
\item \code{yMin}, \code{yMax}: Custom ranges for y-axis instead of error types.
\item \code{IndividualId}: Used for aggregation of simulated population data.
}
}}

\item{metaData}{A list containing metadata for the plot. If NULL, a default
list is constructed from the data. Expected structure includes information
about dimensions and units for both x and y axes.}

\item{mapping}{A ggplot2 aesthetic mapping object. Default is
\code{ggplot2::aes()}. This is added or replaces the default mapping constructed
by the data.}

\item{residualScale}{Either "linear", "log", or "ratio" method for computing
residuals. Default is \code{log}.}

\item{...}{Additional arguments passed to \code{ospsuite.plots::plotQQ}.}
}
\value{
A \code{ggplot2} plot object representing the Q-Q plot.
}
\description{
Plots a Quantile-Quantile plot, grouped by "group".

This function visualizes the distribution of predicted vs observed values
using a Q-Q plot.
}
\examples{
\dontrun{
# Generate a Q-Q plot with default settings
plotQuantileQuantilePlot(myDataCombined)

# Generate a Q-Q plot with linear scale
plotQuantileQuantilePlot(myDataCombined, residualScale = "linear")
}
}
\seealso{
\code{\link[=plotResidualsVsCovariate]{plotResidualsVsCovariate()}}, \code{\link[=plotResidualsAsHistogram]{plotResidualsAsHistogram()}},
\code{\link[=plotPredictedVsObserved]{plotPredictedVsObserved()}}

Other plot functions based on ospsuite.plots: 
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsVsCovariate}()},
\code{\link{plotTimeProfile}()}
}
\concept{plot functions based on ospsuite.plots}
