% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-with-ospsuite-plots.R
\name{plotResidualsAsHistogram}
\alias{plotResidualsAsHistogram}
\title{Plot Residuals Histogram}
\usage{
plotResidualsAsHistogram(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  yUnit = NULL,
  distribution = "normal",
  residualScale = "log",
  ...
)
}
\arguments{
\item{plotData}{An object of class \code{DataCombined} or a \code{data.table}. If a
\code{data.table}, it must include the following:
\itemize{
\item \code{xValues}: Numeric time points.
\item \code{yValues}: Observed or simulated values (numeric).
\item \code{group}: Grouping variable (factor or character).
\item \code{name}: Name for the dataset (factor or character).
\item \code{xUnit}: Unit of the x-axis values (character).
\item \code{yUnit}: Unit of the y-axis values (character).
\item \code{dataType}: Specifies data type—either \code{observed} or \code{simulated}.
\item Optional:
\itemize{
\item \code{yErrorType}: Type of y error, relative to \code{ospsuite::DataErrorType}.
\item \code{yErrorValues}: Numeric error values.
\item \code{yMin}, \code{yMax}: Custom ranges for y-axis instead of error types.
\item \code{IndividualId}: Used for aggregation of simulated population data.
}
}}

\item{metaData}{A list containing metadata for the plot. If NULL, a default
list is constructed from the data. Expected structure includes information
about dimensions and units for both x and y axes.}

\item{mapping}{A ggplot2 aesthetic mapping object. Default is
\code{ggplot2::aes()}. This is added or replaces the default mapping constructed
by the data.}

\item{yUnit}{A character string specifying the target unit for the simulated and
observed y-values used for residual calculation.
If \code{NULL} (default), the most frequent unit in the data is used.
For available units, see \code{ospsuite::ospUnits}.}

\item{distribution}{parameter passed to \code{ospsuite.plots::plotHistogram}.}

\item{residualScale}{Either "linear", "log", or "ratio" method for computing
residuals. Default is \code{log}.}

\item{...}{
  Arguments passed on to \code{\link[ospsuite.plots:plotHistogram]{ospsuite.plots::plotHistogram}}
  \describe{
    \item{\code{geomHistAttributes}}{A \code{list} of arguments passed to \code{ggplot2::geom_histogram} (or \code{geom_bar} if \code{asBarPlot} = TRUE).}
    \item{\code{xScale}}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}
    \item{\code{xScaleArgs}}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}
    \item{\code{yScale}}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}
    \item{\code{yScaleArgs}}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}
    \item{\code{meanFunction}}{Function selection for the display of a vertical line. Options: \code{'none'}, \code{'mean'}, \code{'geomean'}, \code{'median'}, \code{'auto'} (default).
\code{'auto'} selects \code{'mean'} for normal distribution, \code{'geomean'} for lognormal, \code{'median'} for other distributions, and \code{'none'} when no distribution fit.}
  }}
}
\value{
A \code{ggplot2} plot object representing the histogram of residuals.
}
\description{
Plots residuals as a histogram, grouped by "group".

This function generates a histogram of the residuals, providing a visual
representation of their distribution.
}
\examples{
\dontrun{
# Generate a histogram of residuals with default settings
plotResidualsAsHistogram(myDataCombined)

# Generate a histogram with linear scale
plotResidualsAsHistogram(myDataCombined, residualScale = "linear")
}
}
\seealso{
Other plot functions based on ospsuite.plots: 
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotQuantileQuantilePlot}()},
\code{\link{plotResidualsVsCovariate}()},
\code{\link{plotTimeProfile}()}
}
\concept{plot functions based on ospsuite.plots}
