% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residuals-vs-simulated.R
\name{plotResidualsVsSimulated}
\alias{plotResidualsVsSimulated}
\title{Residuals versus time scatter plot}
\usage{
plotResidualsVsSimulated(dataCombined, defaultPlotConfiguration = NULL)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}
}
\description{
Residuals versus time scatter plot
}
\examples{

# TODO: add example

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotPopulationTimeProfile}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
