% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residuals-vs-time.R
\name{plotResidualsVsTime}
\alias{plotResidualsVsTime}
\title{Residuals versus time scatter plot}
\usage{
plotResidualsVsTime(
  dataCombined,
  defaultPlotConfiguration = NULL,
  scaling = "lin"
)
}
\arguments{
\item{dataCombined}{A single instance of \code{DataCombined} class containing both
observed and simulated datasets to be compared.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{scaling}{A character of length one specifying the scale type for residual. can be lin or log.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
# simulated data
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)
simResults <- runSimulations(sim)[[1]]
outputPath <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"

# observed data
obsData <- lapply(
  c("ObsDataAciclovir_1.pkml", "ObsDataAciclovir_2.pkml", "ObsDataAciclovir_3.pkml"),
  function(x) loadDataSetFromPKML(system.file("extdata", x, package = "ospsuite"))
)
names(obsData) <- lapply(obsData, function(x) x$name)


# Create a new instance of `DataCombined` class
myDataCombined <- DataCombined$new()

# Add simulated results
myDataCombined$addSimulationResults(
  simulationResults = simResults,
  quantitiesOrPaths = outputPath,
  groups = "Aciclovir PVB"
)

# Add observed data set
myDataCombined$addDataSets(obsData$`Vergin 1995.Iv`, groups = "Aciclovir PVB")

# Create a new instance of `DefaultPlotConfiguration` class
myPlotConfiguration <- DefaultPlotConfiguration$new()
myPlotConfiguration$title <- "My Plot Title"
myPlotConfiguration$subtitle <- "My Plot Subtitle"
myPlotConfiguration$caption <- "My Sources"

# plot
plotResidualsVsTime(myDataCombined, scaling = "lin", defaultPlotConfiguration = myPlotConfiguration)
}

}
\seealso{
Other plotting:
\code{\link{DefaultPlotConfiguration}},
\code{\link[=plotIndividualTimeProfile]{plotIndividualTimeProfile()}},
\code{\link[=plotObservedVsSimulated]{plotObservedVsSimulated()}},
\code{\link[=plotPopulationTimeProfile]{plotPopulationTimeProfile()}},
\code{\link[=plotResidualsVsSimulated]{plotResidualsVsSimulated()}}
}
\concept{plotting}
