% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residuals-vs-time.R,
%   R/plot-with-ospsuite-plots.R
\name{plotResidualsVsTime}
\alias{plotResidualsVsTime}
\title{Residuals versus time scatter plot}
\usage{
plotResidualsVsTime(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  residualScale = "log",
  ...
)

plotResidualsVsTime(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  residualScale = "log",
  ...
)
}
\arguments{
\item{plotData}{An object of class \code{DataCombined} or a \code{data.table}.
If a \code{data.table}, it must include the following:
\itemize{
\item \code{xValues}: Numeric time points.
\item \code{yValues}: Observed or simulated values (numeric).
\item \code{group}: Grouping variable (factor or character).
\item \code{name}: Name for the dataset (factor or character).
\item \code{xUnit}: Unit of the x-axis values (character).
\item \code{yUnit}: Unit of the y-axis values (character).
\item \code{dataType}: Specifies data type—either \code{observed} or \code{simulated}.
\item Optional:
\itemize{
\item \code{yErrorType}: Type of y error, relative to \code{ospsuite::DataErrorType}.
\item \code{yErrorValues}: Numeric error values.
\item \code{yMin}, \code{yMax}: Custom ranges for y-axis instead of error types.
\item \code{IndividualId}: Used for aggregation pf simulated population data.
}
}}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{mapping}{A ggplot2 aesthetic mapping object. Default is \code{ggplot2::aes()}. This is added or replaces the default mapping constructed by the data.}

\item{residualScale}{A character string specifying the scale for the x and y-axis. Default is 'log'.}

\item{...}{Additional arguments passed to \code{ospsuite.plots::plotResVsCov}.}

\item{scaling}{A character of length one specifying the scale type for residual. can be lin or log.}
}
\value{
A \code{ggplot2} plot object representing residuals vs time.
}
\description{
This function visualizes the residuals over time, providing insights into the accuracy of the model predictions.
}
\examples{
# simulated data
simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
sim <- loadSimulation(simFilePath)
simResults <- runSimulations(sim)[[1]]
outputPath <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"

# observed data
obsData <- lapply(
  c("ObsDataAciclovir_1.pkml", "ObsDataAciclovir_2.pkml", "ObsDataAciclovir_3.pkml"),
  function(x) loadDataSetFromPKML(system.file("extdata", x, package = "ospsuite"))
)
names(obsData) <- lapply(obsData, function(x) x$name)


# Create a new instance of `DataCombined` class
myDataCombined <- DataCombined$new()

# Add simulated results
myDataCombined$addSimulationResults(
  simulationResults = simResults,
  quantitiesOrPaths = outputPath,
  groups = "Aciclovir PVB"
)

# Add observed data set
myDataCombined$addDataSets(obsData$`Vergin 1995.Iv`, groups = "Aciclovir PVB")

# Create a new instance of `DefaultPlotConfiguration` class
myPlotConfiguration <- DefaultPlotConfiguration$new()
myPlotConfiguration$title <- "My Plot Title"
myPlotConfiguration$subtitle <- "My Plot Subtitle"
myPlotConfiguration$caption <- "My Sources"

# plot
plotResidualsVsTime(myDataCombined, scaling = "lin", defaultPlotConfiguration = myPlotConfiguration)

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotObservedVsSimulated}()},
\code{\link{plotPopulationTimeProfile}()},
\code{\link{plotResidualsVsSimulated}()}

Other plot functions based on ospsuite.plots: 
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotQuantileQuantilePlot}()},
\code{\link{plotResidualsVsObserved}()},
\code{\link{plotTimeProfile}()}
}
\concept{plot functions based on ospsuite.plots}
\concept{plotting}
