% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residuals-vs-time.R
\name{plotResidualsVsTime}
\alias{plotResidualsVsTime}
\title{Residuals versus time scatter plot}
\usage{
plotResidualsVsTime(
  dataCombined,
  defaultPlotConfiguration = NULL,
  smoother = NULL
)
}
\arguments{
\item{dataCombined}{A \code{DataCombined} object.}

\item{defaultPlotConfiguration}{A \code{DefaultPlotConfiguration} object, which is
an \code{R6} class object that defines plot properties.}

\item{smoother}{Optional name of smoother function:
\itemize{
\item \code{"loess"} for loess regression
\item \code{"lm"} for linear regression
}}
}
\description{
Residuals versus time scatter plot
}
\examples{

# TODO: add example

}
\seealso{
Other plotting: 
\code{\link{DefaultPlotConfiguration}},
\code{\link{plotIndividualTimeProfile}()},
\code{\link{plotPopulationTimeProfile}()}
}
\concept{plotting}
