% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{Population}
\alias{Population}
\title{Population}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Population
}
\description{
List of individuals used in a population simulation
}
\section{Fields}{

\describe{
\item{\code{count}}{the number of individual in the population}
}}

\section{Methods}{

\describe{
  \count{containerType}{Type of container}
  }
}

\keyword{datasets}
