% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{Population}
\alias{Population}
\title{Population}
\usage{
Population
}
\description{
List of individuals used in a population simulation
}
\section{Fields}{

\describe{
\item{\code{count}}{the number of individual in the population}

\item{\code{allCovariateNames}}{the names of all covariates defined in the population}
}}

\section{Methods}{

\describe{
  \item{has(parameterPath)}{Returns \code{TRUE} if the population has variability defined for \code{parameterPath} otherwise \code{FALSE}}
  \item{setValues(parameterPath, values)}{Adds or update the \code{values} for \code{parameterPath}}
  \item{getValues(parameterPath)}{Returns the variability values defined in the population for \code{parameterPath}}
  \item{getCovariateValues(covariateName)}{Returns the values defined in the population for \code{covariateName}}
  \item{getCovariateValue(covariateName, individualId)}{Returns the covariate value defined in the population for the covariate \code{covariateName} and individual with id\code{individualId}}
  }
}

\keyword{datasets}
