% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-characteristics.R
\docType{class}
\name{PopulationCharacteristics}
\alias{PopulationCharacteristics}
\title{PopulationCharacteristics}
\description{
Characteristics of a population used for population creation
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{PopulationCharacteristics}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{numberOfIndividuals}}{Number of individuals in the population}

\item{\code{proportionOfFemales}}{Proportion of female in the population}

\item{\code{species}}{Specifies the species of the individual. It should be a species available in PK-Sim (see \code{Species})}

\item{\code{population}}{For a Human species, the population of interest. It should be a population available in PK-Sim (see \code{HumanPopulation})}

\item{\code{age}}{Age range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{gestationalAge}}{Gestational Age range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{weight}}{Weight range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{height}}{Height range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{BMI}}{BMI range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{allMoleculeOntogenies}}{All molecule ontogenies defined for this individual characteristics.}

\item{\code{seed}}{Seed used to generate the population}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationCharacteristics-new}{\code{PopulationCharacteristics$new()}}
\item \href{#method-PopulationCharacteristics-print}{\code{PopulationCharacteristics$print()}}
\item \href{#method-PopulationCharacteristics-addMoleculeOntogeny}{\code{PopulationCharacteristics$addMoleculeOntogeny()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationCharacteristics-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationCharacteristics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{PopulationCharacteristics} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationCharacteristics-print"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationCharacteristics-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationCharacteristics-addMoleculeOntogeny"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationCharacteristics-addMoleculeOntogeny}{}}}
\subsection{Method \code{addMoleculeOntogeny()}}{
Add a molecule ontogeny \code{MoleculeOntogeny} to the individual characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$addMoleculeOntogeny(moleculeOntogeny)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moleculeOntogeny}}{Molecule ontogeny to add}
}
\if{html}{\out{</div>}}
}
}
}
