% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-characteristics.R
\docType{class}
\name{PopulationCharacteristics}
\alias{PopulationCharacteristics}
\title{PopulationCharacteristics}
\description{
Characteristics of a population used for population creation
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{PopulationCharacteristics}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{numberOfIndividuals}}{Number of individuals in the population}

\item{\code{proportionOfFemales}}{Proportion of female in the population}

\item{\code{species}}{Specifies the species of the individual. It should be a species available in PK-Sim (see \code{Species})}

\item{\code{population}}{For a Human species, the population of intereset. It should be a population available in PK-Sim (see \code{HumanPopulation})}

\item{\code{age}}{Age range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{gestationalAge}}{Gestational Age range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{weight}}{Weight range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{height}}{Height range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{BMI}}{BMI range of the population as in instance of a \code{ParameterRange} (optional)}

\item{\code{allMoleculeOntogenies}}{All molecule ontogenies defined for this individual characteristics.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationCharacteristics$new()}}
\item \href{#method-print}{\code{PopulationCharacteristics$print()}}
\item \href{#method-addMoleculeOntogeny}{\code{PopulationCharacteristics$addMoleculeOntogeny()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{PopulationCharacteristics} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMoleculeOntogeny"></a>}}
\if{latex}{\out{\hypertarget{method-addMoleculeOntogeny}{}}}
\subsection{Method \code{addMoleculeOntogeny()}}{
Add a molecule ontogeny \code{MoleculeOntogeny} to the individual characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationCharacteristics$addMoleculeOntogeny(moleculeOntogeny)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moleculeOntogeny}}{Molecule ontogeny to add}
}
\if{html}{\out{</div>}}
}
}
}
