% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity.R
\docType{class}
\name{Quantity}
\alias{Quantity}
\title{Quantity}
\usage{
Quantity
}
\description{
A quantity of the model (with unit, value) such as a Parameter or an Amount
}
\section{Fields}{

\describe{
\item{\code{value}}{The value of the quantity in unit}

\item{\code{unit}}{The base unit in which the quantity is defined (Read-Only)}

\item{\code{dimension}}{The dimension in which the quantity is defined  (Read-Only)}
}}

\section{Methods}{

\describe{
  \item{setValue(value, unit=NULL)}{Convert value from unit to the base unit and sets the value in base unit. If unit is null, we assume that the value is in base unit}
  \item{hasUnit(unit)}{Returns \code{TRUE} if the quantity supports the given unit otherwise \code{FALSE}. For the list of supported units, use \code{allUnits}}
  \item{allUnits()}{Returns the list of all supported units}
  }
}

\keyword{datasets}
