% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity.R
\docType{class}
\name{Quantity}
\alias{Quantity}
\title{Quantity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Quantity
}
\description{
A quantity of the model (with unit, value) such as a Parameter or an Amount
}
\section{Fields}{

\describe{
\item{\code{value}}{The value of the quantity}
}}

\keyword{datasets}
