% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity.R
\docType{class}
\name{Quantity}
\alias{Quantity}
\title{Quantity}
\usage{
Quantity
}
\description{
A quantity of the model (with unit, value) such as a Parameter or an Amount
}
\section{Fields}{

\describe{
\item{\code{value}}{The value of the quantity in unit}

\item{\code{unit}}{The base unit in which the quantity is defined (Read-Only)}

\item{\code{dimension}}{The dimension in which the quantity is defined  (Read-Only)}
}}

\keyword{datasets}
