% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-pk-parameter.R
\docType{class}
\name{QuantityPKParameter}
\alias{QuantityPKParameter}
\title{QuantityPKParameter}
\description{
pK-Parameter values for all individuals of a simulation (1 or more) calculated for a specific quantity with path \code{quantityPath}
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{QuantityPKParameter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{values}}{All values for \code{quantityPath} and \code{name}}

\item{\code{quantityPath}}{The path of the quantity for which the values were calculated}

\item{\code{name}}{The name of the pK-Parameter (AUC, Cmax, Tmax etc...)}

\item{\code{unit}}{Base unit in which the pk parameter was calculated}

\item{\code{dimension}}{Dimension in which the pk parameter was calculated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantityPKParameter-new}{\code{QuantityPKParameter$new()}}
\item \href{#method-QuantityPKParameter-print}{\code{QuantityPKParameter$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantityPKParameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantityPKParameter-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantityPKParameter$new(netObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{netObject}}{An \code{rSharp::NetObject} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{QuantityPKParameter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantityPKParameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantityPKParameter-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantityPKParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
