% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-selection.R
\docType{class}
\name{QuantitySelection}
\alias{QuantitySelection}
\title{QuantitySelection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QuantitySelection
}
\description{
List of selected quantities selected as output for a given simulation
}
\keyword{datasets}
