% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runSensitivityAnalysis}
\alias{runSensitivityAnalysis}
\title{Runs a sensitivity analaysis}
\usage{
runSensitivityAnalysis(sensitivityAnalysis,
  sensitivityAnalysisRunOptions = NULL)
}
\arguments{
\item{sensitivityAnalysis}{Instance of a \code{SensitivityAnalysis} to run}

\item{sensitivityAnalysisRunOptions}{Optional instance of a \code{SensitivityAnalysisRunOptions} used during the sensitivity analysis run}
}
\value{
SimulationResults (one entry per Individual)
}
\description{
Runs a sensitivity analaysis
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Create a new sensitivity object for the simulation
sensitivity <- SensitivityAnalysis$new(sim)

# Runs the sensitivity analysis
results <- runSensitivityAnalysis(sensitivity)
}
