% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run a simulation from a pkml file and returns the simulation}
\usage{
runSimulation(simulation)
}
\arguments{
\item{simulation}{Instance of a simulation to simulate.}
}
\value{
List of individual results (one entry per Individual)
}
\description{
Run a simulation from a pkml file and returns the simulation
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
results <- runSimulation(sim)
}
