% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run a single simulation}
\usage{
runSimulation(
  simulation,
  population = NULL,
  agingData = NULL,
  simulationRunOptions = NULL
)
}
\arguments{
\item{simulation}{One \code{Simulation} to simulate.}

\item{population}{Optional instance of a \code{Population} to use for the simulation. This is only used when simulating one simulation
Alternatively, you can also pass the result of \code{createPopulation} directly. In this case, the population will be extracted}

\item{agingData}{Optional instance of \code{AgingData} to use for the simulation. This is only used with a population simulation}

\item{simulationRunOptions}{Optional instance of a \code{SimulationRunOptions} used during the simulation run}
}
\value{
SimulationResults (one entry per Individual) for a single simulation
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Runs one simulation (individual or population) and returns a
\code{SimulationResults} object containing all results of the simulation.
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Running an individual simulation
# results is an instance of `SimulationResults`
results <- runSimulations(sim)[[1]]

# Creating custom simulation run options

simRunOptions <- SimulationRunOptions$new()
simRunOptions$numberOfCores <- 3
simRunOptions$showProgress <- TRUE

# Running a population simulation
popPath <- system.file("extdata", "pop.csv", package = "ospsuite")
population <- loadPopulation(popPath)
results <- runSimulation(sim, population, simulationRunOptions = simRunOptions)
}
\keyword{internal}
