% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runSimulations}
\alias{runSimulations}
\title{Runs multiple simulations concurrently.}
\usage{
runSimulations(
  simulations,
  population = NULL,
  agingData = NULL,
  simulationRunOptions = NULL,
  silentMode = FALSE,
  stopIfFails = FALSE
)
}
\arguments{
\item{simulations}{One \code{Simulation} or list of \code{Simulation} objects
to simulate.}

\item{population}{Optional instance of a \code{Population} to use for the simulation.
Only allowed when simulating one simulation.
Alternatively, you can also pass the result of \code{createPopulation} directly.
In this case, the population will be extracted.}

\item{agingData}{Optional instance of \code{AgingData} to use for the simulation.
This is only used with a population simulation}

\item{simulationRunOptions}{Optional instance of a \code{SimulationRunOptions} used during the simulation run}

\item{silentMode}{If \code{TRUE}, no warnings are displayed if a simulation fails. Default is \code{FALSE}}

\item{stopIfFails}{Wether to stop the execution if one of the simulation did not succeed. default to FALSE.}
}
\value{
A named list of \code{SimulationResults} objects with names being the IDs
of the respective simulations. If a simulation fails, the result for this
simulation is \code{NULL}
}
\description{
Runs multiple simulations concurrently.
}
\details{
For multiple simulations, only individual simulations are possible.
For single simulation, either individual or population simulations can be
performed.
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Running an individual simulation
# Results is a list with one object `SimulationResults`
results <- runSimulations(sim)

# Creating custom simulation run options

simRunOptions <- SimulationRunOptions$new()
simRunOptions$numberOfCores <- 3
simRunOptions$showProgress <- TRUE

# Running a population simulation
popPath <- system.file("extdata", "pop.csv", package = "ospsuite")
population <- loadPopulation(popPath)
results <- runSimulations(sim, population, simulationRunOptions = simRunOptions)

# Running multiple simulations in parallel
sim2 <- loadSimulation(simPath)
sim3 <- loadSimulation(simPath)

# Results is an array of `SimulationResults`
results <- runSimulations(list(sim, sim2, sim3))
}
