% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{runSimulationsFromSnapshot}
\alias{runSimulationsFromSnapshot}
\title{Run Simulations From Snapshot Files}
\usage{
runSimulationsFromSnapshot(
  ...,
  output = ".",
  csv = TRUE,
  pkml = FALSE,
  xml = FALSE
)
}
\arguments{
\item{...}{character strings, path to snapshot files or a directory containing snapshot files}

\item{output}{character string, path to the output directory where to write simulation results}

\item{csv}{logical, whether to export the results as csv (default = TRUE)}

\item{pkml}{logical, whether to export the results as pkml (default = FALSE)}

\item{xml}{logical, whether to export the results as xml (default = FALSE)}
}
\description{
Run Simulations From Snapshot Files
}
