% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{runSimulationsFromSnapshot}
\alias{runSimulationsFromSnapshot}
\title{Run Simulations From Snapshot Files}
\usage{
runSimulationsFromSnapshot(
  ...,
  output = ".",
  RunForAllOutputs = FALSE,
  exportCSV = TRUE,
  exportPKML = FALSE,
  exportJSON = FALSE,
  exportXML = FALSE
)
}
\arguments{
\item{...}{character strings, path to snapshot files or a directory containing snapshot files}

\item{output}{character string, path to the output directory where to write simulation results}

\item{RunForAllOutputs}{logical, whether to run the simulation for all outputs or only OutputSelections (default = FALSE)}

\item{exportCSV}{logical, whether to export the results as csv (default = TRUE)}

\item{exportPKML}{logical, whether to export the results as pkml (default = FALSE)}

\item{exportJSON}{logical, whether to export simulation results as json (default = FALSE)}

\item{exportXML}{logical, whether to export the results as xml (default = FALSE)}
}
\description{
Run Simulations From Snapshot Files
}
\examples{
\dontrun{
runSimulationsFromSnapshot("path/to/my_snapshot.json", csv = TRUE, pkml = TRUE)
}
}
