% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\docType{class}
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{SensitivityAnalysis}
\usage{
SensitivityAnalysis
}
\description{
Supports Sensitivity Analysis workflow to assess the impact of input parameters on the simulation outputs
}
\section{Fields}{

\describe{
\item{\code{simulation}}{Simulation for which the sensitivity analysis should be performed}

\item{\code{parameters}}{List of parameters to use for sensitivity calculation (optional). If undefined, the sensitivity will be performed automatically
on all constant parameters of the simulation. Constant parameter means all parameters with a constant value or a formula parameter
with a value that was overriden by the user}

\item{\code{numberOfSteps}}{Number of steps used for the variation of each parameter (optional, default specified in \code{ospsuiteEnv$sensitivityAnalysisConfig})}

\item{\code{variationRange}}{Variation applied to the parameter (optional, default specified in \code{ospsuiteEnv$sensitivityAnalysisConfig})}
}}

\keyword{datasets}
