% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\docType{class}
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{SensitivityAnalysis}
\usage{
SensitivityAnalysis
}
\description{
Supports Sensitivity Analysis workflow to know which input parameters have most impact on the simulation output
}
\section{Fields}{

\describe{
\item{\code{simulation}}{Simulation for which the sensitivity analysis should be performed}

\item{\code{parameters}}{list of parameters to use for sensitivity calculation (optional)}

\item{\code{numberOfSteps}}{Number of steps used for the variation of each parameter (optional, default is 4)}

\item{\code{variationRange}}{Variation applied to the parameter (optionak, default is 0.1)}
}}

\section{Methods}{

\describe{
  \item{has(parameterPath)}{Returns \code{TRUE} if the population has variability defined for \code{parameterPath} otherwise \code{FALSE}}
  }
}

\keyword{datasets}
