% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis-results.R
\docType{class}
\name{SensitivityAnalysisResults}
\alias{SensitivityAnalysisResults}
\title{SensitivityAnalysisResults}
\usage{
SensitivityAnalysisResults
}
\description{
Results of a sensitivity analysis run (either individual or population simulation)
}
\section{Methods}{

\describe{
  \item{allPKParameterSensitivitiesFor(pkParameterName, outputPath, totalSensitivityThreshold)}
  {Returns the PKParameterSensitivity for a given pkParameter and output participating to a total sensitivity greater or equal to totalSensitivityThreshold}
}
}

\keyword{datasets}
