% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis-run-options.R
\docType{class}
\name{SensitivityAnalysisRunOptions}
\alias{SensitivityAnalysisRunOptions}
\title{SensitivityAnalysisRunOptions}
\usage{
SensitivityAnalysisRunOptions
}
\description{
Options to be passed to the sensitivity analysis engine
}
\section{Fields}{

\describe{
\item{\code{numberOfCoresToUse}}{(Maximal) number of cores to be used. Per default the total number of cores availble to the system minus 1 is taken.}

\item{\code{showProgress}}{Specifies whether progress bar should be shown during sensitivity analysis run. Default is \code{FALSE}}
}}

\keyword{datasets}
