% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{setMoleculeValuesByPath}
\alias{setMoleculeValuesByPath}
\title{Set molecule start values in the simulation by path}
\usage{
setMoleculeValuesByPath(
  moleculePaths,
  values,
  simulation,
  units = NULL,
  stopIfNotFound = TRUE
)
}
\arguments{
\item{moleculePaths}{A single or a list of molecule paths}

\item{values}{A numeric value that should be assigned to the molecule start value or a vector
of numeric values, if the start value of more than one molecule should be changed. Must have the same
length as \code{moleculePaths}}

\item{simulation}{Simulation containing the quantities}

\item{units}{A string or a list of strings defining the units of the
\code{values}. If \code{NULL} (default), values are assumed to be in base units. If
not \code{NULL}, must have the same length as \code{quantityPaths}.}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no quantity exists for
the given path, an error is thrown. If \code{FALSE}, a warning is shown to the
user.}
}
\description{
Set molecule start values in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
setMoleculeValuesByPath("Organism|Liver|A", 1, sim)

setMoleculeValuesByPath(
  c("Organism|Liver|A", "Organism|Liver|B"),
  c(2, 3),
  sim,
  units = c("µmol", "mmol")
)
}
