% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule.R
\name{setMoleculeValuesByPath}
\alias{setMoleculeValuesByPath}
\title{Set molecule start values in the simulation by path}
\usage{
setMoleculeValuesByPath(
  moleculePaths,
  values,
  simulation,
  stopIfNotFound = TRUE,
  units
)
}
\arguments{
\item{moleculePaths}{A single or a list of molecule paths}

\item{values}{A numeric value that should be assigned to the molecule start value or a vector
of numeric values, if the start value of more than one molecule should be changed. Must have the same
length as \code{moleculePaths}}

\item{simulation}{Simulation uses to retrieve quantity instances from given paths.}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default) and no quantity exists for the given path,
an error is thrown. If \code{FALSE}, a warning is shown to the user}

\item{units}{A string or a list of strings defining the units of the \code{values}. If \code{NULL} (default), values
are assumed to be in base units. If not \code{NULL}, must have the same length as \code{quantityPaths}.}
}
\description{
Set molecule start values in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
setMoleculeValuesByPath("Organism|VenousBlood|Plasma|CYP3A4", 1, sim)

setMoleculeValuesByPath(c("Organism|VenousBlood|Plasma|CYP3A4", "Organism|ArterialBlood|Plasma|CYP3A4"), c(2, 3), sim, units = c("µmol", "mmol"))
}
