% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-parameter.R
\name{setParametersValues}
\alias{setParametersValues}
\title{Set values of parameters}
\usage{
setParametersValues(parameters, values)
}
\arguments{
\item{parameters}{A single object of type 'Parameter' or a list of such objects}

\item{values}{A numeric value that should be assigned to the parameter or a vector
of numeric values, if the value of more than one parameter should be changed. Must have the same
length as 'parameters'}
}
\description{
Set values of parameters
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getParameter("Organism|Liver|Volume", sim)
setParametersValues(param, 1)
params <- getAllParametersMatching("Organism|**|Volume", sim)
setParametersValues(params, c(2, 3))
}
\seealso{
\code{\link{getParameter}} and \code{\link{getAllParametersMatching}} to create objects of type Parameter
}
