% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-building-block.R,
%   R/utilities-parameter.R
\name{setParameterValues}
\alias{setParameterValues}
\title{Set or add parameter values to an existing Parameter Values building block.}
\usage{
setParameterValues(parameters, values, units = NULL)

setParameterValues(parameters, values, units = NULL)
}
\arguments{
\item{parameters}{A single or a list of \code{Parameter}}

\item{values}{A numeric value that should be assigned to the parameter or a vector
of numeric values, if the value of more than one parameter should be changed. Must have the same
length as 'parameters'. Alternatively, the value can be a unique number. In that case, the same value will be set in all parameters}

\item{units}{A string or a list of strings defining the units of the
\code{values}. If \code{NULL} (default), values are assumed to be in base units. If
not \code{NULL}, must have the same length as \code{quantities}.}

\item{parameterValuesBuildingBlock}{A \code{BuildingBlock} object of type \verb{Parameter Values}.
The entries will be added to or set in this building block.}

\item{quantityPaths}{A list of full paths of the quantities (usually parameters). Should contain
all path elements and the parameter name, separated by \code{|}.}

\item{formulas}{A list of \code{Formula} objects that will be set for the quantities.
Mixing of \code{formulas} and \code{quantityValues} is not allowed. If \code{formulas} is provided,
\code{quantityValues} should be \code{NULL}. The length of this list should be equal to the length of \code{quantityPaths}.}

\item{dimensions}{A single dimension or a list of dimensions (string names)
of parameter values. Supported dimensions are listed in \code{ospDimension}. By default,
new entries get the \code{Dimensionless} dimension.}

\item{quantityValues}{A list of values for the quantities. Should be \code{NULL} if
the argument \code{formulas} is provided. If not \code{NULL}, the length of this list should be equal to the length of \code{quantityPaths}.}
}
\description{
This functions allows adding or modifying parameter values entries.
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
param <- getParameter("Organism|Liver|Volume", sim)
setParameterValues(param, 1)
params <- getAllParametersMatching("Organism|**|Volume", sim)
setParameterValues(params, c(2, 3), units = c("ml", "l"))
}
\seealso{
\code{\link[=getParameter]{getParameter()}} and \code{\link[=getAllParametersMatching]{getAllParametersMatching()}} to create objects of type Parameter
}
