% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{setQuantityValuesByPath}
\alias{setQuantityValuesByPath}
\title{Set the values of parameters in the simulation by path}
\usage{
setQuantityValuesByPath(quantityPaths, values, simulation)
}
\arguments{
\item{quantityPaths}{A single or a list of absolute quantity path}

\item{values}{A numeric value that should be assigned to the quantities or a vector
of numeric values, if the value of more than one quantity should be changed. Must have the same
length as 'quantityPaths'}

\item{simulation}{Simulation uses to retrieve quantity instances from given paths.}
}
\description{
Set the values of parameters in the simulation by path
}
\examples{

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
setQuantityValuesByPath("Organism|Liver|Volume", 1, sim)

setParameterValuesByPath(list("Organism|Liver|Volume", "Organism|Liver|A"), c(2, 3), sim)
}
