% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Simulation}
\description{
An OSPSuite simulation
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{Simulation}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{root}}{Root container of the simulation (read-only)}

\item{\code{path}}{Path of the root container of the simulation (read-only)}

\item{\code{settings}}{SimulationSettings object for the simulation (read-only)}

\item{\code{solver}}{SimulationSolver object for the simulation (read-only)}

\item{\code{outputSchema}}{outputSchema object for the simulation (read-only)}

\item{\code{outputSelections}}{outputSelections object for the simulation (read-only)}

\item{\code{sourceFile}}{Path to the file the simulation was loaded from (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Simulation$new()}}
\item \href{#method-allStationaryMoleculeNames}{\code{Simulation$allStationaryMoleculeNames()}}
\item \href{#method-allFloatingMoleculeNames}{\code{Simulation$allFloatingMoleculeNames()}}
\item \href{#method-print}{\code{Simulation$print()}}
\item \href{#method-clone}{\code{Simulation$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(ref, sourceFile = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Reference to .NET simulation object}

\item{\code{sourceFile}}{(Optional) File used to load the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Simulation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allStationaryMoleculeNames"></a>}}
\subsection{Method \code{allStationaryMoleculeNames()}}{
Returns the name of all stationary molecules defined in the simulation.
This is a typically a molecule that is individual specific such as en Enzyme, Protein, Transporter, FcRn etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allStationaryMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allFloatingMoleculeNames"></a>}}
\subsection{Method \code{allFloatingMoleculeNames()}}{
Returns the name of all floating molecules defined in the simulation.
This is typically a molecule that is being simulated such as Compound, Inhibitor, DrugComplex.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allFloatingMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
