% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Simulation}
\description{
An OSPSuite simulation
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:ObjectBase]{ospsuite::ObjectBase}} -> \code{Simulation}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{root}}{Root container of the simulation (read-only)}

\item{\code{path}}{Path of the root container of the simulation (read-only)}

\item{\code{solver}}{SimulationSolver object for the simulation (read-only)}

\item{\code{outputSchema}}{outputSchema object for the simulation (read-only)}

\item{\code{outputSelections}}{outputSelections object for the simulation (read-only)}

\item{\code{sourceFile}}{Path to the file the simulation was loaded from (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulation-new}{\code{Simulation$new()}}
\item \href{#method-Simulation-allEndogenousStationaryMoleculeNames}{\code{Simulation$allEndogenousStationaryMoleculeNames()}}
\item \href{#method-Simulation-allXenobioticFloatingMoleculeNames}{\code{Simulation$allXenobioticFloatingMoleculeNames()}}
\item \href{#method-Simulation-allStationaryMoleculeNames}{\code{Simulation$allStationaryMoleculeNames()}}
\item \href{#method-Simulation-allFloatingMoleculeNames}{\code{Simulation$allFloatingMoleculeNames()}}
\item \href{#method-Simulation-molWeightFor}{\code{Simulation$molWeightFor()}}
\item \href{#method-Simulation-allApplicationsFor}{\code{Simulation$allApplicationsFor()}}
\item \href{#method-Simulation-print}{\code{Simulation$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(ref, sourceFile = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Reference to .NET simulation object}

\item{\code{sourceFile}}{(Optional) File used to load the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Simulation} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-allEndogenousStationaryMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-allEndogenousStationaryMoleculeNames}{}}}
\subsection{Method \code{allEndogenousStationaryMoleculeNames()}}{
Returns the name of all endogenous stationary molecules defined in the simulation. (e.g. with the flag IsStationary = TRUE)
This is a typically a molecule that is individual specific such as en Enzyme, Protein, Transporter, FcRn etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allEndogenousStationaryMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-allXenobioticFloatingMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-allXenobioticFloatingMoleculeNames}{}}}
\subsection{Method \code{allXenobioticFloatingMoleculeNames()}}{
Returns the name of all xenobiotoc floating molecules defined in the simulation. (e.g. with the flag IsStationary = FALSE)
This is typically a molecule that is being explicitly simulated such as Compound, Inhibitor, DrugComplex.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allXenobioticFloatingMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-allStationaryMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-allStationaryMoleculeNames}{}}}
\subsection{Method \code{allStationaryMoleculeNames()}}{
Returns the name of all stationary molecules defined in the simulation. (e.g. with the flag IsStationary = TRUE)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allStationaryMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-allFloatingMoleculeNames"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-allFloatingMoleculeNames}{}}}
\subsection{Method \code{allFloatingMoleculeNames()}}{
Returns the name of all floating molecules defined in the simulation. (e.g. with the flag IsStationary = FALSE)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allFloatingMoleculeNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-molWeightFor"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-molWeightFor}{}}}
\subsection{Method \code{molWeightFor()}}{
Returns the mol weight value (in core unit) associated to the quantity with given path or NA if not found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$molWeightFor(quantityPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantityPath}}{Path of quantity used to retrieve the molecular weight}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-allApplicationsFor"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-allApplicationsFor}{}}}
\subsection{Method \code{allApplicationsFor()}}{
Returns the applications ordered by start time associated to the quantity with path \code{quantityPath} or an empty list if not found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$allApplicationsFor(quantityPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantityPath}}{Path of quantity used to retrieve the applications (e.g. applications resulting in this quantity being applied)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-print"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
