% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Simulation}
\usage{
Simulation
}
\description{
An OSPSuite simulation
}
\section{Fields}{

\describe{
\item{\code{root}}{Root container of the simulation (read-only)}

\item{\code{path}}{Path of the root container of the simulation (read-only)}

\item{\code{settings}}{SimulationSettings object for the simulation (read-only)}

\item{\code{sourceFile}}{Path to the file the simulation was loaded from (read-only)}
}}

\keyword{datasets}
