% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-batch.R
\docType{class}
\name{SimulationBatch}
\alias{SimulationBatch}
\title{SimulationBatch}
\description{
Options to be passed to the SimulationBatch
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SimulationBatch}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulation}}{Underlying simulation used for the batch run}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimulationBatch$new()}}
\item \href{#method-run}{\code{SimulationBatch$run()}}
\item \href{#method-finalize}{\code{SimulationBatch$finalize()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatch$new(ref, simulation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{.NET reference object.}

\item{\code{simulation}}{Simulation used in the batch run}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SimulationBatch} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Set the parameter and initial values in the simulation and run the simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatch$run(parameterValues = NULL, initialValues = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterValues}}{Vector of parameter values to set in the simulation (default is \code{NULL})}

\item{\code{initialValues}}{Vector of initial values to set in the simulation  (default is \code{NULL})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SimulationResults} object containing the result of the simulation run
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the reference to the wrapped .NET object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatch$finalize()}\if{html}{\out{</div>}}
}

}
}
