% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-batch-options.R
\docType{class}
\name{SimulationBatchOptions}
\alias{SimulationBatchOptions}
\title{SimulationBatchOptions}
\description{
Options to be passed to the \code{SimulationBatch}.
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SimulationBatchOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{variableParameters}}{Vector of absolute parameter paths to be varied in a simulation batch}

\item{\code{variableMolecules}}{Vector of absolute molecule paths to be varied in a simulation batch}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationBatchOptions-new}{\code{SimulationBatchOptions$new()}}
\item \href{#method-SimulationBatchOptions-print}{\code{SimulationBatchOptions$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationBatchOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationBatchOptions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatchOptions$new(variableParameters = NULL, variableMolecules = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variableParameters}}{Vector of absolute parameter paths to be varied in a simulation batch}

\item{\code{variableMolecules}}{Vector of absolute molecule paths to be varied in a simulation batch}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SimulationBatchOptions} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationBatchOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationBatchOptions-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatchOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
