% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-batch-run-values.R
\docType{class}
\name{SimulationBatchRunValues}
\alias{SimulationBatchRunValues}
\title{SimulationBatchRunValues}
\description{
Options to be passed to the SimulationBatch run
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SimulationBatchRunValues}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameterValues}}{Vector of parameter values used in a batch run}

\item{\code{initialValues}}{Vector of initial values used in a batch run}

\item{\code{id}}{Internal id of the batch run value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationBatchRunValues-new}{\code{SimulationBatchRunValues$new()}}
\item \href{#method-SimulationBatchRunValues-print}{\code{SimulationBatchRunValues$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationBatchRunValues-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationBatchRunValues-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatchRunValues$new(parameterValues = NULL, initialValues = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterValues}}{Vector of parameter values}

\item{\code{initialValues}}{Vector of molecule initial values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SimulationBatchRunValues} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationBatchRunValues-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationBatchRunValues-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationBatchRunValues$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
