% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationConfiguration.R
\docType{class}
\name{SimulationConfiguration}
\alias{SimulationConfiguration}
\title{Simulation configuration}
\value{
A new \code{SimulationConfiguration} object.
}
\description{
Configuration of a simulation. Contains description of the modules used for
the simulation, selected Parameter Values (PV) and Initial Conditions (IC), and molecule calculation methods.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{individual}}{A building block of type "Individual" used in the configuration.
Can be \code{NULL} if no Individual should be applied.}

\item{\code{expressionProfiles}}{A list of building blocks of type "Expression Profile"
used in the configuration. Only one profile per protein is allowed.}

\item{\code{modules}}{A named list of \code{MoBiModule} objects from which to create the simulation.
The order of modules defines the order in which the modules will be combined to a simulation!
Replacing \code{modules} does not automatically reset \code{selectedInitialConditions} or
\code{selectedParameterValues}. Reassign those fields explicitly if the new modules
do not contain the previously selected building blocks.}

\item{\code{selectedInitialConditions}}{A named list with names being the names of the modules,
and the values the names of Initial Conditions Building Blocks}

\item{\code{selectedParameterValues}}{A named list with names being the names of the modules,
and the values the names of Parameter Values Building Blocks
If the modules for which the PV
selection is provided are not in the configuration, throw an error.}

\item{\code{partitionCoefficientOverrides}}{A named list of partition coefficient method
overrides. Names are molecule names, values are the full calculation method names.
Read-only; use \code{setPartitionCoefficientMethods()} to modify.}

\item{\code{cellularPermeabilityOverrides}}{A named list of cellular permeability method
overrides. Names are molecule names, values are the full calculation method names.
Read-only; use \code{setCellularPermeabilityMethods()} to modify.}

\item{\code{settings}}{A \code{SimulationSettings} object defining the simulation settings.
If no settings are provided, default settings will be used upon simulation creation.
Individual properties within the \code{SimulationSettings} object are read-only, but the entire settings object can be replaced with another \code{SimulationSettings} instance (or set to \code{NULL} to use defaults).
Setting the solver settings, output intervals, and output selections must be done after creating the simulation from the configuration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationConfiguration-new}{\code{SimulationConfiguration$new()}}
\item \href{#method-SimulationConfiguration-setPartitionCoefficientMethods}{\code{SimulationConfiguration$setPartitionCoefficientMethods()}}
\item \href{#method-SimulationConfiguration-setCellularPermeabilityMethods}{\code{SimulationConfiguration$setCellularPermeabilityMethods()}}
\item \href{#method-SimulationConfiguration-print}{\code{SimulationConfiguration$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$new(
  modules,
  individual = NULL,
  expressionProfiles = NULL,
  selectedInitialConditions = NULL,
  selectedParameterValues = NULL,
  settings = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modules}}{A list of \code{MoBiModule} objects from which to create in simulation.
The order of modules defines the order in which the modules will be combined to a simulation!}

\item{\code{individual}}{Optional, an individual building block}

\item{\code{expressionProfiles}}{Optional, a list of expression profiles to apply to the simulation.}

\item{\code{selectedInitialConditions}}{By default, the first Initial Conditions
(IC) building block (BB) of each module will be selected. If a module has multiple
IC BBs, it is possible to specify which IC BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the IC BB.
By explicitly setting the value for a specific module to \code{NULL}, no IC BB from the specified module will be applied.
Names that do not match any module passed via \code{modules} raise an error.}

\item{\code{selectedParameterValues}}{By default, the first Parameter Values
(PV) building block (BB) of each module will be selected. If a module has multiple
PV BBs, it is possible to specify which PV BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the PV BB.
By explicitly setting the value for a specific module to \code{NULL}, no PV BB from the specified module will be applied.
Names that do not match any module passed via \code{modules} raise an error.}

\item{\code{settings}}{Optional, a \code{SimulationSettings} object defining the simulation settings. If no settings are provided,
default settings will be used upon simulation creation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-setPartitionCoefficientMethods"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-setPartitionCoefficientMethods}{}}}
\subsection{Method \code{setPartitionCoefficientMethods()}}{
Sets the method used for calculation of partition coefficients. This method can be used to specify the partition coefficient method for each molecule used in the modules. If
not specified, the method set in the Molecules Building Block of the last module in which the molecule is present will be used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$setPartitionCoefficientMethods(
  moleculeName,
  methodName
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moleculeName}}{The name of the molecule for which to set the partition coefficient method.}

\item{\code{methodName}}{The name of the method to use for calculation of partition coefficients for the specified molecule. Should be one of the \code{PartitionCoefficientMethods} enum values.
Use \code{NULL} to remove an existing override.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-setCellularPermeabilityMethods"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-setCellularPermeabilityMethods}{}}}
\subsection{Method \code{setCellularPermeabilityMethods()}}{
Sets the method used for calculation of cellular permeabilities. This method can be used to specify the cellular permeability method for each molecule used in the modules. If
not specified, the method set in the Molecules Building Block of the last module in which the molecule is present will be used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$setCellularPermeabilityMethods(
  moleculeName,
  methodName
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moleculeName}}{The name of the molecule for which to set the cellular permeability method.}

\item{\code{methodName}}{The name of the method to use for calculation of cellular permeabilities for the specified molecule. Should be one of the \code{CellularPermeabilityMethods} enum values.
Use \code{NULL} to remove an existing override.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
