% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationConfiguration.R
\docType{class}
\name{SimulationConfiguration}
\alias{SimulationConfiguration}
\title{Simulation configuration}
\value{
A new \code{SimulationConfiguration} object.
}
\description{
Configuration of a simulation. Contains description of the modules used for
the simulation, selected Parameter Values (PV) and Initial Conditions (IC), and molecule calculation methods.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{partitionCoefficientMethods}}{The method used for calculation of partition coefficients. A named list with names being the molecules used in all modules,
and the values being one of the \code{PartitionCoefficientMethods} enum values.
To set the partition coefficient method for a molecule, provide a named list.
TODO https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1650}

\item{\code{cellularPermeabilityMethods}}{The method used for calculation of cellular permeabilities. A named list with names being the molecules used in all modules,
and the values being one of the \code{CellularPermeabilityMethods} enum values.
To set the cellular permeability method for a molecule, provide a named list.
TODO https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1650}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{individual}}{A building block of type "Individual" used in the configuration.
Can be \code{NULL} if no Individual should be applied.}

\item{\code{expressionProfiles}}{A list of building blocks of type "Expression Profile"
used in the configuration. Only one profile per protein is allowed.}

\item{\code{modules}}{A named list of \code{Module} objects from which to create in simulation.
The order of modules defines the order in which the modules will be combined to a simulation!
When setting the modules, the selection of Initial Conditions and Parameter Values
is reset to the first available ones in the modules.}

\item{\code{selectedInitialConditions}}{A named list with names being the names of the modules,
and the values the names of Initial Conditions Building Blocks}

\item{\code{selectedParameterValues}}{A named list with names being the names of the modules,
and the values the names of Parameter Values Building Blocks
If the modules for which the PV
selection is provided are not in the configuration, throw an error.}

\item{\code{settings}}{A \code{SimulationSettings} object defining the simulation settings.
If no settings are provided, default settings will be used upon simulation creation.
Individual properties within the \code{SimulationSettings} object are read-only, but the entire settings object can be replaced with another \code{SimulationSettings} instance (or set to \code{NULL} to use defaults).
Setting the solver settings, output intervals, and output selections must be done after creating the simulation from the configuration.}

\item{\code{partitionCoefficientMethods}}{The method used for calculation of partition coefficients. A named list with names being the molecules used in all modules,
and the values being one of the \code{PartitionCoefficientMethods} enum values.
To set the partition coefficient method for a molecule, provide a named list.
TODO https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1650}

\item{\code{cellularPermeabilityMethods}}{The method used for calculation of cellular permeabilities. A named list with names being the molecules used in all modules,
and the values being one of the \code{CellularPermeabilityMethods} enum values.
To set the cellular permeability method for a molecule, provide a named list.
TODO https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/1650}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationConfiguration-new}{\code{SimulationConfiguration$new()}}
\item \href{#method-SimulationConfiguration-print}{\code{SimulationConfiguration$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class

Should not be directly used. Instead, use function \code{createSimulationConfiguration()} or the method \code{createSimulationConfiguration()}
from the class \code{MoBiProject}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$new(
  modules,
  individual = NULL,
  expressionProfiles = NULL,
  selectedInitialConditions = NULL,
  selectedParameterValues = NULL,
  settings = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modules}}{A list of \code{Module} objects from which to create in simulation.
The order of modules defines the order in which the modules will be combined to a simulation!}

\item{\code{individual}}{Optional, an individual building block}

\item{\code{expressionProfiles}}{Optional, a list of expression profiles to apply to the simulation.}

\item{\code{selectedInitialConditions}}{By default, the first Initial Conditions
(IC) building block (BB) of each module will be selected. If a module has multiple
IC BBs, it is possible to specify which IC BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the IC BB.
By explicitly setting the value for a specific module to \code{NULL}, no IC BB from the specified module will be applied.
If the list contains a module name that is not part of the provided modules, it will be ignored.}

\item{\code{selectedParameterValues}}{By default, the first Parameter Values
(PV) building block (BB) of each module will be selected. If a module has multiple
PV BBs, it is possible to specify which PV BB to apply by providing a named list,
where the name should be the name of the module and the value the name of the PV BB.
By explicitly setting the value for a specific module to \code{NULL}, no PV BB from the specified module will be applied.
If the list contains a module name that is not part of the provided modules, it will be ignored.}

\item{\code{settings}}{Optional, a \code{SimulationSettings} object defining the simulation settings. If no settings are provided,
default settings will be used upon simulation creation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationConfiguration-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationConfiguration$print(printClassProperties = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{printClassProperties}}{Logical, whether to print class properties (default: \code{FALSE}). If \code{TRUE}, calls first the \code{print} method of the parent class.
Useful for debugging.}

\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
