% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-pk-analyses.R
\docType{class}
\name{SimulationPKAnalyses}
\alias{SimulationPKAnalyses}
\title{SimulationPKAnalyses}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimulationPKAnalyses
}
\description{
PKAnalyses of a simulation (either individual or population simulation)
}
\section{Methods}{

\describe{
  \item{allPKParametersFor(quantityPath)}{All pK Parameters defined for a given path}
  \item{pKParameterFor(quantityPath, pkParameter)}{The pK Parameter defined for the given path and name}
  }
}

\keyword{datasets}
