% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\docType{class}
\name{SimulationResults}
\alias{SimulationResults}
\title{SimulationResults}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimulationResults
}
\description{
Results of a simulation run (either individual or population simulation)
}
\keyword{datasets}
