% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\docType{class}
\name{SimulationResults}
\alias{SimulationResults}
\title{SimulationResults}
\usage{
SimulationResults
}
\description{
Results of a simulation run (either individual or population simulation)
}
\section{Fields}{

\describe{
\item{\code{count}}{the number of individual results (\code{Count==1} generally means that we are dealing with an individual simulation results).}

\item{\code{simulation}}{Reference to the \code{Simulation} used to calculate or import the results (Read-Only).}

\item{\code{timeValues}}{Vector of simulated time output values}

\item{\code{allQuantityPaths}}{List of all paths for which results are defined.}

\item{\code{allIndividualId}}{List of Ids of all individuals that have been simulated}
}}

\keyword{datasets}
