% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-run-options.R
\docType{class}
\name{SimulationRunOptions}
\alias{SimulationRunOptions}
\title{SimulationRunOptions}
\usage{
SimulationRunOptions
}
\description{
Options to be passed to the simulation engine
}
\section{Fields}{

\describe{
\item{\code{numberOfCoresToUse}}{(Maximal) number of cores to be used (1 per default)}

\item{\code{checkForNegativeValues}}{Specifies whether negative values check is on or off. Default is \code{TRUE}}
}}

\keyword{datasets}
