% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-run-options.R
\docType{class}
\name{SimulationRunOptions}
\alias{SimulationRunOptions}
\title{SimulationRunOptions}
\description{
Options to be passed to the simulation engine
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SimulationRunOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{numberOfCores}}{(Maximal) number of cores to be used. This is only relevant when simulating a population simulation.
Default is \code{getOSPSuiteSetting("numberOfCores")}.}

\item{\code{checkForNegativeValues}}{Specifies whether negative values check is on or off. Default is \code{TRUE}}

\item{\code{showProgress}}{Specifies whether progress bar should be shown during simulation run. Default is \code{getOSPSuiteSetting("showProgress")}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationRunOptions-new}{\code{SimulationRunOptions$new()}}
\item \href{#method-SimulationRunOptions-print}{\code{SimulationRunOptions$print()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationRunOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationRunOptions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationRunOptions$new(
  numberOfCores = NULL,
  checkForNegativeValues = NULL,
  showProgress = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{Number of cores to use for the simulation. Default value is \code{getOSPSuiteSetting("numberOfCores")}}

\item{\code{checkForNegativeValues}}{Should the solver check for negative values. Default is \code{TRUE}}

\item{\code{showProgress}}{Should a progress information be displayed. Default value is \code{getOSPSuiteSetting("showProgress")}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SimulationRunOptions} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationRunOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationRunOptions-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationRunOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
