% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-settings.R
\docType{class}
\name{SimulationSettings}
\alias{SimulationSettings}
\title{SimulationSettings}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimulationSettings
}
\description{
Settings associated with a given simulation
}
\section{Fields}{

\describe{
\item{\code{solver}}{Container containing all solver parameters for the simulation (read-only)}

\item{\code{outputSelections}}{All selected quantities (species, observers, parameters) that will be part of the simulated results}
}}

\keyword{datasets}
