% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-characteristics.R
\docType{class}
\name{SnapshotParameter}
\alias{SnapshotParameter}
\title{SnapshotParameter}
\usage{
SnapshotParameter
}
\description{
A parameter typically used in the definition of \code{OriginData} covariates (Height, Weight etc...)
}
\section{Fields}{

\describe{
\item{\code{value}}{Parameter value}

\item{\code{unit}}{Unit in which the value is defined}
}}

\keyword{datasets}
