% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot-parameter.R
\docType{class}
\name{SnapshotParameter}
\alias{SnapshotParameter}
\title{SnapshotParameter}
\description{
A parameter typically used in the definition of \code{IndividualCharacteristics} covariates (Height, Weight etc...)
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SnapshotParameter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{Parameter value}

\item{\code{unit}}{Unit in which the value is defined}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SnapshotParameter-new}{\code{SnapshotParameter$new()}}
\item \href{#method-SnapshotParameter-print}{\code{SnapshotParameter$print()}}
\item \href{#method-SnapshotParameter-printValue}{\code{SnapshotParameter$printValue()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotParameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotParameter-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotParameter$new(ref = NULL, value = NULL, unit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref}}{Optional .NET reference object. If not defined, a new instance will be created}

\item{\code{value}}{Optional value of the parameter.}

\item{\code{unit}}{Optional unit of the value specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SnapshotParameter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotParameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotParameter-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotParameter-printValue"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotParameter-printValue}{}}}
\subsection{Method \code{printValue()}}{
Print the the parameter in one line
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotParameter$printValue(caption)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{caption}}{Caption to display before the value of the parameter}
}
\if{html}{\out{</div>}}
}
}
}
