% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-settings.R
\docType{class}
\name{SolverSettings}
\alias{SolverSettings}
\title{SolverSettings}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SolverSettings
}
\description{
Solver settings associated with a given simulation
}
\section{Fields}{

\describe{
\item{\code{useJacobian}}{Use of Jacobian matrix during calculations}

\item{\code{h0}}{Initial time step size}

\item{\code{hMin}}{Minimum absolute value of step size allowed}

\item{\code{hMax}}{Maximum absolute value of step size allowed}

\item{\code{mxStep}}{Maximum number of internal steps to be taken by the solver in its attempt to reach tout}

\item{\code{relTol}}{Relative tolerance of unknowns}

\item{\code{absTol}}{Absolute tolerance of unknowns}
}}

\keyword{datasets}
