% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-settings.R
\docType{class}
\name{SolverSettings}
\alias{SolverSettings}
\title{SolverSettings}
\description{
Solver settings associated with a given simulation
}
\section{Super classes}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{SolverSettings}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{useJacobian}}{Use of Jacobian matrix during calculations}

\item{\code{h0}}{Initial time step size}

\item{\code{hMin}}{Minimum absolute value of step size allowed}

\item{\code{hMax}}{Maximum absolute value of step size allowed}

\item{\code{mxStep}}{Maximum number of internal steps to be taken by the solver in its attempt to reach tout}

\item{\code{relTol}}{Relative tolerance of unknowns}

\item{\code{absTol}}{Absolute tolerance of unknowns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{SolverSettings$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize">}\href{../../ospsuite/html/DotNetWrapper.html#method-initialize}{\code{ospsuite::DotNetWrapper$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverSettings$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
