% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-sim.R
\docType{data}
\name{Species}
\alias{Species}
\title{Default species defined in PK-Sim}
\format{An object of class \code{list} of length 8.}
\usage{
Species
}
\description{
Default species defined in PK-Sim
}
\keyword{datasets}
