% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-path.R
\docType{data}
\name{StandardContainer}
\alias{StandardContainer}
\title{Standard containers typically available in a PBPK simulation}
\format{An object of class \code{list} of length 4.}
\usage{
StandardContainer
}
\description{
Standard containers typically available in a PBPK simulation
}
\keyword{datasets}
