% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-sim.R
\docType{data}
\name{StandardOntogeny}
\alias{StandardOntogeny}
\title{List of ontogeny supported in PK-Sim}
\format{An object of class \code{list} of length 16.}
\usage{
StandardOntogeny
}
\description{
List of ontogeny supported in PK-Sim
}
\keyword{datasets}
