% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-path.R
\docType{data}
\name{StandardPath}
\alias{StandardPath}
\title{Standard parameter paths typically available in a PBPK simulation}
\format{An object of class \code{list} of length 8.}
\usage{
StandardPath
}
\description{
Standard parameter paths typically available in a PBPK simulation
}
\keyword{datasets}
