% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{StandardPKParameter}
\alias{StandardPKParameter}
\title{Standard PK-Parameters types defined in OSPSuite
This is only used to defined how a user defined PK Parameter should be calculated}
\usage{
StandardPKParameter
}
\description{
Standard PK-Parameters types defined in OSPSuite
This is only used to defined how a user defined PK Parameter should be calculated
}
