% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-pk-parameter.R
\docType{data}
\name{StandardPKParameter}
\alias{StandardPKParameter}
\title{Standard PK-Parameters defined in OSPSuite}
\format{An object of class \code{list} of length 16.}
\usage{
StandardPKParameter
}
\description{
Standard PK-Parameters defined in OSPSuite
}
\keyword{datasets}
