% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter.R
\docType{data}
\name{StandardPKParameter}
\alias{StandardPKParameter}
\title{Standard PK-Parameters types defined in OSPSuite
This is only used to defined how a user defined PK Parameter should be calculated}
\format{An object of class \code{list} of length 16.}
\usage{
StandardPKParameter
}
\description{
Standard PK-Parameters types defined in OSPSuite
This is only used to defined how a user defined PK Parameter should be calculated
}
\keyword{datasets}
