% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toBaseUnit}
\alias{toBaseUnit}
\title{Converts a value given in a specified unit into the base unit of a quantity}
\usage{
toBaseUnit(quantity, value, unit)
}
\arguments{
\item{quantity}{Instance of a quantity from which the base unit will be retrieved}

\item{value}{Value in unit.}

\item{unit}{Unit of value}
}
\description{
Converts a value given in a specified unit into the base unit of a quantity
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in unit (1000 ml) to the base unit (l) => 1
valueInBaseUnit <- toBaseUnit(par, 1000, "ml")
}
