% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toDisplayUnit}
\alias{toDisplayUnit}
\title{Converts a value given in base unit of a quantity into the display unit of a quantity}
\usage{
toDisplayUnit(quantity, values)
}
\arguments{
\item{quantity}{Instance of a quantity from which the base unit will be retrieved}

\item{values}{Value in base unit (single or vector)}
}
\description{
Converts a value given in base unit of a quantity into the display unit of a quantity
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in base unit (1L) to display unit
valueInMl <- toDisplayUnit(par, 1)

valuesInDisplayUnit <- toDisplayUnit(par, c(1, 5, 5))
}
