% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-path.R
\name{toPathString}
\alias{toPathString}
\title{Convert a path array to a path as string with entries separated by '|'}
\usage{
toPathString(...)
}
\arguments{
\item{pathArray}{An array of path entries}
}
\value{
A string built using each entry of the pathArray
}
\description{
Convert a path array to a path as string with entries separated by '|'
}
\examples{

path <- toPathString(c("Organism", "Organ", "Liver"))
}
