% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-units.R
\name{toUnit}
\alias{toUnit}
\title{Converts a value given in base unit of a quantity into a target unit}
\usage{
toUnit(quantity, values, targetUnit)
}
\arguments{
\item{quantity}{Instance of a quantity from which the base unit will be retrieved}

\item{values}{Value in base unit (single or vector)}

\item{targetUnit}{Unit to convert to}
}
\description{
Converts a value given in base unit of a quantity into a target unit
}
\examples{
simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
par <- getParameter("Organism|Liver|Volume", sim)

# Converts the value in base unit (1L) to ml => 1000
valueInMl <- toUnit(par, 1, "ml")

valuesInMl <- toUnit(par, c(1, 5, 5), "ml")
}
