% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-defined-pk-parameter.R
\docType{class}
\name{UserDefinedPKParameter}
\alias{UserDefinedPKParameter}
\title{UserDefinedPKParameter}
\description{
Definition of a user defined PKParameter that can be calculated on top of the standard PK Parameters
}
\section{Super classes}{
\code{\link[rSharp:NetObject]{rSharp::NetObject}} -> \code{\link[ospsuite:DotNetWrapper]{ospsuite::DotNetWrapper}} -> \code{\link[ospsuite:PKParameter]{ospsuite::PKParameter}} -> \code{UserDefinedPKParameter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{startTime}}{Start time in minutes for the calculation of the PK-Parameter.
If not specified, the time will start at the first time point of the simulation (optional)}

\item{\code{startTimeOffset}}{Offset in minutes to apply to the start time or
to the start time of the application identified by  \code{startApplicationIndex}. (0 by default).}

\item{\code{endTime}}{End time in minutes for the calculation of the PK-Parameter.
If not specified, the time will end at the last time point of the simulation (optional)}

\item{\code{endTimeOffset}}{Offset in minutes to apply to the end time or
to the start time of the application identified by \code{endApplicationIndex}. (0 by default).}

\item{\code{startApplicationIndex}}{1-based Index of the application to use to determine the start time for the calculation of the PK-Parameter.
If not specified, the time will start at the first time point of the simulation (optional)}

\item{\code{endApplicationIndex}}{1-based Index of the application to use to determine the end time for the calculation of the PK-Parameter.
If not specified, the time will end at the last time point of the simulation (optional)}

\item{\code{normalizationFactor}}{Factor to use to normalized the calculated PK-Parameter. (typically DrugMass, Dose, DosePerBodyWeight).
It is the responsibility of the caller to ensure that the value is in the correct unit. (optional)}

\item{\code{concentrationThreshold}}{Used in conjunction with the \code{threshold} parameter type.
If defined, the time at which this concentration was reached will be calculated}

\item{\code{standardPKParameter}}{Based parameter to use to perform the PK-Analysis calculation.
See \code{StandardPKParameter} enum for all possible pk parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserDefinedPKParameter-print}{\code{UserDefinedPKParameter$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="call"><a href='../../rSharp/html/NetObject.html#method-NetObject-call'><code>rSharp::NetObject$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="get"><a href='../../rSharp/html/NetObject.html#method-NetObject-get'><code>rSharp::NetObject$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getFields"><a href='../../rSharp/html/NetObject.html#method-NetObject-getFields'><code>rSharp::NetObject$getFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMemberSignature"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMemberSignature'><code>rSharp::NetObject$getMemberSignature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getMethods"><a href='../../rSharp/html/NetObject.html#method-NetObject-getMethods'><code>rSharp::NetObject$getMethods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="getProperties"><a href='../../rSharp/html/NetObject.html#method-NetObject-getProperties'><code>rSharp::NetObject$getProperties()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rSharp" data-topic="NetObject" data-id="set"><a href='../../rSharp/html/NetObject.html#method-NetObject-set'><code>rSharp::NetObject$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite" data-topic="DotNetWrapper" data-id="initialize"><a href='../../ospsuite/html/DotNetWrapper.html#method-DotNetWrapper-initialize'><code>ospsuite::DotNetWrapper$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDefinedPKParameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-UserDefinedPKParameter-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDefinedPKParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
