% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-point.R
\docType{class}
\name{ValuePoint}
\alias{ValuePoint}
\title{ValuePoint}
\usage{
ValuePoint
}
\description{
A point in a table formula
}
\section{Fields}{

\describe{
\item{\code{x}}{The x value of the point.}

\item{\code{y}}{the y value of the point.}

\item{\code{restartSolver}}{Indicates whether the solver should be restarted when this point is reached. Default is \code{FALSE}}
}}

\keyword{datasets}
